% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{adjustDates}
\alias{adjustDates}
\title{Adjust time/start dates of a loaded object}
\usage{
adjustDates(timePars)
}
\arguments{
\item{timePars}{Object containing the relevant time parameters}
}
\value{
A list with dates (POSIXct) start and end, defining the interval [start, end)
}
\description{
Adjust time/start dates of a loaded object
}
\details{
Sub-daily information is displayed only in case of subdaily data
}
\author{
J Bedia \email{joaquin.bedia@gmail.com}
}
\keyword{internal}

