% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biasCorrection.R
\name{adjustPrecipFreq}
\alias{adjustPrecipFreq}
\title{adjustPrecipFreq}
\usage{
adjustPrecipFreq(obs, pred, threshold)
}
\arguments{
\item{obs}{A vector (e.g. station data) containing the observed climate data for the training period}

\item{pred}{A vector containing the simulated climate by the model for the training period.}

\item{threshold}{The minimum value that is considered as a non-zero precipitation.}
}
\description{
Adjusts precipitation frequency in 'p' (prediction) to the observed frequency in 'o'. 
It constitutes a preprocess to bias correct precipitation data following Themeßl et al. (2012).
}
\author{
S. Herrera and M. Iturbide
}
\keyword{internal}
