% Generated by roxygen2 (4.0.1): do not edit by hand
\name{aggr6hToDailyMean}
\alias{aggr6hToDailyMean}
\title{Performs the aggregation of 6 hourly data to daily mean}
\usage{
aggr6hToDailyMean(x, aggr.fun)
}
\arguments{
\item{x}{a vector of data (a time series actually)}
}
\value{
The vector aggregated by the mean of each four elements
}
\description{
This is an auxiliary routine used in \code{\link{toDD}} within the \code{makeSubset} functions
}
\details{
The function assumes that we are loading 6-h data so that the length of x is always divisible by 4.
Otherwise it will raise an error. Because selections are based on seasons and years, there is no reason that this error
 ever occurs unless we are taking the previous forecast time for deaccumulation purposes (but so far there is not any
 de-accumulable 6h variable, and it is not likely that one is ever going to come across any...).
}
\author{
J Bedia \email{joaquin.bedia@gmail.com}
}
\references{
\url{http://stackoverflow.com/questions/18434305/apply-a-function-to-every-n-elements-of-a-vector}
}
\keyword{internal}

