% Generated by roxygen2 (4.0.1): do not edit by hand
\name{analogs}
\alias{analogs}
\title{Analog downscaling}
\usage{
analogs(obs, pred, sim, n.neigh = 1, sel.fun = c("random", "mean"))
}
\arguments{
\item{obs}{Predictand correpsonding to an observational dataset. This could be either an stations object as
returned by \code{\link{loadStationData}} or a field.}

\item{pred}{Predictor(s) dataset, i.e., explanatory variables used to build the model (typically reanalysis data).
This can be a single field, or more often a multifield, if several variables are to be used as predictors or their
principal components (as returned by \code{\link{prinComp}}).}

\item{sim}{Simulation data. Field/multifield (possibly also multi-member) data providing the model simulations to be
downscaled.}

\item{n.neigh}{Integer indicating the number of closest neigbours to retain for analog construction. Default to 1.}

\item{sel.fun}{Criterion for the construction of analogs when several neigbours are chosen. Ignored when \code{n.neig = 1}.
Current values are \code{"random"} (the default) and \code{"mean"}. See details.}
}
\description{
Implementation of the downscaling analogs method
}
\details{
\strong{Spatial consistency}

Several checks of spatial consistency are performed. In particular, note that both 'pred' (reanalysis) and 'sim' (model
simulations) should be in the same grid. This consistency must be ensured by the user prior to entering these arguments,
for instance by means of the \code{\link{interpGridData}} function in conjunction with the \code{\link{getGrid}} method.

\strong{Scaling and centering}

When the climate variables are used as predictors instead of the PCs, these are previously centered and scaled
using the mean and sigma parameters globally computed for the whole spatial domain (This is equivalent to the \dQuote{field})
method in the \code{\link{prinComp}} function. The simulation data will use the parameters obtained when scaling and centering
the predictors dataset. In case that the predictors come from a PC analysis object (as returned by \code{\link{prinComp}}), the
parameters for rescaling the simulation data are passed by the predictors.

\strong{Construction of analogs using multiple neighbours}

The argument \code{sel.fun} controls how the analogs are constructed when considering more than the first neighbour (argument
\code{n.neigh} > 1). In this case the \code{"random"} choice randomly selects one of the \code{n.neigh} neighbours, while
}
\author{
J. Bedia \email{joaquin.bedia@gmail.com}
}
\references{
\itemize{
\item Benestad, R.E., Hanssen-Bauer, I. and Chen, D., 2008. Empirical-Statistical Downscaling,
 1st ed. World Scientific Publishing, Singapore
\item Gutierrez, J.M. \emph{et al.}, 2013. Reassessing Statistical downscaling techniques for
 their robust application under climate change conditions. J. Clim. 26, 171-188
\item
Bedia, J. \emph{et al.}, 2013. Robust projections of Fire Weather Index in the Mediterranean
 using statistical downscaling. Clim. Change 120, 229–247.
}
}
\seealso{
\code{\link{prinComp}} for details on principal component/EOF analysis
\code{\link{loadMultiField}}, \code{\link{makeMultiField}} for multifield creation
\code{\link{loadGridData}} and \code{\link{loadStationData}} for loading fields and station data respectively.

Other downscaling: \code{\link{biasCorrection}};
  \code{\link{calibrateProj}}; \code{\link{isimip}}
}

