% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsANALOGS.R
\name{analogs.test}
\alias{analogs.test}
\title{Donwscaling with the analogs method (test).}
\usage{
analogs.test(newdata, x, y, dates, info)
}
\arguments{
\item{newdata}{The grid data.  Class: matrix.}

\item{x}{The grid data where to search the analogs. Class: matrix.}

\item{y}{The observations data of the same days of x. Class: matrix.}

\item{dates}{A list containing both the newdata grid  and x grid, dates.}

\item{info}{A list containing the information required to perform the analogs method: number of analogs, function, window. These parameters are grouped in a list in \code{\link[downscaleR]{analogs.train}}.}
}
\value{
A matrix containing the predictions.
}
\description{
Donwscaling with the analogs method (test).
}
\details{
The selected functions use the base R functions: \code{\link[base]{mean}}, \code{\link[stats]{median}}, \code{\link[stats]{weighted.mean}}, \code{\link[stats]{quantile}}, \code{\link[base]{Extremes}}.
}
\author{
J. Bano-Medina
}
