% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsANALOGS.R
\name{analogs.train}
\alias{analogs.train}
\title{Donwscaling with the analogs method (train).}
\usage{
analogs.train(x, y, n.analogs = 4, sel.fun = "mean", window = 0,
  n.random = NULL)
}
\arguments{
\item{x}{The grid data. Class: matrix.}

\item{y}{The observations data. Class: matrix.}

\item{n.analogs}{Number of analogs. Default is 4.}

\item{sel.fun}{Function applied to the analogs. Can be "mean", "wmean" (i.e., weighted mean), "max", "min", "median", "prcXX" (i.e., prc85 means the 85th percentile of the analogs values distribution). Default is "mean".}

\item{window}{Window of days removed when selecting analogs. Default is 7. This means that the 7 days after the date and the 7 days before the date are removed.}

\item{n.random}{An integer. Choose N random analogs among the closest n.analogs.}
}
\value{
A list containing the grid data, the observations and a list with information concerning the analogs.
}
\description{
Donwscaling with the analogs method (train).
}
\details{
The analogs actually is not a model in the sense that it is not really trained. The model would be the data where to search the analogs. For this reason this function only saves the information to search analogs in a list.
}
\author{
J. Bano-Medina
}
