% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/timeAggregation.R
\name{annualAggregationMain}
\alias{annualAggregationMain}
\title{Annual aggregation of the data in a field}
\usage{
annualAggregationMain(data, FUN = c("mean", "min", "max", "sum"), id)
}
\arguments{
\item{data}{array (can be multimember).}

\item{FUN}{a character indicating the aggregation function. Possibilities are: "mean", "min", "max", "sum".}

\item{id}{as argument INDEX in function tapply.}
}
\value{
array  with the time dimension annually aggregated.
}
\description{
Aggregates data sub-annual data annually by specifiyng an aggregation function
}
\author{
M. Iturbide \email{maibide@gmail.com}
}

