% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biasCorrection.R
\name{biasCorrection1D}
\alias{biasCorrection1D}
\title{Bias correction methods on 1D data}
\usage{
biasCorrection1D(o, p, s, method, scaling.type, fitdistr.args, precip,
  pr.threshold, n.quantiles, extrapolation, theta, parallel = FALSE,
  max.ncores = 16, ncores = NULL)
}
\arguments{
\item{o}{A vector (e.g. station data) containing the observed climate data for the training period}

\item{p}{A vector containing the simulated climate by the model for the training period.}

\item{s}{A vector containing the simulated climate for the variable used in \code{p}, but considering the test period.}

\item{method}{method applied. Current accepted values are \code{"eqm"}, \code{"delta"},
\code{"scaling"}, \code{"pqm"}, \code{"gpqm"}, \code{"variance"}, \code{"loci"} and \code{"ptr"}.}

\item{scaling.type}{Character indicating the type of the scaling method. Options are \code{"additive"} (default)
or \code{"multiplicative"} (see details). This argument is ignored if \code{"scaling"} is not selected as the bias correction method.}

\item{fitdistr.args}{Distribution to use when applying the "pqm" method (parametric quantile mapping).
Choices are "gamma" or "gaussian" (see details).}

\item{precip}{Logical indicating if o, p, s is precipitation data.}

\item{pr.threshold}{The minimum value that is considered as a non-zero precipitation. Ignored for
\code{varcode} values different from \code{"pr"}. Default to 1 (assuming mm).}

\item{extrapolation}{Character indicating the extrapolation method to be applied to correct values in
\code{"s"} that are out of the range of \code{"p"}. Extrapolation is applied only to the \code{"eqm"} method, 
thus, this argument is ignored if other bias correction method is selected. Default is \code{"none"} (do not extrapolate).}

\item{theta}{numeric indicating  upper threshold (and lower for the left tail of the distributions, if needed) 
above which precipitation (temperature) values are fitted to a Generalized Pareto Distribution (GPD). 
Values below this threshold are fitted to a gamma (normal) distribution. By default, 'theta' is the 95th 
percentile (5th percentile for the left tail). Only for \code{"gpqm"} method.}

\item{parallel}{Logical. Should parallel execution be used?}

\item{max.ncores}{Integer. Upper bound for user-defined number of cores.}

\item{ncores}{Integer number of cores used in parallel computation. Self-selected number of
cores is used when \code{ncpus = NULL} (the default), or when \code{maxcores} exceeds the default \code{ncores} value.}
}
\description{
Implementation of several standard bias correction methods
}
\author{
M. Iturbide
}
\keyword{internal}
