% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbleValidation.R
\name{bubbleValidation}
\alias{bubbleValidation}
\title{Bubble plot}
\usage{
bubbleValidation(mm.obj, obs, select.year, score = TRUE,
  size.as.probability = TRUE)
}
\arguments{
\item{mm.obj}{A multi-member object with predictions, either a grid or a multi-member station object as a result of
downscaling of a forecast using station data. See details.}

\item{obs}{The benchmarking observations for forecast verification.}

\item{select.year}{Year within the whole verification period to display the results for.}

\item{score}{Logical. Whether to include or not the relative operating characteristic skill score (ROCSS). See details.}

\item{size.as.probability}{Logical. Whether to include the tercile probabilities (magnitude proportional to bubble radius)
in the plot. See details.}
}
\description{
Bubble plot for the visualization of the skill of an ensemble forecast prediction. It provides a
 spatially-explicit representation of the skill, resolution and reliability of a probabilistic predictive system in
 a single map.
}
\details{
For each member, the daily predictions are averaged to obtain a single seasonal forecast. The corresponding terciles 
for each ensemble member are then computed for the analysis period. Thus, each particular grid point, member and season,
are categorized into three categories (above, between or below), according to their respective climatological 
terciles. Then, a probabilistic forecast is computed year by year by considering the number of members falling 
within each category. For instance, probabilities below 1/3 are very low, indicating that a minority of the members 
falls in the tercile. Conversely, probabilities above 2/3 indicate a high level of member agreement (more than 66\% of members
falling in the same tercile). Color represents the tercile with the highest probability for the selected year. The bubble size 
indicates the probability of that tercile. This option is not plotted if the size.as.probability argument is FALSE.

Finally, the ROC Skill Score (ROCSS) is computed. For each tercile, it provides a quantitative measure of the forecast skill,
and it is commonly used to evaluate the performance of probabilistic systems (Joliffe and Stephenson 2003). The value of 
this score ranges from 1 (perfect forecast system) to -1 (perfectly bad forecast system). A value zero indicates no skill 
compared with a random prediction. The transparency of the bubble is associated to the ROCSS (negative values are
plotted with x).  This option is not plotted if the score argument is FALSE.
}
\note{
The computation of climatological terciles requires a representative period to obtain meaningful results.
}
\author{
J. Bedia, M.D. Frias and J. Fernandez
}
\references{
Jolliffe, I. T. and Stephenson, D. B. 2003. Forecast Verification: A Practitioner's Guide in 
 Atmospheric Science, Wiley, NY
}
\seealso{
Other visualization: \code{\link{plotMeanGrid}},
  \code{\link{quickDiagnostics}},
  \code{\link{tercileValidation}}
}

