% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calibrateProj}
\alias{calibrateProj}
\title{Bias correction methods}
\usage{
calibrateProj(obs, pred, sim, method = c("qqmap", "delta", "scaling",
  "unbiasing", "piani"), varcode = c("tas", "hurs", "pr", "wss"),
  pr.threshold = 1)
}
\description{
Implementation of several standard bias correction methods
}
\author{
S. Herrera \email{sixto@predictia.es}
}

