% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/climatology.R
\name{climatology}
\alias{climatology}
\title{Compute a grid climatology}
\usage{
climatology(grid, clim.fun = list(FUN = "mean", na.rm = TRUE),
  by.member = TRUE, parallel = FALSE, max.ncores = 16, ncores = NULL)
}
\arguments{
\item{grid}{Input grid}

\item{clim.fun}{Function to compute the climatology. This is specified as a list,
indicating the name of the aggregation function in first place (as character), and other optional arguments
 to be passed to the aggregation function. Default to mean (i.e., \code{clim.fun = list(FUN="mean",na.rm = TRUE)}).}

\item{by.member}{Logical. In case of multimember grids, should the climatology be computed sepparately
for each member (\code{by.member=TRUE}), or a single climatology calculated from the ensemble mean
 (\code{by.member=FALSE})?. Default to \code{TRUE}.}

\item{parallel}{Logical. Should parallel execution be used?}

\item{max.ncores}{Integer. Upper bound for user-defined number of cores.}

\item{ncores}{Integer number of cores used in parallel computation. Self-selected number of
cores is used when \code{ncpus = NULL} (the default), or when \code{maxcores} exceeds the default \code{ncores} value.}
}
\value{
A grid corresponding to the climatology. See details.
}
\description{
Calculates the climatology (i.e., complete temporal aggregation, typically the mean)
of the input grid.
}
\details{
Two attributes are appended to the grid:
\itemize{
\item \code{climatology:fun}, added to the \code{Data} component of the grid,
indicating the function used to compute the climatology.
\item \code{season}, added to the \code{Dates} component (if not yet existing), in order to provide information
 on the season for which the climatology has been computed.
}
}
\section{Parallel Processing}{


Parallel processing is enabled using the \pkg{parallel} package.
Parallelization is undertaken by a FORK-type parallel socket cluster formed by \code{ncores}.
If \code{ncores} is not specified (default), \code{ncores} will be one less than the autodetected number of cores.
The maximum number of cores used for parallel processing can be set with the \code{max.ncores} argument,
although this will be reset to the auto-detected number of cores minus 1 if this number is exceeded. Note that not all
code, but just some critical loops within the function are parallelized.

In practice, parallelization does not always result in smaller execution times, due to the parallel overhead.
However, parallel computing may potentially provide a significant speedup for the
particular case of large multimember datasets or large grids.

Parallel computing is currently not available for Windows machines.
}
\examples{
\dontrun{
#' # Maximum July surface temp forecast climatology
data("tasmax_forecast")
# (Parallelization option has no effect under WinOS)
# Aggregate all members before computing the climatology
tx_mean.clim <- climatology(tasmax_forecast,
                            by.member = FALSE,
                            parallel = TRUE)
# Note that time dimension is not dropped, and the new attributes
str(tx_mean.clim$Data)
str(tx_mean.clim$Dates)
# Compute a climatology for each member sepparately
tx_mean_15mem.clim <- climatology(tasmax_forecast,
                                  by.member = TRUE,
                                  parallel = TRUE)
str(tx_mean_15mem.clim$Data)
# 9 different climatologies, one for each member

# Flexible aggregation function definition:
# Example: climatology of the absolute maximum daily precipitation (Winter 2000):
data("NCEP_Iberia_tp")
tpmax <- climatology(NCEP_Iberia_tp,
                    clim.fun = list(FUN = "max"))
plotMeanGrid(tpmax)
}
}
\author{
J. Bedia
}

