% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dataInventory.ASCII}
\alias{dataInventory.ASCII}
\title{Make data inventory from a station dataset in standard ASCII format}
\usage{
dataInventory.ASCII(dataset, rs)
}
\arguments{
\item{dataset}{path to the directory containng the dataset}

\item{rs}{}
}
\description{
Make data inventory from a station dataset in standard ASCII format
}
\author{
J Bedia \email{joaquin.bedia@gmail.com}
}
\keyword{internal}

