% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataInventory.R
\name{dataInventory.NetCDF}
\alias{dataInventory.NetCDF}
\title{Inventory of a gridded dataset}
\usage{
dataInventory.NetCDF(dataset)
}
\arguments{
\item{dataset}{A full path to the file describing the dataset (NcML)}
}
\value{
A (named) list whose length is determined by the number of variables stored in the dataset,
its names corresponding to the short names of the variables.
For each variable, information on the variable long name, data type, units and
characteristics of its dimensions is provided.
}
\description{
Returns a list with summary information about the variables stored in a gridded dataset.
Sub-routine of \code{dataInventory}
}
\author{
J. Bedia \email{joaquin.bedia@gmail.com}
}
\keyword{internal}

