% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dateReplacement}
\alias{dateReplacement}
\title{Compute dates of a downscaled observational dataset}
\usage{
dateReplacement(obs.dates, sim.dates)
}
\arguments{
\item{obs.dates}{The \code{Dates} slot of the 'obs' input in the downscaling method}

\item{sim.dates}{The \code{Dates} slot of the 'sim' input in the downscaling method}
}
\value{
A new \code{Dates} list that preserves the temporal extent of the downscaled simulations but considering
the temporal resolution of the downscaled variable
}
\description{
The function calculates the appropiate Dates slot in the returned output of downscaling functions,
considering the possible mismatches in time resolution between predictors and predictand, the multifield dates slot etc.
}
\details{
The function is intended for internal use only. Sometimes the time resolution of the predictors does not match
that of the downscaled variable (e.g., suppose that instantaneous surface temperature at 12:00 UTC is used as predictor
 of daily minimum temperature). In addition, in case of multiple predictors the \code{Dates} slot of the simulated series
 has several start/end time lists, one for each predictor, while there is only one predictand. For this reason,
 the function takes care of adjusting adequately the returned \code{Dates} slot.
}
\author{
J. Bedia \email{joaquin.bedia@gmail.com}
}
\keyword{internal}

