% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deaccum.R
\name{deaccum}
\alias{deaccum}
\title{Performs deaccumulation}
\usage{
deaccum(x, t.ranges, dff)
}
\arguments{
\item{x}{a vector of (accumulated) data.}

\item{t.ranges}{a vector defining the start/end indices for each annual season, as returned by
\code{dictionaryTransform}. This is used to restart the deaccumulation at the beginning of each season.}

\item{dff}{Logical. Is deaccumulation performed from the first value of the time series?.
Passed by dictionaryTransform,coming from getForecastTimeDomain.S4. See Details.}
}
\value{
a vector of deaccumulated data
}
\description{
Used for deaccumulation of precipitation in certain model data (e.g. System4). Sub-routine
of \code{dictionaryTransform}.
}
\details{
When leadMonth equals 0, there is not a previous day for starting deaccumulation, and therefore the
 first value of the time series is taken 'as is'. Otherwise, one value before the start has to be taken
 to preserve time series length (this is previously done by getForecastDomain.S4).
}
\author{
J. Bedia \email{joaquin.bedia@gmail.com}
}
\keyword{internal}

