% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biasCorrection.R
\name{delta}
\alias{delta}
\title{Delta method for bias correction}
\usage{
delta(o, p, s)
}
\arguments{
\item{o}{A vector (e.g. station data) containing the observed climate data for the training period}

\item{p}{A vector containing the simulated climate by the model for the training period.}

\item{s}{A vector containing the simulated climate for the variable used in \code{x}, but considering the test period.}
}
\description{
Implementation of Delta method for bias correction
}
\author{
S. Herrera and M. Iturbide
}
\keyword{internal}
