% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/detrendGrid.R
\name{detrendGrid}
\alias{detrendGrid}
\title{Linear detrending}
\usage{
detrendGrid(grid, parallel = FALSE, max.ncores = 16, ncores = NULL)
}
\arguments{
\item{grid}{Input grid (possibly multimember)}

\item{parallel}{Logical. Should parallel execution be used?}

\item{max.ncores}{Integer. Upper bound for user-defined number of cores.}

\item{ncores}{Integer number of cores used in parallel computation. Self-selected number of
cores is used when \code{ncpus = NULL} (the default), or when \code{maxcores} exceeds the default \code{ncores} value.}
}
\value{
A detrended grid. Values are residuals.
}
\description{
Perform a linear detrending along the time dimension of a grid
}
\details{
Performs a simple linear detrending by fitting a linear model and retaining the residuals.
An attribute indicating the linear detrending is added to the \code{Variable} component of the output grid.
}
\section{Parallel Processing}{


Parallel processing is enabled using the \pkg{parallel} package.
Parallelization is undertaken by a FORK-type parallel socket cluster formed by \code{ncores}.
If \code{ncores} is not specified (default), \code{ncores} will be one less than the autodetected number of cores.
The maximum number of cores used for parallel processing can be set with the \code{max.ncores} argument,
although this will be reset to the auto-detected number of cores minus 1 if this number is exceeded. Note that not all
code, but just some critical loops within the function are parallelized.

In practice, parallelization does not always result in smaller execution times, due to the parallel overhead.
However, parallel computing may potentially provide a significant speedup for the
particular case of large multimember datasets or large grids.

Parallel computing is currently not available for Windows machines.
}
\examples{
data("iberia_ncep_ta850")
monthly <- aggregateGrid(iberia_ncep_ta850, aggr.m = list(FUN = "mean"))
plot(monthly$Data[,4,2], ty = 'l')
abline(reg = lm(monthly$Data[,4,2] ~ I(1:length(monthly$Data[,4,2]))))
det <- detrendGrid(monthly, parallel = FALSE)
# Detrended series in red
lines(det$Data[,4,2], col = "red")
abline(reg = lm(det$Data[,4,2] ~ I(1:length(det$Data[,4,2]))), col = "red")
}
\author{
J Bedia
}

