% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaryLookup.R
\name{dictionaryLookup}
\alias{dictionaryLookup}
\title{Searches variable string in the dictionary}
\usage{
dictionaryLookup(dicPath, var, time)
}
\arguments{
\item{dicPath}{Full path to the dictionary file (a csv file with extension \sQuote{.dic}).}

\item{var}{Character string with the (standard) name of the variable}

\item{time}{Time specification.}
}
\value{
A data.frame of 1 row with the mapping information
}
\description{
Searches variable string provided in the dictionary to map it in the vocabulary, in order to get
all the necessary information for variable homogenization. It also includes a new column specifying the
aggregation function to be applied (if any).
}
\author{
J. Bedia \email{joaquin.bedia@gmail.com}
}
\references{
\url{http://meteo.unican.es/trac/wiki/udg/ecoms/RPackage/homogeneization}
}
\keyword{internal}

