% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaryTransformGrid.R
\name{dictionaryTransformGrid}
\alias{dictionaryTransformGrid}
\title{Performs variable transformation}
\usage{
dictionaryTransformGrid(dic, timePars, mdArray)
}
\arguments{
\item{dic}{Dictionary line for the variable, as returned by \code{dictionaryLookup} or \code{dictionaryLookup.ECOMS}}

\item{timePars}{A list of time selection parameters, as returned by \code{getTimeDomain}}

\item{mdArray}{A n-dimensional array, as returned by \code{makeSubset}}
}
\value{
The transformed n dimensional array of data. See details.
}
\description{
Performs variable transformation according to dictionary specifications.
Sub-routine of \code{loadGridDataset}.
}
\details{
When performing deaccumulation, the shape of the output array may change. This is due to the re-ordering
of the dimensions done by \code{apply}, and also in the length of the time dimension, as usually the first time value
of each runtime is lost in the deaccumulation. The function handles the re-ordering of the "dimensions" label in this case.
}
\note{
The current implementation does not support deaccumulation of CFSv2 variables (not needed so far...)
}
\author{
J. Bedia \email{joaquin.bedia@gmail.com}
}
\keyword{internal}

