% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscalePredict.R
\name{downs.predict}
\alias{downs.predict}
\title{Switch to selected downscale method.}
\usage{
downs.predict(x, method, atomic_model)
}
\arguments{
\item{x}{The grid data. Class: matrix.}

\item{method}{The method of the given model.}

\item{atomic_model}{An object containing the statistical model of the selected method.}
}
\value{
A matrix with the predictions.
}
\description{
Internal function of \code{\link[downscaleR]{downscalePredict}} that switches to the corresponding method.
}
\details{
This function is internal and should not be used by the user. The user should use \code{\link[downscaleR]{downscalePredict}}.
}
\author{
J. Bano-Medina
}
