% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscale-package.R
\docType{package}
\name{downscale-package}
\alias{downscale}
\alias{downscale-package}
\title{downscale: What the Package Does}
\description{
What the package does.
}
\author{
\strong{Maintainer}: Nicolas Casajus \email{nicolas.casajus@fondationbiodiversite.fr}

}
\keyword{internal}
