% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscale.R
\name{downscale}
\alias{downscale}
\title{Perfect-prog downscaling}
\usage{
downscale(y, x, newdata = x, method = c("analogs", "glm", "lm"),
  simulate = c("no", "yes"), n.analogs = 1, sel.fun = c("mean", "wmean",
  "max", "min", "median"), wet.threshold = 0.1, n.pcs = NULL,
  cross.val = c("none", "loocv", "kfold"), folds = NULL)
}
\arguments{
\item{y}{The observations dataset. It should be an object as returned by \pkg{loadeR}.}

\item{x}{The input grid. It should be an object as returned by \pkg{loadeR}.}

\item{newdata}{It should be an object as returned by \pkg{loadeR} and consistent with x. Default is newdata = x.}

\item{method}{Downscaling method. Options are c = ("analogs","glm","lm"). Glm can only be set when downscaling precipitation.}

\item{simulate}{Character. Options are \code{"no"}, \code{"yes"}.}

\item{n.analogs}{Applies only when \code{method="analogs"} (otherwise ignored). Integer indicating the number of closest neigbours to retain for analog construction. Default to 1.}

\item{sel.fun}{Applies only when \code{method="analogs"} (otherwise ignored). Criterion for the construction of analogs when several neigbours are chosen. Ignored when \code{n = 1}.
Current values are \code{"mean"} (the default), \code{"wmean"},  \code{"max"},  \code{"min"} and  \code{"median"}.}

\item{wet.threshold}{Value below which precipitation amount is considered zero}

\item{n.pcs}{Integer indicating the number of EOFs to be used as predictors}

\item{cross.val}{Should cross-validation be performed? methods available are leave-one-out (\code{"loocv"})
and k-fold (\code{"kfold"}). Default to \code{"none"}, which does not perform cross-validation.}

\item{folds}{Only requiered if \code{cross.val = "kfold"}, otherwise ignored. Could be a fraction, value between (0,1) indicating the fraction of the data that will define the train set, 
or an integer indicating the number of folds. It can also be a list of folds indicating the years of each fold.}
}
\value{
The prediction structure.
}
\description{
Workhorse function to call the different perfect-prog downscaling methods
}
\details{
\strong{Scaling and centering}
When the climate variables are used as predictors instead of the PCs, these are previously centered and scaled
using the mean and sigma parameters globally computed for the whole spatial domain.
}
\examples{
x <- makeMultiGrid(NCEP_Iberia_hus850, NCEP_Iberia_ta850)
newdata <- subsetGrid(x, years = 1994:1995)
x <- subsetGrid(x, years = 1985:1993)
# Loading predictands
y <- VALUE_Iberia_pr
y <- getTemporalIntersection(obs = y,prd = x, "obs" )
x <- getTemporalIntersection(obs = y,prd = x, "prd" )
### Analogs ###
# None
yp <- downscale(y,x,method = "analogs")
yp <- downscale(y,x,newdata,method = "analogs")
# kfold
yp <- downscale(y,x,method = "analogs", n.pcs = 15,
                cross.val = "kfold", folds = list(c("1985","1986","1987"),
                                                  c("1988","1989","1990"),
                                                  c("1991","1992","1993")))
### GLM ###
# None
yp <- downscale(y,x,method = "glm", simulate = "no",  n.pcs = 10,
                wet.threshold = 1)
yp <- downscale(y,x,method = "glm", simulate = "yes", n.pcs = 10,
                wet.threshold = 1)
# kfold
yp <- downscale(y,x,method = "glm", simulate = "no", n.pcs = 10,
                cross.val = "kfold", folds = list(c("1985","1986","1987"),
                                                  c("1988","1989","1990"),
                                                  c("1991","1992","1993")))
}
\seealso{
\url{https://github.com/SantanderMetGroup/downscaleR/wiki/training-downscaling-models} for detailed examples.
}
