% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biasCorrection.R
\name{dqm}
\alias{dqm}
\title{Detrended quantile matching}
\usage{
dqm(o, p, s, precip, pr.threshold, n.quantiles, detrend = TRUE)
}
\arguments{
\item{o}{A vector (e.g. station data) containing the observed climate data for the training period}

\item{p}{A vector containing the simulated climate by the model for the training period.}

\item{s}{A vector containing the simulated climate for the variable used in \code{p}, but considering the test period.}

\item{precip}{Logical indicating if o, p, s is precipitation data.}

\item{pr.threshold}{Integer. The minimum value that is considered as a non-zero precipitation.}

\item{n.quantiles}{Integer. Maximum number of quantiles to estimate. Default: same as data length.}

\item{detrend}{logical. Detrend data prior to bias correction? Default. TRUE.}
}
\description{
Detrended quantile matching with delta-method extrapolation
}
\details{
DQM method developed by A. Canon, from \url{https://github.com/pacificclimate/ClimDown}, \url{https://cran.r-project.org/web/packages/ClimDown/}.
}
\references{
Cannon, A.J., S.R. Sobie, and T.Q. Murdock (2015) Bias Correction of GCM Precipitation by Quantile Mapping: How Well Do Methods Preserve Changes in Quantiles and Extremes?. J. Climate, 28, 6938–6959, \url{https://doi.org/10.1175/JCLI-D-14-00754.1}
}
\author{
A. Cannon (acannon@uvic.ca), A. Casanueva
}
\keyword{internal}
