% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{draw.world.lines}
\alias{draw.world.lines}
\title{Land borders}
\source{
Postprocessed from the original shapefile from Natural Earth (http://www.naturalearthdata.com/downloads/110m-physical-vectors/)
}
\usage{
draw.world.lines(...)
}
\arguments{
\item{...}{Graphical parameters passed to \code{\link{lines}}.}
}
\value{
Draws a simplied land border areas as lines onto the map
}
\description{
Add land borders to a map
}
\details{
The function loads a built-in world segments dataset created ad hoc to avoid dependencies on other packages (i.e. 'maps').
Geographical lonlat coordinates in wgs84.
}
\author{
J. Bedia
}
\keyword{internal}

