% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyVery2grid.R
\name{easyVeri2grid}
\alias{easyVeri2grid}
\title{easyVerification matrix to climatological grid conversion}
\usage{
easyVeri2grid(easyVeri.mat, obs.grid, verifun = NULL)
}
\arguments{
\item{easyVeri.mat}{A matrix containing the verification score,
as returned by \code{\link[easyVerification]{veriApply}}}

\item{obs.grid}{The grid containing the verifying observations used in the call to \code{veryApply}
producing the score matrix.}

\item{verifun}{Optional. Character string indicating the value of the \code{verifun} value. Just for 
a better traceability and metadata completeness.}
}
\value{
A climatological grid.
}
\description{
Convert a xyz-type score matrix as returned by \code{veriApply} to a climatological grid
}
\author{
J. Bedia
}
\seealso{
\code{\link{climatology}}, \code{\link{plotClimatology}}.
}

