% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscale.predict.R
\name{elm.predict}
\alias{elm.predict}
\title{Donwscaling with a given extreme learning machine model (ELM).}
\usage{
elm.predict(x, weights, info)
}
\arguments{
\item{x}{The grid data. Class: matrix.}

\item{weights}{Matrix of coefficients as returned from \code{\link[downscaleR]{elm.train}}}

\item{info}{This is a list containing the information/parameters used in \code{\link[downscaleR]{elm.train}}. Containes the optional parameters: fitting, neurons, Act.F, area.region and area.module. See \code{\link[downscaleR]{elm.train}} for information relevant to these parameters.}
}
\value{
The predicted matrix.
}
\description{
Donwscaling with a given extreme learning machine (ELM) calculated in \code{\link[downscaleR]{downscale.predict}} via \code{\link[downscaleR]{elm.train}}.
}
\details{
This function is internal and should not be used by the user. The user should use \code{\link[downscaleR]{downscale.predict}}.
}
\author{
J. Bano-Medina
}
