% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdas.R
\docType{data}
\name{EOBS_Iberia_tas}
\alias{EOBS_Iberia_tas}
\title{Grid containing E-OBS observation data of temperature for the Iberian Peninsula.}
\format{A grid}
\source{
subset of the E-OBS observational gridded dataset
}
\description{
EOBS_Iberia_tas is a grid object returned by \code{loadECOMS} from package \pkg{loadeR.ECOMS}
}
\seealso{
\code{\link{makeMultiGrid}}, \code{\link[loadeR]{loadGridData}}
}
\keyword{gridded}
\keyword{observations}

