% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rdas.R
\docType{data}
\name{EOBS_Iberia_tp}
\alias{EOBS_Iberia_tp}
\title{Grid containing E-OBS observation data of precipitation for the Iberian Peninsula.}
\format{A grid}
\source{
subset of the E-OBS observational gridded dataset
}
\description{
EOBS_Iberia_tp is a grid object returned by function loadGridData
(package loadeR):
season = c(12,1,2),
years = 1991:2000,
var="tas",
lonLim = c(-10,5),
latLim= c(34,44)
}
\seealso{
\code{\link{makeMultiField}}, \code{\link[loadeR]{loadGridData}}
}
\keyword{gridded}
\keyword{observations}

