% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid3Dto2Dmatrix.R
\name{expandIrregularCoords}
\alias{expandIrregularCoords}
\title{Expand irregular grid  coords}
\usage{
expandIrregularCoords(irregularXYcoords)
}
\arguments{
\item{irregularXYcoords}{And irregular xyCoords element of an irregular grid}
}
\value{
a 2D matrix of x-y coordinates
}
\description{
Expand the coordinates of irregular grids into a 2D matrix of x-y coordinates
}
\author{
J Bedia
}
\keyword{internal}

