% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fieldFromPCs}
\alias{fieldFromPCs}
\title{Reconstruct a Field from EOFs and principal components}
\usage{
fieldFromPCs(prinCompObj, var)
}
\arguments{
\item{prinCompObj}{A EOF analysis object as returned by \code{\link{prinComp}}}

\item{var}{Character string indicating the variable to be re-constructed}
}
\value{
A list similar to the \code{\link{loadGridData}} output, but simplified. See details.
}
\description{
Reconstructs a field of a climatic variable from the outputs of a
principal components analysis
}
\details{
The output of this function returns the minimum required information to use the
 \code{\link{plotMeanField}} method, and is intended for comparison and visual analysis
 of the differences between the original fields and the reconstructed ones, for instance
 in determining an optimal number of PCs etc... Hence, the temporal information (i.e.,
  the \code{Dates} object) is lost, and should be retrieved from the original field/multifield
  object used to compute the PC/EOF analysis.
}
\author{
J. Bedia \email{joaquin.bedia@gmail.com}
}
\seealso{
Other loading.grid: \code{\link{getGrid}};
  \code{\link{interpGridData}}; \code{\link{loadGridData}};
  \code{\link{loadMultiField}}; \code{\link{prinComp}}

Other multifield: \code{\link{loadMultiField}};
  \code{\link{prinComp}}
}

