% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/findPointXYindex.R
\name{findPointXYindex}
\alias{findPointXYindex}
\title{Define user-defined geolocation parameters}
\usage{
findPointXYindex(lonLim, latLim, gcs)
}
\arguments{
\item{lonLim}{x domain definition}

\item{latLim}{y domain definition}

\item{gcs}{the grid coordinate system, formally a \sQuote{GridCoordSys} java object}
}
\value{
a list with the new lonLim and latLim parameters (if changed) and the pointXYindex
 definition, consisting of a integer vector of length two.
}
\description{
Finds the XY index position of single point selections
}
\details{
The function applies the java method \sQuote{findXYindexFromCoord}, being the input
data the point coordinates in the native units. Thus, if longitudes are coded in the form [0-360] in the dataset,
the input lonLim coordinate [-180 - 180] is converted prior to passing it to the method.
 The resulting pointXYindex is ultimately passed to the java method \sQuote{readDataSlice} after subsetting.
The function also takes into account the possibility of selection of \dQuote{strips}, i.e.,
a selection across a whole or part of a meridian/parallel.
}
\author{
J Bedia \email{joaquin.bedia@gmail.com}
}
\references{
\url{http://www.unidata.ucar.edu/software/thredds/v4.3/netcdf-java/v4.3/javadocAll/ucar/nc2/dt/grid/GridCoordSys.html#findXYindexFromLatLon\%28double,\%20double,\%20int[]\%29}
}
\keyword{internal}

