% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findVerticalLevel.R
\name{findVerticalLevel}
\alias{findVerticalLevel}
\title{Finds vertical level from variable definition}
\usage{
findVerticalLevel(var)
}
\arguments{
\item{var}{Character string defining the (standandar) variable defined}
}
\value{
A list with the variable name (string) and vertical level (double)
}
\description{
Parses the variable name as passed by the 'var' argument and extracts the level and variable name
}
\details{
The level output of this function is passed to \code{\link{getVerticalLevelPars}}
}
\author{
J Bedia \email{joaquin.bedia@gmail.com}
}
\keyword{internal}

