% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biasCorrection.R
\name{flatMemberDim}
\alias{flatMemberDim}
\title{Concatenate members}
\usage{
flatMemberDim(grid, station)
}
\arguments{
\item{grid}{Input (multimember) grid}
}
\value{
A grid without members, with additional attributes to retrieve the original structure after bias correction
}
\description{
Concatenate members as a single time series for using their joint distribution in bias correction
}
\seealso{
\code{\link{recoverMemberDim}}, for recovering the original structure after bias correction.
}
\author{
J Bedia
}
\keyword{internal}
