% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get2DmatCoordinates}
\alias{get2DmatCoordinates}
\title{Get grid coordinates as 2D matrix}
\usage{
get2DmatCoordinates(grid)
}
\arguments{
\item{grid}{An input grid}
}
\value{
A 2D matrix of x-y coordinates (in this order)
}
\description{
Obtain grid coordinates as 2D matrix
}
\author{
J. Bedia
}
\keyword{internal}
