% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/misc.R
\name{getCoordinates}
\alias{getCoordinates}
\title{Get geographical coordinates of a climate data object}
\usage{
getCoordinates(obj)
}
\arguments{
\item{obj}{Any object extending the station or field classes}
}
\value{
A list with x and y components
}
\description{
Returns the coordinates of a climate data object, either stations
 or field
}
\author{
J. Bedia \email{joaquin.bedia@gmail.com}
}

