% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getLatLonDomainStations.R
\name{getLatLonDomainStations}
\alias{getLatLonDomainStations}
\title{Define geolocation of station datasets}
\usage{
getLatLonDomainStations(lonLim, latLim, lons, lats)
}
\arguments{
\item{lonLim}{Numeric. Definition of geolocation in X}

\item{latLim}{Numeric. Definition of geolocation in Y}

\item{lons}{Numeric. All available X coordinates in dataset}

\item{lats}{Numeric. All available Y coordinates in dataset}
}
\value{
A list with index positions of the requested data and a 2D matrix
of XY coordinates
}
\description{
Given the geolocation arguments, finds the index positions and coordinates of the
requested data, either single point loactions (finds closest --euclidean-- point) or
data within the given bounding box.
}
\author{
J. Bedia \email{joaquin.bedia@gmail.com}
}
\keyword{internal}

