% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getShape}
\alias{getShape}
\title{Retrieve array shape}
\usage{
getShape(obj, dimension = NULL)
}
\arguments{
\item{obj}{A grid or station object}
}
\value{
An integer vector with dim values, labelled with the \code{"dimension"} attribute names
}
\description{
Retrieve array attributes 'dimensions' and 'dim'
}
\author{
J. Bedia
}
\keyword{internal}

