% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getTimeDomain.R
\name{getTimeDomain}
\alias{getTimeDomain}
\title{Selection of time slices of gridded datasets}
\usage{
getTimeDomain(grid, dic, season, years, time, aggr.d, aggr.m)
}
\arguments{
\item{grid}{A java \sQuote{GeoGrid}}

\item{dic}{dictionary output, as returned by \code{dictionaryLookup} or NULL is not in use}

\item{season}{Vector of months defining the requested season. Passed by
\code{loadGridDataset}.}

\item{years}{Vector of selected years. Passed by \code{loadGridDataset}.}

\item{time}{Verification time defined as a character string (e.g. \dQuote{06} for data
verifying at 06:00:00). Only applies for sub-daily datasets.}

\item{aggr.d}{}

\item{aggr.m}{}
}
\value{
A list with the selected verification dates and a list
 with the index values defined as java objects of the class \sQuote{ucar.ma2.Range}.
 Output passed to \code{loadGridDataset}.
}
\description{
Performs the selection of time slices based on season and year specification
 (including year-crossing seasons). Sub-routine of \code{loadGridDataset}
}
\details{
The indices of time positions are returned as a list, in order to read
discontinuous data in the time dimension from NetCDF files in a more efficient way
(i.e. seasons for different years). The corresponding calendar dates (as POSIXct), are returned
in the \code{dateSliceList} element.
}
\author{
J. Bedia \email{joaquin.bedia@gmail.com}
}
\keyword{internal}

