% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getTimeDomainStations.R
\name{getTimeDomainStations}
\alias{getTimeDomainStations}
\title{Time  index positions for station dataset selections}
\usage{
getTimeDomainStations(timeDates, season, years)
}
\arguments{
\item{timeDates}{a POSIXlt vector of time dates}

\item{season}{A vector of months defining the season selected}

\item{years}{A vector of (continuous) year selection}
}
\value{
A list with a vector of time index positions and the corresponding POSIXlt dates
}
\description{
Get time index positions for loading ascii station data
}
\author{
J. Bedia \email{joaquin.bedia@gmail.com}
}
\keyword{internal}

