% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getVerticalLevelPars}
\alias{getVerticalLevelPars}
\title{Definition of vertical dimension slices}
\usage{
getVerticalLevelPars(grid, level)
}
\arguments{
\item{gcs}{An object of the java class \sQuote{GeoGrid})}

\item{level}{Vertical level. Passed by \code{loadGridDataset}, obtained via \code{findVerticalLevel}}
}
\value{
A list with the level value and either a java Range or a java null reference
defining the index of the vertical axis (null if no vertical levels exist)
}
\description{
Returns the selected level value (if any) and a suitable java structure. This is a subroutine
of \code{loadGridDataset}, whose output is passed to \code{makeSubset}.
}
\details{
The function opens the grid and checks all possible level values.
}
\author{
J. Bedia \email{joaquin.bedia@gmail.com}
}
\keyword{internal}

