% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biasCorrection.R
\name{getWindowIndex}
\alias{getWindowIndex}
\title{Get the index of the window days.}
\usage{
getWindowIndex(y, newdata, window, delta.method = FALSE)
}
\arguments{
\item{y}{A grid or station data containing the observed climate data for the training period}

\item{newdata}{A grid containing the simulated climate for the test period.}

\item{window}{vector of length = 2 specifying the time window width used to calibrate and the target days (days that are being corrected).
The window is centered on the target day/s (window width >= target days).}

\item{delta.method}{logical (default is FALSE)}

\item{method}{method applied. Current accepted values are \code{"eqm"}, \code{"delta"},
\code{"scaling"}, \code{"gqm"} and \code{"gpqm"} \code{"variance"},\code{"loci"} and \code{"ptr"}. See details.}
}
\description{
Get the index of the days that corresponding to the window and the target days centered in it.
}
\author{
M. Iturbide
}
\keyword{internal}
