% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getYearsAsINDEX}
\alias{getYearsAsINDEX}
\title{Get years as a factor}
\usage{
getYearsAsINDEX(obj)
}
\arguments{
\item{obj}{Any object extending the station or field classes}
}
\value{
A vector of years of the same length as the time dimension of the object,
seasonally-adjusted in the case of year-crossing seasons (e.g. DJF). See details.
}
\description{
Extract the year as a factor (e.g. for computing annual statistics)
}
\details{
The function performs a very basic operation, extracting the year element from the
dates previously converted to POSIXlt. The trick lies in the year-crossing seasons. For instance:
 by convention, winter 2001 encompasses December 2000 and January, February 2001. Therefore, in order to compute
annual statistics for a year-crossing season, it is necessary to modify first the vector of years,
and assign year 2001 to the preceding December. Similarly, the next December 2001 belongs to winter 2002,
 and so on... The function is useful for computing and/or plotting annual statistics, seasonal climatologies ...
}
\section{Warning}{

The function should no be used to extract the actual years vector
}
\examples{
data(iberia_ncep_hus850)
getSeason(iberia_ncep_hus850)
# Winter 1991-2010
range(iberia_ncep_hus850$Dates$start)
## Time series for the first point
# Dates vector
time <- as.POSIXlt(iberia_ncep_hus850$Dates$start, tz = "GMT")
hus850 <- iberia_ncep_hus850$Data[ ,1,1]
plot(time, hus850, ty = "l")
## Computation of the annual series for winter specific humidity:
par(mfrow = c(2,1))
## Wrong:
years <- as.POSIXlt(iberia_ncep_hus850$Dates$start)$year + 1900
x <- tapply(hus850, INDEX = list(years), FUN = mean)
plot(unique(years), x, ty = "b")
points(1990, x[1], col = "red", cex = 2, lwd = 2)
## Correct:
years <- getYearsAsINDEX(iberia_ncep_hus850)
x <- tapply(hus850, INDEX = years, FUN = mean)
plot(unique(years), x, ty = "b")
par(mfrow = c(1,1))
}
\author{
J. Bedia \email{joaquin.bedia@gmail.com}
}

