% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscale.train.R
\name{glm.train}
\alias{glm.train}
\title{Donwscaling with generalized linear models (GLM).}
\usage{
glm.train(x, y, family)
}
\arguments{
\item{x}{The grid data. Class: matrix.}

\item{y}{The observations data. Class: matrix.}

\item{family}{Family of the generalized linear model according to \code{\link[stats]{family}}. This is an optional parameter and there is no default.}
}
\value{
The GLM model as returned from \code{\link[stats]{glm}}.
}
\description{
Donwscaling with generalized linear models (GLM) with the base function \code{\link[stats]{glm}}.
}
\details{
The indicated optional parameters MUST be included in \code{\link[downscaleR]{downscale.predict}}, in this case \code{family} is the only optional parameter. This function is internal and should not be used by the user. The user should use \code{\link[downscaleR]{downscale.predict}}.
}
\author{
J. Bano-Medina
}
