% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biasCorrection.R
\name{gqm}
\alias{gqm}
\title{Gamma Quantile Mapping method for bias correction}
\usage{
gqm(o, p, s, precip, pr.threshold)
}
\arguments{
\item{o}{A vector (e.g. station data) containing the observed climate data for the training period}

\item{p}{A vector containing the simulated climate by the model for the training period.}

\item{s}{A vector containing the simulated climate for the variable used in \code{x}, but considering the test period.}

\item{precip}{Logical indicating if o, p, s is precipitation data.}

\item{pr.threshold}{The minimum value that is considered as a non-zero precipitation. Ignored for
\code{varcode} values different from \code{"pr"}. Default to 1 (assuming mm).}
}
\description{
Implementation of Gamma Quantile Mapping method for bias correction
}
\author{
S. Herrera and M. Iturbide
}
\keyword{internal}

