% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{grabGrob}
\alias{grabGrob}
\title{Obtain a grob object from ordinary plot}
\usage{
grabGrob()
}
\value{
A grob object
}
\description{
Obtain a grob object from ordinary plot
}
\author{
M. Iturbide
}
\keyword{internal}
