% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/grid2Netcdf.R
\name{grid2NetCDF}
\alias{grid2NetCDF}
\title{Write NetCDF}
\usage{
grid2NetCDF(data, NetCDFOutFile = "out.nc4", missval = 1e+20,
  globalAttributes = NULL, varAttributes = NULL, prec = "float",
  compression = 4, shuffle = TRUE)
}
\arguments{
\item{data}{A grid data object coming from \code{\link{loadGridData}} or \code{\link{interpGridData}}
or the function \code{\link[ecomsUDG.Raccess]{loadECOMS}} of package \pkg{ecomsUDG.Raccess}.}

\item{NetCDFOutFile}{Name of the file created by the function. ("out.nc4", default)}

\item{missval}{Value associated with the NA value used in R. (1e20, default)}

\item{globalAttributes}{List of global attributes included in the NetCDF file. (NULL, default)}

\item{varAttributes}{List of attributes to be included in the variable written in the NetCDF file. (NULL, default)}

\item{prec}{Precision to write the attribute. If not specified, the written precision is the same as the variable whose attribute this is. This can be overridden by specifying this argument with a value of "short", "float", "double", or "text".}

\item{compression}{If set to an integer between 1 (least compression) and 9 (most compression), this enables compression for the variable as it is written to the file. Turning compression on forces the created file to be in netcdf version 4 format, which will not be compatible with older software that only reads netcdf version 3 files. (4, default)}

\item{shuffle}{Turns on (if TRUE) or off (if FALSE, the default) the shuffle filter. According to netcdf docs, turning the shuffle filter on can improve compression for integer variables. Turning the shuffle filter on forces the created file to be in netcdf version 4 format, which will not be compatible with older software that only reads netcdf version 3 files.}
}
\value{
A NetCDF-4 file with the variable and attributes defined in the inputs.
}
\description{
Export downscaleR object to NetCDF
}
\examples{
\dontrun{
# The following panels show an illustrative use of ECOMS-UDG and downscaleR to obtain the bias corrected series of mean temperature for the period DJFMAM (one-month lead time; i.e. with the initializations from November) over Europe. WFDEI is used as reference.
# Note that, in order to facilitate the use of the resulting bias corrected data in different impact applications, the resulting bias corrected data can be easily exported to netcdf format.

obs <- loadECOMS(dataset = "WFDEI", var = "tas", season = c(12,1,2,3,4,5), lonLim = c(-15,35), latLim = c(32, 75), years = c(2001:2010))
prd <- loadECOMS(dataset = "System4_seasonal_15", var = "tas", time = "DD", season = c(12,1,2,3,4,5), members = 1:2, leadMonth = 1, lonLim = c(-15,35), latLim = c(32, 75), years = c(2001:2010))
prd <- interpGridData(prd, new.grid = getGrid(obs), method = "nearest");

# Bias correction and plotting
prd.bc <- biasCorrection(obs, prd, prd, method = "qqmap", multi.member = TRUE, window = 30)
plotMeanField(obs)
plotMeanField(prd, multi.member = FALSE)
plotMeanField(prd.bc, multi.member = FALSE)

# Exporting to netcdf4
fileName <- "tas_qqmap_System4_WFDEI_2001_2010.nc4"
grid2NetCDF(prd.bc, NetCDFOutFile = fileName, missval = 1e20, prec = "float")
# Including attributes:
globalAttributes <- list(institution = "Santander Meteorology Group, http://www.meteo.unican.es/")
varAttributes <- list(description = "Bias corrected daily mean temperature.")
grid2NetCDF(prd.bc, NetCDFOutFile = fileName, missval = 1e20, prec = "float", globalAttributes = globalAttributes, varAttributes = varAttributes)
}
}
\author{
Sixto Herrera \email{herreras@unican.es}, Wietse Franssen \email{wietse.franssen@wur.nl}
}
\references{
\itemize{
\item David Pierce \email{dpierce@ucsd.edu}, Interface to Unidata netCDF (version 4 or earlier) format data files, http://dwpierce.com/software
}
}

