% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/grid2PCs.R
\name{grid2PCs}
\alias{grid2PCs}
\title{Projection of a grid onto an EOF}
\usage{
grid2PCs(prinCompObj, grid, n.pcs = NULL)
}
\arguments{
\item{prinCompObj}{An object created by \code{\link{prinComp}}}

\item{grid}{Input grid to project. Multigrids are not allowed.}

\item{n.pcs}{Number of principal components to be retained. Default to the number of EOFs contained in the input \code{prinCompObj}.}
}
\value{
A list, each component corresponding to a member (a list of length 1 if the projected grid is not multimember),
containing a matrix of PCs, being the number of columns defined by \code{n.pcs}, and the number of rows by the
length of the time dimension.
}
\description{
Projection of a grid onto a user-defined EOF
}
\details{
The function is intended to project a grid onto user specified EOFs, contained in \code{prinCompObj}.
Note that the grid should have the same spatial extent and resolution than the original EOF, so input grids from
different models should be first interpolated. Also, to ensure consistency, both objects should contain the same
variable name.

Note that the function is not currently implemented to deal with decadal predictions
 (i.e., 'runtime' dimension is not handled)
}
\author{
J Bedia
}
\seealso{
\code{\link{prinComp}} for EOF analysis.
}

