% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid3Dto2Dmatrix.R
\name{grid3Dto2Dmatrix}
\alias{grid3Dto2Dmatrix}
\title{Conversion of a 3D grid to a 2D matrix}
\usage{
grid3Dto2Dmatrix(grid3D)
}
\arguments{
\item{grid3D}{A 3-dimensional grid with longitude, latitude and time dimensions.}
}
\value{
A 2-dimensional matrix with time in rows and grid-points in columns.
}
\description{
Converts 3D grids of the form [time,lat,lon]
 to 2D matrices of the form [time, grid-point].
}
\details{
Global attributes of the output matrix are provided for metadata description and spatiotemporal collocation.
}
\author{
J. Bedia
}

