% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rdas.R
\docType{data}
\name{Iberia_NCEP_tas}
\alias{Iberia_NCEP_tas}
\title{Field containing NCEP reanalysis data of temperature for the Iberian Peninsula.}
\format{A field}
\source{
subset of NCEP reanalysis data, which is accessible
through the \strong{ECOMS User Data Gateway (ECOMS-UDG)}
}
\description{
Iberia_NCEP_tas is an field object returned by loadECOMS
function (package ecomsUDG.Raccess):
season = c(12,1,2),
years = 1991-2000,
var="tas",
lonLim = c(-10,5),
latLim= c(34,44),
time = "DD",
aggr.d = "min".
}
\seealso{
\code{\link{makeMultiField}}, \code{\link[ecomsUDG.Raccess]{loadECOMS}}
}
\keyword{NCEP}
\keyword{reanalysis}

