% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interannualOutlook.R
\name{interannualOutlook}
\alias{interannualOutlook}
\title{Plots annual series of the inputs and the output of downscaling}
\usage{
interannualOutlook(obs, pred, downscaled, location = c(-40, -5),
  yrange = NULL)
}
\arguments{
\item{obs}{Field of observations.}

\item{pred}{Field of model data.}

\item{downscaled}{Field of downscaling output.}

\item{location}{coordinates of a location in the geographic domain of the field.}

\item{yrange}{equivalent to ylim argument for plot function}
}
\value{
A plot
}
\description{
Plots annual series of the inputs and the output of downscaling
}
\author{
M. Iturbide \email{maibide@gmail.com}
}

