% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isimip3.R
\name{isimip3}
\alias{isimip3}
\title{ISIMIP v3 method for bias correction}
\usage{
isimip3(
  o,
  p,
  s,
  dates,
  lower_bound = c(NULL),
  lower_threshold = c(NULL),
  upper_bound = c(NULL),
  upper_threshold = c(NULL),
  randomization_seed = NULL,
  detrend = array(data = FALSE, dim = 1),
  rotation_matrices = c(NULL),
  n_quantiles = 50,
  distribution = c("normal"),
  trend_preservation = array(data = "additive", dim = 1),
  adjust_p_values = array(data = FALSE, dim = 1),
  if_all_invalid_use = c(NULL),
  invalid_value_warnings = FALSE
)
}
\arguments{
\item{o}{A vector (e.g. station data) containing the observed climate data for the training period}

\item{p}{A vector containing the simulated climate by the model for the training period.}

\item{s}{A vector containing the simulated climate for the variable used in \code{p}, but considering the test period.}

\item{dates}{List containing the dates corresponding to Observations (obs_hist), training period (sim_hist) and test period (sim_fut). 
Every array represents the years of the time steps of the time series data. Used for detrending.}

\item{lower_bound}{: list of floats, optional. Lower bounds of values in x_obs_hist, x_sim_hist, and x_sim_fut.}

\item{lower_threshold}{: list of floats, optional. Lower thresholds of values in x_obs_hist, x_sim_hist, and x_sim_fut. 
All values below this threshold are replaced by random numbers between lower_bound and lower_threshold before bias adjustment.}

\item{upper_bound}{: list of floats, optional. Upper bounds of values in x_obs_hist, x_sim_hist, and x_sim_fut.}

\item{upper_threshold}{: list of floats, optional. Upper thresholds of values in x_obs_hist, x_sim_hist, and x_sim_fut. 
All values above this threshold are replaced by random numbers between upper_threshold and upper_bound before bias adjustment.}

\item{randomization_seed}{: int, optional. Used to seed the random number generator before replacing invalid values and values beyond 
the specified thresholds.}

\item{detrend}{: list of booleans, optional. Detrend time series before bias adjustment and put trend back in afterwards.}

\item{rotation_matrices}{: list of (n,n) ndarrays, optional. List of orthogonal matrices defining a sequence of rotations in variable 
space, where n is the number of variables.}

\item{n_quantiles}{: int, optional. Number of quantile-quantile pairs used for non-parametric quantile mapping.}

\item{distribution}{: list of strs, optional. Kind of distribution used for parametric quantile mapping:
{"normal", "weibull", "gamma", "beta", "rice"}.}

\item{trend_preservation}{: list of strs, optional. Kind of trend preservation used for non-parametric quantile mapping: 
{"additive", "multiplicative", "mixed", "bounded"}.}

\item{adjust_p_values}{: list of booleans, optional. Adjust p-values for a perfect match in the reference period.}

\item{if_all_invalid_use}{: list of floats, optional. Used to replace invalid values if there are no valid values. An error is raised 
if there are no valid values and this parameter is None.}

\item{invalid_value_warnings}{: boolean, optional. Raise user warnings when invalid values are replaced bafore bias adjustment.}
}
\description{
Implementation of ISIMIP v3 method for bias correction:
  1. Replaces invalid values in time series.
  2. Detrends time series if desired.
  3. Replaces values beyond thresholds by random numbers.
  4. Adjusts inter-variable copula.
  5. Adjusts marginal distributions for every variable.
  6. Replaces values beyond thresholds by the respective bound.
  7. Restores trends.
  Removes invalid values from masked arrays and store resulting numpy arrays
}
\author{
S. Herrera and M. Iturbide
}
\keyword{internal}
