% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateGrid.R
\name{latWeighting}
\alias{latWeighting}
\title{Weight as a function of latitude cosine}
\usage{
latWeighting(grid)
}
\description{
Computes weights as a function of latitude to compute areal averages
}
\author{
J. Bedia, S. Siegert
}
\keyword{internal}

