% Generated by roxygen2 (4.0.1): do not edit by hand
\name{loadGridData}
\alias{loadGridData}
\alias{loading}
\title{Load a gridded dataset}
\usage{
loadGridData(dataset, var, dictionary = TRUE, lonLim = NULL,
  latLim = NULL, season = NULL, years = NULL, time = "none")
}
\arguments{
\item{dataset}{}

\item{var}{}

\item{dictionary}{}

\item{lonLim}{}

\item{latLim}{}

\item{season}{}

\item{years}{}

\item{time}{}
}
\value{
a list of elements
}
\description{
Load a user-defined spatio-temporal slice from a gridded dataset
}
\author{
J. Bedia \email{joaquin.bedia@gmail.com}
}

