% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loadStationData.R
\name{loadStationData.ASCII}
\alias{loadStationData.ASCII}
\title{Load station data in standard ASCII format}
\usage{
loadStationData.ASCII(dataset, var, stationID, lonLim, latLim, season, years,
  tz)
}
\arguments{
\item{dataset}{Directory containing the stations dataset}

\item{var}{Character string. Name of the variable, as defined in the dataset}

\item{stationID}{Character string. Optional, Id code(s) of the station(s) selected}

\item{lonLim}{numeric vector of length 1 or two defining X coordinate(s).
See \code{\link{loadStationData}} for details.}

\item{latLim}{numeric vector of length 1 or two defining Y coordinate(s).
See \code{\link{loadStationData}} for details.}

\item{season}{Numeric vector of months defining the season}

\item{years}{Numeric vector of years}
}
\value{
A list with several components. See \code{\link{loadStationData}} for details.
}
\description{
Load station data in standard ASCII format, being the standard defined in the framework
of the action COST VALUE
}
\author{
J. Bedia \email{joaquin.bedia@gmail.com}
}
\references{
\url{https://github.com/SantanderMetGroup/downscaleR/wiki/Observation-Data-format}
}
\keyword{internal}

