% Generated by roxygen2 (4.0.1): do not edit by hand
\name{loadStationData}
\alias{loadStationData}
\alias{loading}
\title{Load station data}
\usage{
loadStationData(source.dir, file.format = c("ascii", "netcdf"), var,
  stationID = NULL, lonLim = NULL, latLim = NULL, season = NULL,
  years = NULL, tz = "GMT")
}
\arguments{
\item{source.dir}{A valid path to the directory containing the station files}

\item{file.format}{Wether the stations data are stored in a netCDF or ASCII (default) file. See details for standard format definition.}

\item{var}{Character string indicating the variable to be loaded. Note that the notation depends on wether the dictionary is being used or not.}

\item{dictionary}{Optional. A full path to the file containing the dictionary. See details}

\item{stationID}{Optional. A character vector indicating the code names of the stations to be loaded.}

\item{lonLim}{Vector of length = 2, with minimum and maximum longitude coordinates of the bounding box
selected. For single-point queries, a numeric value with the longitude coordinate. See details.}

\item{latLim}{Same as \code{lonLim}, but for the selection of the latitudinal range. See details.}

\item{season}{An integer vector specifying the desired season (in months, January = 1 ..., December = 12).
Options include one to several (contiguous) months. If NULL (default), full year selections is performed (same as \code{season = 1:12})}

\item{years}{Optional vector of years to select. Default (NULL) to all available years}

\item{tz}{A time zone specification to be used for the conversion of dates, if one is required
(i.e., if the time zone of the dataset does not correspond to the system-specific one; see
\code{\link[base]{timezones}} for details). The default assumes "GMT" (UTC, Universal Time, Coordinated),
but be very careful as station datasets are quite heterogeneous.}
}
\value{
a list with the following elements:
\itemize{
\item \code{Variable}. Name of the variable
\item \code{Data}. Dates are ordered by rows and Stations in columns. Names are station codes
\item \code{xyCoords}. A 2-D matrix with longitude and latitudes of the stations
\item \code{Dates}. A list with the verification time interval of each record in the time series.
 This is represented by a list with two elements: \code{start} and \code{end}, representing the
 lower and upper bounds of the verification period
\item \code{Metadata}. A list of variable length depending on the available metadata associated
to each observation. If no metadata are provided, at least the station codes (compulsory) are displayed.
}
}
\description{
Load observations data from station datasets in standard ASCII format.
}
\author{
J. Bedia \email{joaquin.bedia@gmail.com}
}
\references{
\url{https://github.com/SantanderMetGroup/downscaleR/wiki/Observation-Data-format}
}
\seealso{
\code{\link{loadGridData}}, \code{\link{dataInventory}}
}

