% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biasCorrection.R
\name{loci}
\alias{loci}
\title{Local intensity scaling of precipitation}
\usage{
loci(o, p, s, precip, pr.threshold)
}
\arguments{
\item{o}{A vector (e.g. station data) containing the observed climate data for the training period}

\item{p}{A vector containing the simulated climate by the model for the training or test period.}

\item{s}{A vector containing the simulated climate for the variable used in \code{p}, but considering the test period.}

\item{precip}{Logical indicating if o, p, s is precipitation data.}

\item{pr.threshold}{The minimum value that is considered as a non-zero precipitation. Ignored when 
\code{precip = FALSE}. See details in function \code{biasCorrection}.}
}
\description{
Implementation of Local intensity scaling of precipitation method for bias correction based on Vincent Moron's local_scaling function in weaclim toolbox in Matlab
}
\author{
B. Szabo-Takacs
}
