% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loginUDG.R
\name{loginUDG}
\alias{loginUDG}
\title{Login to the Santander Met Group Thredds Data Server}
\usage{
loginUDG(username, password, proxy.host = NULL, proxy.port = NULL)
}
\arguments{
\item{username}{A character string with a valid user ID. See details.}

\item{password}{Character string. Authorized password. See details.}

\item{proxy.host}{In case of a proxy connection, the name of the proxy host, as a character string}

\item{proxy.port}{In case of a proxy connection, an integer indicating the proxy port}
}
\description{
Provides HTTP authentication for accessing the Santander Met Group's THREDDS Data Server
}
\details{
The Santander Met Group has deployed a THREDDS Data Server to enable access to different climate databases/data collections,
implementing a fine-grained user authorization and using remote data access protocols with data subsetting capabilities.
Prior to data access, users must log in. Registration can be obtained via the THREDDS Administration Panel (\href{http://www.meteo.unican.es/tap}{TAP}),
 indicating the group (Project) you belong to (e.g. CORDEX, SPECS ...), which will grant access to certain databases.
 Further details on registration for data access can be viewed in this \href{http://meteo.unican.es/ecoms-udg/DataServer/Registration}{example link}.
}
\author{
J Bedia \email{joaquin.bedia@gmail.com}, M. Vega.
}

