% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMultiGrid.R
\name{makeMultiGrid}
\alias{makeMultiGrid}
\title{Multigrid constructor}
\usage{
makeMultiGrid(..., spatial.tolerance = 0.001)
}
\arguments{
\item{...}{Input grids to form the multifield. These must be compatible in time and space (see details).}

\item{spatial.tolerance}{numeric. Coordinate differences smaller than \code{spatial.tolerance} will be considered equal 
coordinates. Default to 0.001 --assuming that degrees are being used it seems a reasonable rounding error after interpolation--.
This value is passed to the \code{\link{identical}} function to check for spatial consistency of the input grids.}
}
\value{
A (multimember) multigrid object encompassing the different input (multimember) grids
}
\description{
Constructs a (possibly multimember) multigrid from different (multimember) grids.
A multigrid can be considered as a \dQuote{stack} of grids with similar spatiotemporal extents,
 useful to handle sets of predictors as a single block.
}
\details{
The function makes a number of checks in order to test the spatiotemporal compatibility of the input multi-member grids.
 Regarding the temporal concordance, it is implicitly assumed that all temporal data from the different
 multimember grids correspond to the same time zone ("GMT"). The time zone itself is not important, as long as it is the
 same across datasets, because temporal consistency is checked on a daily basis (not hourly), allowing the inclusion of 
 predictors with different verification times and temporal aggregations. For instance, instantaneous geopotential at 12:00 
 is compatible with mean daily surface temperature, always that both variables correspond to the same days. Different time 
 resolutions are not compatible and will return an error (for instance, 6-hourly data is incompatible with daily values, 
 because their respective time series for a given season have different lengths).
 
 The spatial consistency of the input grids is also checked. In order to avoid possible errors from the user, the spatial
  consistency (i.e., equal XY coordinates) of the input grids must be ensured before attempting the creation of the multigrid,
  otherwise giving an error. This can be achieved either through the specification of the same 'lonLim' and 'latLim' argument
  values when loading the grids, or using the \code{\link{interpData}} interpolator in conjuntion with the \code{\link{getGrid}}
  method.
}
\note{
A multigrid can not be passed to the interpolator \code{\link{interpData}} directly. Instead, the 
multimember grids should be interpolated individually prior to multigrid construction.
}
\examples{
# Creation of a multigrid from three different grids:
data(iberia_ncep_ta850)
data(iberia_ncep_hus850)
data(iberia_ncep_psl)
# An example of different temporal aggregations, temporally compatible: 
# sea-level pressure is a daily mean, while specific humidity and air temperature 
# (850 mb surface isobaric pressure level) are instantaneous data verifying at 12:00 UTC:
# air temperature
range(iberia_ncep_ta850$Dates$start)
range(iberia_ncep_ta850$Dates$end) # start and end are identical (instantaneous)
# sea-level pressure
range(iberia_ncep_psl$Dates$start)
range(iberia_ncep_psl$Dates$end) # start and end differ in 24 h (daily mean)
mf <- makeMultiGrid(iberia_ncep_hus850, iberia_ncep_psl, iberia_ncep_ta850)
# The new object inherits the global attributes from the first grid, as it is assumed
# that all input grids come from the same data source:
attributes(mf)
# The data structure has now one additional dimension ("var"), along which the data arrays
# have been binded:
str(mf$Data)
plotMeanGrid(mf)

# Example of multimember multigrid creation from several multimember grids:
# Load three different multimember grids with the same spatiotemporal ranges:
data(tasmax_forecast)
data(tasmin_forecast)
data(tp_forecast)
mm.mf <- makeMultiGrid(tasmax_forecast, tasmin_forecast, tp_forecast)
# 'plotMeanGrid' can just handle the multi-member mean for each variable in this case:
plotMeanGrid(mm.mf)
}
\author{
J. bedia
}
\seealso{
\code{\link{interpData}} for spatial consistency of input grids.
}

