% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSubset.R
\name{makeSubset}
\alias{makeSubset}
\title{Reads an arbitrary data slice}
\usage{
makeSubset(grid, timePars, levelPars, latLon)
}
\arguments{
\item{timePars}{A list of time parameters as returnde by \code{getTimeDomain}.}

\item{levelPars}{Vertical level parameters, as returned by \code{getVerticalLevelPars}.
The element \code{zRange} is a \sQuote{ucar.ma2.Range} or a null reference.}

\item{latLon}{A list of geospatial parameters, as returned by
\code{getLatLonDomain}.}

\item{A}{grid of the java class \sQuote{ucar.nc2.dt.grid.GeoGrid}}
}
\value{
A n-dimensional array with the selected subset data.
}
\description{
Reads an arbitrary data slice from a new GeoGrid that is a logical subset of
the original GeoGrid.
}
\note{
The process is somewhat tricky because R cannot adequately represent NDjavaArrays.
Several tests have shown that R (unlike MatLab) puts zeroes where they shouldn't be
 when using the \sQuote{copyToNDjavaArray} method. Hence, the data returned from
 \sQuote{readDataSlice} are converted to a 1D array (\sQuote{copyTo1DjavaArray} method),
and then put back in a n-dimensional R array. This is achieved by reversing
the GeoGrid dimension ordering. In addition, the particular case of single-coordinate selections
(either in X and/or in Y dimensions) must be adequately handled given that the \sQuote{makeSubset} output 
has in this case a different shape than the final output after \dQuote{slicing}.
}
\author{
J. Bedia \email{joaquin.bedia@gmail.com}, with contributions of A. Cofin\~no and S. Herrera.
}
\references{
\url{https://www.unidata.ucar.edu/software/thredds/v4.3/netcdf-java/v4.3/javadoc/ucar/nc2/dt/grid/GeoGrid.html}
}
\keyword{internal}

