% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mat2Dto3Darray}
\alias{mat2Dto3Darray}
\title{Conversion 2D matrix into a 3D array}
\usage{
mat2Dto3Darray(mat2D, x, y)
}
\arguments{
\item{mat2D}{A 2D matrix with time in rows and lonlat in columns, as returned
by \code{\link{array3Dto2Dmat}}}

\item{x}{unique X coordinates of the points, in ascending order}

\item{y}{As argument \code{x}, for the Y coordinates}
}
\value{
A 3-dimensional array with the dimensions ordered: [time,lat,lon]
}
\description{
Converts a 2D matrix of the form [time, lonlat] to a 3D array of the form
[time,lat,lon], in this order. Mainly for PCA analysis and field reconstruction.
}
\details{
The function is the inverse of \code{\link{array3Dto2Dmat}}
}
\author{
J. Bedia \email{joaquin.bedia@gmail.com}
}
\seealso{
\code{\link{array3Dto2Dmat}}, which performs the inverse operation.
}
\keyword{internal}

