% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aggregateGrid.R
\name{memberAggregation}
\alias{memberAggregation}
\title{Member aggregation}
\usage{
memberAggregation(grid, aggr.mem, parallel, max.ncores, ncores)
}
\arguments{
\item{grid}{A multimember grid to apply the aggregation}

\item{aggr.mem}{Character string indicatins the aggregation function}

\item{parallel.pars}{Arguments defining the parallelization options, as passed by \code{\link{parallelCheck}}}
}
\description{
Aggregate a grid along its member dimension
}
\details{
The function preserves the metadadata associated with member information (i.e. initialization dates and member names). In addition,
an attribute indicating the member aggregation function is added to the \code{Variable} component.
}
\author{
J. Bedia
}
\keyword{internal}

