% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsMLR.R
\name{mlr.predict}
\alias{mlr.predict}
\title{Donwscaling with a given multiple linear regression model (MLR).}
\usage{
mlr.predict(x, weights, info)
}
\arguments{
\item{x}{The grid data. Class: matrix.}

\item{weights}{Object as returned from \code{\link[downscaleR]{mlr.train}}}

\item{info}{This is a list containing the information/parameters used in \code{\link[downscaleR]{mlr.train}}. Basically is used to distinguish between Least Squares (LS) and Moore-Penrose (MP).}
}
\value{
The predicted matrix.
}
\description{
Donwscaling with a given multiple linear regression (MLR) calculated in \code{\link[downscaleR]{downscale.predict}} via \code{\link[downscaleR]{mlr.train}}.
}
\details{
Predicts by using the base function \code{\link[stats]{predict}}. This function is internal and should not be used by the user. The user should use \code{\link[downscaleR]{downscale.predict}}.
}
\author{
J. Bano-Medina
}
