% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsMLR.R
\name{mlr.train}
\alias{mlr.train}
\title{Donwscaling with multiple linear regression (MLR).}
\usage{
mlr.train(x, y, fitting = "LS")
}
\arguments{
\item{x}{The grid data. Class: matrix.}

\item{y}{The observations data. Class: matrix.}

\item{fitting}{Type of fitting used: c("LS","MP","MP+L2"). There are three options: least squares ("LS") that downscales via \code{\link[stats]{lm}}, Moore-Penrose ("MP") that uses the generalized inverse via \code{\link[MASS]{ginv}} or the L2 penalty, commonly known as ridge regression ("MP+L2"). This is an optional parameter and the default is "LS".}
}
\value{
If fitting = "LS", then returns a MLR model as returned from \code{\link[stats]{lm}} whereas if fitting = ("MP" or "MP+L2") then returns just the estimated coefficients/weights.
}
\description{
Donwscaling with multiple linear regression (MLR).
}
\details{
If "MP" or "MP+L2" are selected, the bias is input as an additional input neuron in the weight matrix with value equal 1. The indicated optional parameters can be changed by including them in \code{\link[downscaleR]{downscale.predict}}, in this case \code{fitting} is the only optional parameter. This function is internal and should not be used by the user. The user should use \code{\link[downscaleR]{downscale.predict}}.
}
\author{
J. Bano-Medina
}
