% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mm2mm.mf}
\alias{mm2mm.mf}
\title{Constructor of a multimember multifield from various multimember fields}
\usage{
mm2mm.mf(mm.list)
}
\arguments{
\item{mm.list}{Multimember list. A list containing the different input multimember fields.}
}
\value{
A multimember multifield object
}
\description{
Constructs a multimember multifield from different multimember fields.
 Useful for the creation of multifields from the multimember fields returned by
 ecomsUDG.Raccess::loadECOMS in seasonal forecast downscaling applications.
}
\details{
The function makes a number of checks in order to test the spatiotemporal compatibility of the input multi-member fields.
 Regarding the temporal concordance, it is implicitly assumeds that all temporal data from the different
 multimember fields correspond to the same time zone ("GMT"). The time zone itself is not important, as long as it is the
 same across datasets, because temporal consistency is checked on a daily basis, allowing the inclusion of predictors with
 different verification times and temporal aggregations. For instance, instantaneous geopotential at 12:00 is compatible with
 mean daily surface temperature, always that both variables correspond to the same days. Different time resolutions are not
 compatible and will return an error (for instance, 6-hourly data is incompatible with daily values, because their
 respective time series have different lengths).
}
\author{
J. bedia \email{joaquin.bedia@gmail.com}
}
\seealso{
\code{\link[ecomsUDG.Raccess]{loadECOMS}}, \code{\link{loadMultifield}}
}

