% Generated by roxygen2 (4.0.1): do not edit by hand
\name{multiPlot}
\alias{multiPlot}
\title{Make multi-panel plots}
\usage{
multiPlot(gridData, split.dim.name, titles)
}
\arguments{
\item{gridData}{a grid dataset as returned by any of the loading functions}

\item{name}{of the dimension used for splitting: either \code{var} or \code{member} for multi-predictor and
multi-member displays respectively}
}
\value{
Prints the graphical display
}
\description{
Sub-routine of plotMeanField for dividing the graphical window into different subplots,
for multi-member or multi-variable displays
}
\author{
J Bedia \email{joaquin.bedia@gmail.com}
}
\keyword{internal}

