% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biasCorrection.R
\name{mva}
\alias{mva}
\title{Mean and Variance Adjustment}
\usage{
mva(o, p, s)
}
\arguments{
\item{o}{A vector (e.g. station data) containing the observed climate data for the training period}

\item{p}{A vector containing the simulated climate by the model for the training period.}

\item{s}{A vector containing the simulated climate for the variable used in \code{p}, but considering the test period.}
}
\description{
Mean and Variance Adjustment method for bias correction
}
\author{
M. Iturbide
}
\keyword{internal}
