% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rdas.R
\docType{data}
\name{NCEP_Iberia_tp}
\alias{NCEP_Iberia_tp}
\title{Grid containing NCEP reanalysis data of precipitation for the Iberian Peninsula.}
\format{A grid}
\source{
subset of NCEP reanalysis data, which is accessible through the \strong{ECOMS User Data Gateway (ECOMS-UDG)}
}
\description{
NCEP_Iberia_tp is a grid object returned by loadECOMS from package loadeR.ECOMS
}
\keyword{NCEP}
\keyword{reanalysis}

