% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{ndays}
\alias{ndays}
\title{Calculate the number of days of the current month}
\usage{
ndays(d)
}
\arguments{
\item{d}{A date (character) in format YYYY-MM-DD...}
}
\value{
The number of days of the current month
}
\description{
Calculate the number of days of the current month
}
\references{
\url{http://stackoverflow.com/questions/6243088/find-out-the-number-of-days-of-a-month-in-r}
}

