% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biasCorrection.R
\name{norain}
\alias{norain}
\title{norain}
\usage{
norain(o, p, threshold)
}
\arguments{
\item{o}{A vector (e.g. station data) containing the observed climate data for the training period}

\item{p}{A vector containing the simulated climate by the model for the training period.}

\item{threshold}{The minimum value that is considered as a non-zero precipitation. Ignored for
\code{varcode} values different from \code{"pr"}. Default to 1 (assuming mm).}
}
\description{
presprocess to bias correct precipitation data
}
\author{
S. Herrera and M. Iturbide
}
\keyword{internal}
