% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parallelCheck.R
\name{parallelCheck}
\alias{parallelCheck}
\title{Parallel availability check}
\usage{
parallelCheck(parallel, max.ncores = 16, ncores = NULL)
}
\arguments{
\item{parallel}{Logical, should parallel execution be used?}

\item{ncores}{number of cores used in parallel computation, self-selected number of cores
is used when \code{is.null(ncpus)} (the default).}

\item{maxcores}{upper bound for self-selected number of cores}
}
\value{
A list with two elements:
\itemize{
\item \code{hasparallel}, a logical flag indicating if parallelization is enabled and
\item \code{cl}: parallel socket cluster object, or NULL.
}
}
\description{
Check availability of parallel package and set parameters
}
\details{
The function checks if package parallel is available. Then, it is checked whether the FORK nodes
can be initialized
}
\author{
J. Bedia, relying on previous code by Jonas Bhend and Matteo de Felice
}
\keyword{internal}

