% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PC2field.R
\name{PC2field}
\alias{PC2field}
\title{Extract field data from PCs and EOFs}
\usage{
PC2field(pc.obj = NULL, orig = FALSE, EOFs = NULL, PCs = NULL)
}
\arguments{
\item{pc.obj}{a list returned by function \code{\link{prinComp}}. Default is "NULL" (see Details).}

\item{orig}{logical (default is FALSE). If TRUE and argument 'pc.obj' is used,
field data is reovered from the 'orig' slot, that contains the original
field data before PCs calculation with function \code{\link{prinComp}}.}

\item{EOFs}{matrix with EOFs in rows (default is NULL).}

\item{PCs}{matrix with PCs in columns (default is "NULL").}
}
\value{
A field/multifield (if 'pc.obj' is used) or a 2D matrix.
}
\description{
Extracts field data from PCs and EOFs
}
\details{
The field/multifield can be built from the EOFs and PCs in 'pc.obj' or from matrixes of EOFs and PCs.
}
\author{
M. Iturbide \email{maibide@gmail.com}
}

