% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEOF.R
\name{plotEOF}
\alias{plotEOF}
\title{Plot an arbitrary number of EOFs}
\usage{
plotEOF(prinCompObj, var = NULL, member = 1, n.eofs = NULL)
}
\arguments{
\item{prinCompObj}{A PCA object as returned by \code{\link{prinComp}}}

\item{var}{Character string indicating the variable whose EOFs are to be displayed. If the PCA analysis has
been applied to 1 single field, this argument can be omitted.}

\item{member}{An integer indicating the position of the member whose EOFs are to be displayed. Default 1, 
corresponding to the first member. Ignored for non multimember fields.}

\item{n.eofs}{Number of EOFs to be displayed. Default to NULL, indicating that all computed EOFS
will be represented}
}
\value{
A plot with as many panels as EOFs requested, in the original units of the variable
# @export
}
\description{
Plots an arbitrary number of EOFs. Useful to have a quick overview of the main spatial modes
 of a (possibly multimember) field.
}
\examples{
\dontrun{ 
# Winter temperature at 850 mb isobaric surface pressure level is loaded (period 1981--2010):
data(iberia_ncep_ta850)
# PCA analysis, retaining the PCs that explain 95\\\% of the total variance:
pca <- prinComp(iberia_ncep_ta850, v.exp = .95)
# Plot of all EOFs
plotEOF(pca)
# Plot the first 4 EOFs:
plotEOF(pca, n.eofs = 4)

# Example with PCA analysis of a multifield (multiple variables)
# Download dataset
dir.create("mydirectory")
download.file("http://meteo.unican.es/work/downscaler/data/Iberia_NCEP.tar.gz", 
destfile = "mydirectory/Iberia_NCEP.tar.gz")
# Extract files from the tar.gz file
untar("mydirectory/NCEP_Iberia.tar.gz", exdir = "mydirectory")
# First, the path to the ncml file is defined:
ncep <- "mydirectory/Iberia_NCEP/Iberia_NCEP.ncml"
# load geopotential heigth at 500 mb, temperature at 1000 mb and sea-level pressure
multifield <- loadMultiField(ncep, vars = c("z@500", "ta@1000", "psl"),
              season = c(12,1,2), years = 1981:2010)
# PCA analysis, retaining the first 9 PCs of each variable:
pca2 <- prinComp(multifield, n.eofs = 9)
names(pca2)
# EOFs for geopotential
plotEOF(pca2, "z")
plotEOF(pca2, "ta", n.eofs = 4)
plotEOF(pca2, "psl", n.eofs = 2)
}

}
\author{
J. Bedia
}
\seealso{
\code{\link{prinComp}}
}

