% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotMeanField}
\alias{plotMeanField}
\title{Plot a map of the mean value of a grid dataset}
\usage{
plotMeanField(gridData, multi.member = FALSE)
}
\arguments{
\item{gridData}{A grid dataset}

\item{multi.member}{Should members be plotted sepparately (TRUE), or just a plot
 of the multi-member mean (FALSE, default)?. Ignored if the dataset has no members.}
}
\value{
a plot of the mean/multi-member/multi-variable field with a world map superposed
}
\description{
Plot the spatial mean of a gridded variable, or variables in the case of multi-fields.
}
\details{
The function is a wrapper of the \code{\link[fields]{image.plot}} function
in package \pkg{fields}
}
\note{
The function plots a simple temporal mean of the loaded object in the form of
a map. It does not handle other temporal aggregations.
}
\author{
J Bedia \email{joaquin.bedia@gmail.com}
}

