% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMeanGrid.R
\name{plotMeanGrid}
\alias{plotMeanGrid}
\title{Plot a map of the mean value of a grid dataset}
\usage{
plotMeanGrid(gridData, multi.member = FALSE)
}
\arguments{
\item{gridData}{A grid dataset}

\item{multi.member}{Should members be plotted sepparately (TRUE), or just a plot
of the multi-member mean (FALSE, default)?. Ignored if the dataset has no members.}
}
\value{
a plot of the mean/multi-member/multi-variable grid with a world map superposed
}
\description{
Plot the spatial mean of a gridded variable, or variables in the case of multigrids.
}
\details{
The function is a wrapper of the \code{\link[fields]{image.plot}} function
in package \pkg{fields}.

The world lines superposed onto the maps have been extracted from the land polygons at 1:50m scale
freely available at \url{http://www.naturalearthdata.com/downloads/50m-physical-vectors/}, version 2.0.0.
}
\note{
The function plots a simple temporal mean of the loaded object in the form of
a map. It does not handle other temporal aggregations.
}
\examples{
# A grid
data(iberia_ncep_ta850)
plotMeanGrid(iberia_ncep_ta850)
# A multigrid
data(iberia_ncep_hus850)
mf <- makeMultiGrid(iberia_ncep_ta850, iberia_ncep_hus850)
plotMeanGrid(mf)
# A multimember grid
data(tasmax_forecast)
plotMeanGrid(tasmax_forecast) # multimember mean
plotMeanGrid(tasmax_forecast, multi.member = TRUE) # by members
# A multimember multigrid
data(tasmin_forecast)
mm.mf <- makeMultiGrid(tasmax_forecast, tasmin_forecast)
plotMeanGrid(mm.mf) # Note: multi-member not supported in this case

}
\author{
J Bedia
}
\seealso{
Other visualization: \code{\link{quickDiagnostics}},
  \code{\link{tercileValidation}}
}

