% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppModelSetup.R
\name{ppModelSetup}
\alias{ppModelSetup}
\title{Data preparation for perfect-prog model construction}
\usage{
ppModelSetup(y, x, newdata)
}
\arguments{
\item{y}{A grid or station data containing the observed climate data for the training period}

\item{x}{A grid containing the simulated climate by the model for the training period. This can be either
the same variable as \code{obs}, in the case of model calibration (bias correction and related techniques) or MOS (model
 output statistics) approaches, or a set of predictors in the case of \emph{perfect prog} downscaling approaches (possibly
 after principal component analysis via the \code{\link{prinComp}} output).}

\item{newdata}{A grid containing the simulated climate for the variables used in \code{x}, but considering the test period.
If \code{x} is a multimember grid (the member dimension is > 1), the function assumes that the order of 
the members is the same in \code{newdata}.}
}
\value{
A list with differet components required by the downscaling functions:
\itemize{
\item \code{stations} Logical indicating whether the predictand comes from station data
\item \code{multi.member} Logical indicating whether newdata is multi-member or not
\item \code{pred.mat} 2D matrix of predictors. Predictors are arranged in columns and time in rows
\item \code{sim.mat} A list of 2D matrices containing the prediction data. The list is of length \emph{n},
being \emph{n} the number of members considered (n = 1 for deterministic/single member predictions).
\item \code{sim.dates} The Dates element of the newdata (either a list of start/end lists for
several predictors or a list of two with the estart/end of a single predictor. See \code{\link{dateReplacement}}
for details).
\item \code{init.dates} Initialization dates inherited from 'newdata' if a forecast. NULL otherwise.
\item \code{member.names} Names of the members inherited from 'newdata' if a forecast. NULL otherwise.
}
}
\description{
Performs the various data pre-processing and formatting steps required
to implement the different perfect-prog downscaling methods.
}
\details{
The function accepts either PCA or raw grids as predictors. In the first case, it handles the 
mapping of EOFs onto the newdata grids.
}
\author{
J. Bedia
}
\keyword{internal}
