% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biasCorrection.R
\name{pqm}
\alias{pqm}
\title{Parametric Quantile Mapping method for bias correction}
\usage{
pqm(o, p, s, fitdistr.args, precip, pr.threshold)
}
\arguments{
\item{o}{A vector (e.g. station data) containing the observed climate data for the training period}

\item{p}{A vector containing the simulated climate by the model for the training period.}

\item{s}{A vector containing the simulated climate for the variable used in \code{x}, but considering the test period.}

\item{fitdistr.args}{Further arguments passed to function \code{\link[MASS]{fitdistr}} 
(\code{densfun}, \code{start}, \code{...}). Only used when applying the "pqm" method 
(parametric quantile mapping). Please, read the \code{\link[MASS]{fitdistr}} help 
document  carefully before parameter setting in \code{fitdistr.args}.}

\item{precip}{Logical indicating if o, p, s is precipitation data.}

\item{pr.threshold}{The minimum value that is considered as a non-zero precipitation. Ignored for
\code{varcode} values different from \code{"pr"}. Default to 1 (assuming mm).}
}
\description{
Implementation of Parametric Quantile Mapping method for bias correction
}
\author{
S. Herrera and M. Iturbide
}
\keyword{internal}
