% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.R
\name{predictor.nn.values}
\alias{predictor.nn.values}
\title{Construct local predictor matrices}
\usage{
predictor.nn.values(nn.indices.list, grid, fun = NULL)
}
\arguments{
\item{nn.indices.list}{A list of index positions as returned by \code{\link{predictor.nn.indices}}}

\item{grid}{The grid containing the local predictor information. It can be either the predictors (training)
grid or the simulation (prediction) grid, depending on the situation.}

\item{fun}{A list of funtions for neighbour aggregation for each local predictor variable.
Default to \code{NULL}, and no aggregation is performed}
}
\value{
A nested list of 2D matrices with the following structure: sites/members
}
\description{
Constructs the local predictor matrices given their spatial index position 
(as returned by \code{\link{predictor.nn.indices}}) and additional parameters.
}
\seealso{
Other downscaling.helpers: \code{\link{predictor.nn.indices}},
  \code{\link{prepareData}}, \code{\link{prepareNewData}}
}
\author{
J Bedia
}
\concept{downscaling.helpers}
\keyword{internal}
