% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_newdata.R
\name{prepare_newdata}
\alias{prepare_newdata}
\title{Prepare newdata for predictions}
\usage{
prepare_newdata(newdata, predictor)
}
\arguments{
\item{newdata}{A grid containing the prediction data.}

\item{predictor}{A predictor structure, as returned by \code{\link{prepare_predictors}}}
}
\value{
A named list with the components required by the downscaling method in order to perform the predictions
}
\description{
Prepare the prediction data according to the definition of the experiment
}
\examples{
# See the dedicated vignette by typing:
# utils::vignette(topic = "configuring_newdata", package = "downscaleR")
}
\seealso{
Other downscaling.helpers: \code{\link{predictor.nn.indices}},
  \code{\link{predictor.nn.values}},
  \code{\link{prepare_predictors}}
}
\author{
J Bedia
}
