% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.R
\name{prepareData}
\alias{prepareData}
\title{Configuration of data for downscaling}
\usage{
prepareData(x, y, global.vars = NULL, combined.only = TRUE,
  spatial.predictors = NULL, local.predictors = NULL,
  extended.predictors = NULL)
}
\arguments{
\item{x}{A grid (usually a multigrid) of predictor fields}

\item{y}{A grid (usually a stations grid, but not necessarily) of observations (predictands)}

\item{global.vars}{An optional character vector with the short names of the variables of the input \code{x} 
multigrid to be retained as global predictors (use the \code{\link{getVarNames}} helper if not sure about variable names).
This argument just produces a call to \code{\link[transformeR]{subsetGrid}}, but it is included here for better
flexibility in downscaling experiments (predictor screening...). For instance, it allows to use some 
specific variables contained in \code{x} as local predictors and the remaining ones, specified in \code{subset.vars},
as either raw global predictors or to construct the combined PC.}

\item{combined.only}{Optional, and only used if spatial.predictors parameters are passed. Should the combined PC be used as the only
global predictor? Default to TRUE. Otherwise, the combined PC constructed with \code{which.combine} argument in 
\code{\link{prinComp}} is append to the PCs of the remaining variables within the grid.}

\item{spatial.predictors}{Default to \code{NULL}, and not used. Otherwise, a named list of arguments in the form \code{argument = value},
with the arguments to be passed to \code{\link[transformeR]{prinComp}} to perform Principal Component Analysis
of the predictors grid (\code{x}). See Details on principal component analysis of predictors.}

\item{local.predictors}{Default to \code{NULL}, and not used. Otherwise, a named list of arguments in the form \code{argument = value},
with the following arguments:
\itemize{
  \item \code{vars}: names of the variables in \code{x} to be used as local predictors
  \item \code{fun}: Optional. Aggregation function for the selected local neighbours.
  The aggregation function is specified as a list, indicating the name of the aggregation function in
   first place (as character), and other optional arguments to be passed to the aggregation function.
   For instance, to compute the average skipping missing values: \code{fun = list(FUN= "mean", na.rm = TRUE)}.
   Default to NULL, meaning that no aggregation is performed.
  \item \code{n}: Integer. Number of nearest neighbours to use. If a single value is introduced, and there is more
  than one variable in \code{vars}, the same value is used for all variables. Otherwise, this should be a vector of the same
  length as \code{vars} to indicate a different number of nearest neighbours for different variables.
}}

\item{extended.predictors}{This is a parameter related to the extreme learning machine and reservoir computing framework where input data is randomly projected into a new space of size \code{n}. Default to \code{NULL}, and not used. Otherwise, a named list of arguments in the form \code{argument = value},
with the following arguments:
\itemize{
  \item \code{n}: A numeric value. Indicates the size of the random nonlinear dimension where the input data is projected.
  \item \code{module}: A numeric value (Optional). Indicates the size of the mask's module. Belongs to a specific type of ELM called RF-ELM.
}}
}
\value{
A named list with components \code{y} (the predictand), \code{x.global} (global predictors, 2D matrix), \code{x.local} (local predictors, a list) 
and \code{pca} (\code{\link[transformeR]{prinComp}} output), and other attributes. See Examples.
}
\description{
Configuration of data for flexible downscaling experiment definition
}
\details{
\strong{Temporal consistency}
 Note that \code{x} (predictors) and \code{y} predictands are checked for temporal consistency
  prior to downscaling. In case of partial temporal overlapping, both are internnaly intersected for exact temporal matching.
 
 \strong{Principal Component Analysis}
 Always that spatial.predictors is used, a combined PC will be returned (unless one single predictor is used, case in which no combination is possible).
 Note that the variables of the predictor grid used to construct the combined PC can be flexibly controlled through the optional argument
 \code{subset.vars}.
}
\examples{
# Loading data
data("VALUE_Iberia_tas")
y <- VALUE_Iberia_tas 
data("NCEP_Iberia_hus850", "NCEP_Iberia_psl", "NCEP_Iberia_ta850")
x <- makeMultiGrid(NCEP_Iberia_hus850, NCEP_Iberia_psl, NCEP_Iberia_ta850)
# Raw data
data <- prepareData(x = x, y = y)
# Using PCs as predictors. Number of EOFS: 10,5,5 for the 3 input variables
data <- prepareData(x = x, y = y, spatial.predictors = list(n.eofs = c(10,5,5)))
# Using joined PCs as predictors. Explained variance 95\%
data <- prepareData(x = x, y = y, 
spatial.predictors = list(v.exp = 0.95, which.combine =getVarNames(x)))
# Using local predictors: the 4 closest gridboxes
data <- prepareData(x = x, y = y,local.predictors = list(n=4, vars = getVarNames(x)))
# Using joined PCs and local predictors: the 4 closest gridboxes
data <- prepareData(x = x, y = y,local.predictors = list(n=4, vars = getVarNames(x)),
spatial.predictors = list(v.exp = 0.95, which.combine =getVarNames(x)))
}
\seealso{
\href{https://github.com/SantanderMetGroup/downscaleR/wiki/preparing-predictor-data}{downscaleR Wiki} for preparing predictors for downscaling and seasonal forecasting.

Other downscaling.helpers: \code{\link{predictor.nn.indices}},
  \code{\link{predictor.nn.values}},
  \code{\link{prepareNewData}}
}
\author{
J. Bedia, D. San-Martín and J.M. Gutiérrez
}
\concept{downscaling.helpers}
