% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareNewData.R
\name{prepareNewData}
\alias{prepareNewData}
\title{Prepare newdata for predictions}
\usage{
prepareNewData(newdata, data.structure)
}
\arguments{
\item{newdata}{A grid containing the prediction data.}

\item{data.structure}{A structure, as returned by \code{\link{prepareData}}}
}
\value{
A named list with the components required by the downscaling method in order to perform the predictions
}
\description{
Prepare the prediction data according to the definition of the experiment
}
\examples{
# Loading data
data("VALUE_Iberia_tas")
y <- VALUE_Iberia_tas 
data("NCEP_Iberia_hus850", "NCEP_Iberia_psl", "NCEP_Iberia_ta850")
x <- makeMultiGrid(NCEP_Iberia_hus850, NCEP_Iberia_psl, NCEP_Iberia_ta850)
# Calculating EOFs
data <- prepareData(x = x, y = y, spatial.predictors = list(v.exp = 0.95))
# Projecting a new dataset to the calculated EOFs
newdata <- prepareNewData(x,data)
}
\seealso{
\href{https://github.com/SantanderMetGroup/downscaleR/wiki/preparing-predictor-data}{downscaleR Wiki} for preparing predictors for downscaling and seasonal forecasting.

Other downscaling.helpers: \code{\link{predictor.nn.indices}},
  \code{\link{predictor.nn.values}},
  \code{\link{prepareData}}
}
\author{
J Bedia
}
\concept{downscaling.helpers}
