% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prinComp.R
\name{prinComp}
\alias{prinComp}
\title{Principal Component Analysis of grid/multigrid/multimember data}
\usage{
prinComp(gridData, n.eofs = NULL, v.exp = NULL, scaling = c("field",
  "gridbox"), parallel = FALSE, max.ncores = 16, ncores = NULL)
}
\arguments{
\item{gridData}{A grid, multigrid, multimember grid or multimember multigrid object}

\item{n.eofs}{Number of EOFs to be retained. Default to \code{NULL}, indicating
that either all EOFs are kept, or that the next argument will be used as criterion
for its determination. See next argument and details.}

\item{v.exp}{Maximum fraction of explained variance, in the range (0,1]. Used to determine the number of EOFs 
to be retained, as an alternative to \code{n.eofs}. Default to \code{NULL}. See details.}

\item{scaling}{Method for performing the scaling (and centering) of the input raw data matrix.
Currently accepted choices are \code{"field"} (the default) and \code{"gridbox"}. See details.}

\item{parallel}{Logical. Should parallel execution be used?}

\item{max.ncores}{Integer. Upper bound for user-defined number of cores.}

\item{ncores}{Integer number of cores used in parallel computation. Self-selected number of
cores is used when \code{ncpus = NULL} (the default), or when \code{maxcores} exceeds the default \code{ncores} value.}
}
\value{
A list of \emph{N + 1} elements for multigrids, where \emph{N} is the number of input variables used
 and the last element contains the results of the combined PCA (See details). The list is named as the variables,
 including the last element named \code{"COMBINED"}. In case of single grids (1 variable only), a list of length 1
  (without the combined element). For each element of the list, the following objects are returned, either in the form of
  another list (1 element for each member) for multimembers, or not in the case of non multimember inputs:
 \itemize{
 \item \code{PCs}: A matrix of principal components, arranged in columns by decreasing importance order 
 \item \code{EOFs}: A matrix of EOFs, arranged in columns by decreasing importance order
 \item \code{orig}: The original variable in the form of a 2D-matrix. This is standardized according
 to the approach indicated in argument \code{scaling}.
 }
 The \dQuote{order of importance} is given by the explained variance of each PC, as indicated
 in the attribute \code{"explained_variance"} as a cumulative vector.
 Additional information is returned via the remaining attributes (see details), including geo-referencing and time.
}
\description{
Performs a Principal Component Analysis of grid, multigrid or multi-member data
}
\details{
\strong{Number of EOFs}
\code{n.eofs} and \code{v.exp} are alternative choices for the determination
 of the number of EOFs (hence also the corresponding PCs) to be retained. If both are \code{NULL} (the default)
 , all EOFs will be retained. If both are given a value different from \code{NULL}, 
 the \code{n.eofs} argument will prevail, and \code{v.exp} will be ignored, with a warning.
 Note that when dealing with multigrids, the \code{n.eofs} argument will return the same number of EOFs
 for each variable by sepparate and also for the combined grid, while \code{v.exp} will select
 a different number, depending on the explained variance in each case. 
 
\strong{Scaling and centering}

In order to eliminate the effect of the varying scales of the different climate variables, the input
data matrix is always scaled and centered, and there is no choice to avoid this step. However, the mean
and standard deviation can be either computed for each grid box individually (\code{"gridbox"}) or for all
grid-boxes together (i.e., at the field level, \code{"field"}). The last case is the preferred choice and has
been set as default, returning one single mean and sigma parameter for each variable. If the \code{"gridbox"}
approach is selected, a vector of length \emph{n}, where \emph{n} is the number of grid-cells composing the 
grid, is returned for both the mean and sigma parameters (this is equivalent to using the \code{\link{scale}}
function with the input data matrix). 

The method used is returned as a global attribute of the returned object (\code{"scaled:method"}), and the 
\emph{mu} and \emph{sigma} parameters are returned as attributes of the corresponding variables
 (\code{"scaled:scale"} and \code{"scaled:center"} respectively).

\strong{Combined EOF analysis}

When dealing with multigrid data, apart from the PCA analysis performed on each variable indivisually,
a combined analysis considering all variables together is done. This is always returned in the last element
of the output list.
}
\note{
Performing PCA analysis on multimember multigrids may become time-consuming and computationally expensive. 
It is therefore advisable to avoid the use of this option for large datasets, and iterate over single
multimember grids instead.
}
\section{Parallel Processing}{


Parallel processing is enabled using the \pkg{parallel} package. 
Parallelization is undertaken by a FORK-type parallel socket cluster formed by \code{ncores}.
If \code{ncores} is not specified (default), \code{ncores} will be one less than the autodetected number of cores.
The maximum number of cores used for parallel processing can be set with the \code{max.ncores} argument, 
although this will be reset to the auto-detected number of cores minus 1 if this number is exceeded. Note that not all 
code, but just some critical loops within the function are parallelized.

In practice, parallelization does not always result in smaller execution times, due to the parallel overhead.
However, parallel computing may potentially provide a significant speedup for the 
particular case of large multimember datasets or large grids.
 
Parallel computing is currently not available for Windows machines.
}
\examples{
\dontrun{
# Download dataset
download.file("http://meteo.unican.es/work/downscaler/data/Iberia_NCEP.tar.gz", 
destfile = "mydirectory/Iberia_NCEP.tar.gz")
# Extract files from the tar.gz file
untar("mydirectory/NCEP_Iberia.tar.gz", exdir = "mydirectory")
# First, the path to the ncml file is defined:
ncep <- "mydirectory/Iberia_NCEP/Iberia_NCEP.ncml"
# A multigrid containing a set of variables is loaded (e.g. data for spring spanning the 
# 30-year period 1981--2010):
require(loadeR)
hus <- loadGridData(ncep, var = "hus@850", season = 3:5, years = 1981:2010)
ta <- loadGridData(ncep, var = "ta@850", season = 3:5, years = 1981:2010)
psl <- loadGridData(ncep, var = "psl", season = 3:5, years = 1981:2010)
multigrid <- makeMultiGrid(hus, ta, psl)
# In this example, we retain the PCs explaining the 99\\\% of the variance
pca <- prinComp(multigrid, v.exp = .99)
# Note that, apart from computing the principal components and EOFs for each grid, 
# it also returns, in the last element of the output list,
# the results of a PC analysis of all the variables combined (named "COMBINED"):
names(pca)
str(pca)
# The different attributes of the pca object provide information regarding the variables involved
# and the geo-referencing information
str(attributes(pca))
# In addition, for each variable (and their combination), the scaling and centering parameters 
# are also returned. These are either a single value in case of grid scaling (the default), or a
# vector of values, one for each grid-cell, for the gridbox scaling. For instance, the parameters
# for the specific humidity grid are:
attributes(pca$hus[[1]])$`scaled:center`
attributes(pca$hus[[1]])$`scaled:scale`
# In addition, the (cumulative) explained variance of each PC is also returned:
vexp <- attributes(pca$hus[[1]])$explained_variance
# The classical "scree plot":
barplot(1-vexp, names.arg = paste("PC",1:length(vexp)), las = 2, 
        ylab = "Fraction of unexplained variance")
 
# This is an example using a multimember object:
data(tasmax_forecast)
# In order make the computation faster, we interpolate to a coarser grid of 2.5 deg
range(tasmax_forecast$xyCoords$x)
range(tasmax_forecast$xyCoords$y)
multimember <- interpGrid(tasmax_forecast, 
                              new.coordinates = list(c(-10,29.5, 2.5), c(35.5, 64.5, 2.5)))
# In this case we retain the first 15 EOFs:
pca.mm <- prinComp(multimember, n.eofs = 15) 
# Note that now the results of the PCA for the variable are a list, with the results 
# for each member sepparately considered
str(pca.mm)

# The most complex situation comes from multimember multigrids:
data(tasmin_forecast)
# We interpolate using an identical grid than the previous example:
multimember2 <- interpGrid(tasmin_forecast, new.coordinates = getGrid(multimember))
# Now the multimember multigrid is constructed
mm.multigrid <- makeMultiGrid(multimember, multimember2)
pca.mm.mf <- prinComp(mm.multigrid, n.eofs = 10)
# Now there is a "COMBINED" element at the end of the output list
str(pca.mm.mf)
}
}
\author{
J. Bedia
}
\references{
Guti\'{e}rrez, J.M., R. Ancell, A. S. Cofi\~{n}o and C. Sordo (2004). Redes Probabil\'{i}sticas
 y Neuronales en las Ciencias Atmosf\'{e}ricas. MIMAM, Spain. 279 pp.
  \url{http://www.meteo.unican.es/en/books/dataMiningMeteo}
}
\seealso{
\code{\link{gridFromPCs}}, \code{\link{plotEOF}}
}

