% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biasCorrection.R
\name{ptr}
\alias{ptr}
\title{Power transformation of precipitation}
\usage{
ptr(o, p, s, precip)
}
\arguments{
\item{o}{A vector (e.g. station data) containing the observed climate data for the training period}

\item{p}{A vector containing the simulated climate by the model for the training period.}

\item{s}{A vector containing the simulated climate for the variable used in \code{p}, but considering the test period.}

\item{precip}{Logical indicating if o, p, s is precipitation data.}
}
\description{
Implementation of Power transformation of precipitation method for bias correction
}
\author{
S. Herrera and B. Szabo-Takacs
}

