% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quickDiagnostics.R
\name{quickDiagnostics}
\alias{quickDiagnostics}
\title{Plots daily/annual series and the annual correlation map of different field objects}
\usage{
quickDiagnostics(obs, sim, downscaled = NULL, location = c(-42.5, -3),
  type = c("daily", "interannual"), yrange = NULL)
}
\arguments{
\item{obs}{Field of observations.}

\item{sim}{Field of model data.}

\item{downscaled}{Field of downscaling output.}

\item{location}{Coordinates of a location in the geographic domain of the field.}

\item{yrange}{equivalent to ylim argument for plot function}
}
\value{
A plot
}
\description{
Plots daily/annual series and the annual correlation map of different field objects
}
\author{
M. Iturbide \email{maibide@gmail.com}
}

