% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isimip3.R
\name{re_scale_by_upper_bound_climatology}
\alias{re_scale_by_upper_bound_climatology}
\title{Re-Scaling by upper bound climatology}
\usage{
re_scale_by_upper_bound_climatology(data, dates, rescale)
}
\arguments{
\item{data}{A vector containing the climate data}

\item{dates}{A vector containing the dates to which corresponds the data.}

\item{rescale}{A list with the elements needed (day-of-year and scale corresponding to each day-of-year) to re-scale the bias adjusted data.}
}
\description{
Re-Scaling by upper bound climatology
}
\author{
S. Herrera (sixto.herrera@unican.es)
}
\keyword{internal}
