% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biasCorrection.R
\name{recoverMemberDim}
\alias{recoverMemberDim}
\title{Recover member multimember structure}
\usage{
recoverMemberDim(plain.grid, bc.grid, newdata)
}
\arguments{
\item{plain.grid}{A \dQuote{flattened} grid used as predictor in \code{biasCorrection} (the 'pred' object)}

\item{bc.grid}{The bias-corrected output (the 'bc' object), still without its member structure}

\item{newdata}{The 'newdata' object, needed to recover relevant metadata (i.e. initialization dates and member names)}
}
\value{
A (bias-corrected) multimember grid
}
\description{
Recover member multimember structure after application of \code{\link{flatMemberDim}}
}
\seealso{
\code{\link{flatMemberDim}}, for \dQuote{flattening} the member structure
}
\author{
J Bedia
}
\keyword{internal}
