% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{renameDims}
\alias{renameDims}
\title{Set the 'dimensions' attribute}
\usage{
renameDims(obs, multi.member)
}
\arguments{
\item{obs}{A observations object}

\item{multi.member}{Logical indicating if simulation data is a multimember}
}
\value{
A character vector indicating the dimensions of the output object
}
\description{
Sets the 'dimensions' attribute of model out Data objects after downscaling
}
\author{
J. Bedia
}
\keyword{internal}

