% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescaleGrid.R
\name{rescaleGrid}
\alias{rescaleGrid}
\title{Grid rescaling}
\usage{
rescaleGrid(grid, grid.clim = NULL, ref = NULL, by.member = TRUE,
  parallel = FALSE, max.ncores = 16, ncores = NULL)
}
\arguments{
\item{grid}{Input grid to be rescaled}

\item{grid.clim}{Reference climatology to be subtracted to the input grid. If \code{NULL} (the default),
the climatology is directly computed from the grid via \code{\link{climatology}}, either member-by-member
or using the ensemble mean climatology, as specified by the \code{by.member} flag.}

\item{ref}{Reference grid. After subtracting to grid its climatology (as defined by \code{grid.clim}), the mean
climatology of this additional grid is added. Default to NULL, so no reference grid is used.}

\item{by.member}{Logical. In case of multimember grids, should the climatology be computed sepparately
for each member (\code{by.member=TRUE}), or a single climatology calculated from the ensemble mean
 (\code{by.member=FALSE})?. Default to \code{TRUE}. Argument passed to \code{\link{climatology}}.}

\item{parallel}{Logical. Should parallel execution be used?}

\item{max.ncores}{Integer. Upper bound for user-defined number of cores.}

\item{ncores}{Integer number of cores used in parallel computation. Self-selected number of
cores is used when \code{ncpus = NULL} (the default), or when \code{maxcores} exceeds the default \code{ncores} value.}
}
\value{
A rescaled grid
}
\description{
Rescale a grid
}
\details{
The reference grid (\code{ref}) is used to correct the input grid, as follows:

\deqn{grid' = grid - mu_grid + mu_ref}

, where \emph{mu} corresponds to the monthly climatological mean considering the training period,
and \emph{grid'} is the corrected test data. The way \emph{mu_ref} and \emph{mu_grid} are computed in case
of multimember grids is controlled by the argument \code{by.member}.

The \code{ref} usually corresponds to the control run of the GCM in the training period in climate change applications,
or the hindcast data for the training period in s2d applications. Note that by default \code{ref = NULL}. In this 
case it will be assumed to be the \code{pred} grid. This can be used for instance when train and test correspond
to the same model.
}

