% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescaleMonthlyMeans.R
\name{rescaleMonthlyMeans}
\alias{rescaleMonthlyMeans}
\title{Annual cycle scaling of simulation data}
\usage{
rescaleMonthlyMeans(pred, sim, ref = NULL, ensemble = FALSE)
}
\arguments{
\item{pred}{Predictor field}

\item{sim}{Simulation field}

\item{ref}{Optional reference field. A field from where the scaling and centering parameters are taken. See details.}

\item{ensemble}{Logical flag. Should the correction of the mean be performed w.r.t. the ensemble mean (\code{TRUE})
or member by member independently (\code{ensemble = FALSE})?. Ignored if \code{ref} (and thus \code{sim}) are not
 multimembers. Default to \code{FALSE}.}
}
\value{
A field with the rescaled simulation data (\code{sim}), with the centering parameters
 indicated as an attribute
}
\description{
Annual cycle scaling of simulation data w.r.t. the predictors
}
\details{
The reference field is used to correct the simulation (test) data, as follows:

\deqn{sim' = sim - mu_ref + mu_pred}

, where \emph{mu} corresponds to the monthly climatological mean considering the training period,
and \emph{sim'} is the corrected simulation (test) data. The way \emph{mu_ref} is computed in case
of multimember fields is controlled by the argument \code{ensemble}.

The \code{ref} usually corresponds to the control run of the GCM in the training period in climate change applications,
or the hindcast data for the training period in s2d applications. Note that by default \code{ref = NULL}. In this 
case it will be assumed to be the \code{pred} field. This can be used for instance when train and test correspond
to the same model.
}
\author{
J. Bedia \email{joaquin.bedia@gmail.com}
}
\keyword{internal}

