% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/revArrayLatDim.R
\name{revArrayLatDim}
\alias{revArrayLatDim}
\title{Reverse latitude coordinates ordering}
\usage{
revArrayLatDim(mdArray, grid)
}
\arguments{
\item{mdArray}{A n-dimensional array of data as returned by \code{makeSubset}}

\item{grid}{A java-class \sQuote{GeoGrid}}

\item{dimNamesRef}{Character vector with the (ordered) names of the dimensions of the array}
}
\value{
A n-dimensional array with the ordering of the lat dimension reversed
}
\description{
Reverses the order of the latitudinal coordinates when the dataset is read in
 reverse y order. Sub-routine of \code{makeSubset}.
}
\note{
The code is partially based on an example provided at \url{http://stackoverflow.com/a/14502298}
 by Hadley Wickham
}
\author{
J. Bedia \email{joaquin.bedia@gmail.com}
}
\references{
\url{http://adv-r.had.co.nz/Computing-on-the-language.html}
}
\keyword{internal}

