% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rdas.R
\docType{data}
\name{S4_Iberia_tas}
\alias{S4_Iberia_tas}
\title{grid containing the first 5 members of the System4 seasonal forecasting data of 15 members.
Contains mean temperature data for the Iberian Peninsula.}
\format{A grid}
\source{
subset of System4 seasonal forecasting data of 15 members accesible
through the \strong{ECOMS User Data Gateway (ECOMS-UDG)}
}
\description{
S4_Iberia_tas is a grid object returned by \code{loadECOMS} from package \pkg{loadeR.ECOMS}
}
\seealso{
\code{\link{makeMultiGrid}}, \code{\link[loadeR.ECOMS]{loadECOMS}}
}
\keyword{forecasting}
\keyword{seasonal}

