% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rdas.R
\docType{data}
\name{S4_Iberia_tp_fut}
\alias{S4_Iberia_tp_fut}
\title{grid containing the first 5 members of the System4 seasonal forecasting data of 15 members.
Contains precipitation data for the Iberian Peninsula.}
\format{A grid}
\source{
subset of System4 seasonal forecasting data of 15 members accesible
through the \strong{ECOMS User Data Gateway (ECOMS-UDG)}
}
\description{
S4_Iberia_tp_fut is a grid object returned by loadECOMS
function (package loadeR.ECOMS):
season = c(12,1,2),
years = 2001:2010,
var="tas",
lonLim = c(-10,5),
latLim= c(34,44),
time = "DD",
aggr.d = "min".
}
\seealso{
\code{\link{makeMultiField}}, \code{\link[loadeR.ECOMS]{loadECOMS}}
}
\keyword{forecasting}
\keyword{seasonal}

