% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isimip3.R
\name{scale_by_upper_bound_climatology}
\alias{scale_by_upper_bound_climatology}
\title{Scaling by upper bound climatology}
\usage{
scale_by_upper_bound_climatology(data, dates, halfwin = 0)
}
\arguments{
\item{data}{A vector containing the climate data}

\item{dates}{A vector containing the dates to which corresponds the data.}

\item{halfwin}{A number defining the length of running windows used in the calculations of 
climatologies of upper bounds that are used to scale values. The
window length is set to halfwin * 2 + 1 time
steps. If halfwin == 0 then no rescaling is done.}
}
\description{
Scaling by upper bound climatology
}
\author{
S. Herrera (sixto.herrera@unican.es)
}
\keyword{internal}
