% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biasCorrection.R
\name{scaling}
\alias{scaling}
\title{Scaling method for bias correction}
\usage{
scaling(o, p, s, scaling.type)
}
\arguments{
\item{o}{A vector (e.g. station data) containing the observed climate data for the training period}

\item{p}{A vector containing the simulated climate by the model for the training period.}

\item{s}{A vector containing the simulated climate for the variable used in \code{x}, but considering the test period.}

\item{scaling.type}{Character indicating the type of the scaling method. Options are \code{"additive"} (default)
or \code{"multiplicative"} (see details). This argument is ignored if \code{"scaling"} is not selected as the bias correction method.}
}
\description{
Implementation of Scaling method for bias correction
}
\author{
S. Herrera and M. Iturbide
}
\keyword{internal}
