% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanVarDimensions.R
\name{scanVarDimensions}
\alias{scanVarDimensions}
\title{Retrieve dimension information for a gridded variable}
\usage{
scanVarDimensions(grid)
}
\arguments{
\item{grid}{A java object of the class \sQuote{ucar.nc2.dt.grid.GeoGrid}}
}
\value{
a list of length \emph{N}, being N the number of dimensions defining the grid shape.
}
\description{
Goes through variable dimensions and retrieves information for the inventory.
This is a sub-routine of \code{dataInventory.NetCDF}, \code{makeSubset}.
}
\details{
All possible dimensions of a gridded datasets are searched and included if existing. This is done
 following the canonical dimension ordering [runtime, member, time, level, lat, lon], although function
 does not assume this ordering and therefore non-standard orderings should be adequately treated
 (this might be the case for instance when creating a new aggregation dimension via NcML).
 In case of existing vertical levels for a given dataset, all possible dataset level values are scanned
 one by one and those that do not exist for that particular variable are removed. This way, the inventory returns
 only the non-empty level values for each variable. In case of horizontal 2D axes,
 coordinate values are truncated to 1D for conciseness in the inventory.
}
\author{
J. Bedia \email{joaquin.bedia@gmail.com}
}
\references{
\url{http://www.unidata.ucar.edu/software/thredds/v4.3/netcdf-java/v4.3/javadocAll/ucar/nc2/dt/grid/GeoGrid.html}
}
\keyword{internal}

