% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataInventory.R
\name{stationInfo}
\alias{stationInfo}
\title{Retrieve station info}
\usage{
stationInfo(dataset, plot = TRUE)
}
\arguments{
\item{dataset}{A character string indicating the database to be accessed. For station data in standard ASCII format,
this is the path to the directory the dataset lives in.}

\item{plot}{Logical indicating if a map displaying the station locations should be displayed. Default to TRUE.}
}
\value{
A data.frame with the station codes in its first column, followed by longitude and latitude,
 and the rest of metadata (if any) in the following columns. If \code{plot = TRUE}, also a map of the station locations.
}
\description{
Get a quick overview of the stations contained in a stations dataset
}
\note{
For an adequate map display the station coordinates must be in decimal degrees.
}
\examples{
\donttest{
dir.create("mydirectory")
download.file("http://meteo.unican.es/work/downscaler/data/VALUE_ECA_86_v2.tar.gz", 
              destfile = "mydirectory/VALUE_ECA_86_v2.tar.gz")
# Extract files from the tar.gz file
untar("mydirectory/VALUE_ECA_86_v2.tar.gz", exdir = "mydirectory")
# Data inventory
value <- "mydirectory/VALUE_ECA_86_v2"
print(stationInfo(value))
} 

}
\author{
J. Bedia \email{joaquin.bedia@gmail.com}
}
\seealso{
\code{\link{dataInventory}} to obtain a more exhaustive report the dataset.
}

