% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetGrid.R
\name{subsetDimension}
\alias{subsetDimension}
\title{Select an arbitrary subset from a grid or multigrid along one of its dimensions}
\usage{
subsetDimension(grid, dimension = NULL, indices = NULL)
}
\arguments{
\item{grid}{The input grid to be subset. This is either a grid, as returned e.g. by \code{loadGridData} from package \pkg{loadeR} or a
multigrid, as returned by \code{makeMultiGrid}, or other types of multimember grids
(possibly multimember multigrids) as returned e.g. by \code{loadECOMS}, from package \pkg{loadeR.ECOMS}.}

\item{dimension}{Character vector indicating the dimension along which the positions indicated by the \code{indices} paraneter.}

\item{indices}{An integer vector indicating \strong{the positions} of the dimension to be extracted.}
}
\value{
A new grid object that is a logical subset of the input grid along the specified dimension.
}
\description{
Create a new grid/multigrid that is a subset of the input grid 
along the selected dimension
}
\details{
The attribute \code{subset} will be added taking the value of the \code{dimension} parameter.
}
\examples{
# Example - Member subset
data(tasmax_forecast)
plotMeanGrid(tasmax_forecast, TRUE)
# Selection of a smaller domain over the Iberian Peninsula and members 3 and 7
sub <- subsetDimension(tasmax_forecast,
                   dimension = "member",
                   indices = c(1,3))
plotMeanGrid(sub, multi.member = TRUE)
}
\author{
J. Bedia and S. Herrera
}
\seealso{
Other subsetting: \code{\link{subsetGrid}},
  \code{\link{subsetMembers}}, \code{\link{subsetSeason}},
  \code{\link{subsetSpatial}}, \code{\link{subsetVar}},
  \code{\link{subsetYears}}
}

