% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subsetField.R
\name{subsetField}
\alias{subsetField}
\title{Select an arbitrary spatial/temporal subset from a field or multifield}
\usage{
subsetField(field, lonLim = NULL, latLim = NULL, years = NULL)
}
\arguments{
\item{field}{The input data to be subset. This is either a field, as returned e.g. by \code{loadGridData}
or a Station dataset as returned by \code{loadStationData}.}

\item{lonLim}{Vector of length = 2, with minimum and maximum longitude coordinates, in decimal degrees,
of the bounding box defining the subset. For single-point subsets, a numeric value with the
longitude coordinate. If \code{NULL} (default), no subsetting is performed on the longitude dimension}

\item{latLim}{Same as \code{lonLim} argument, but for latitude.}

\item{years}{The years to be selected. Note that this can be either a continuous or discontinuous
series of years, the latter option often used in a cross-validation framework.
 See details for year-crossing seasons. Default to \code{NULL} (no subsetting is performed on the time dimension).}
}
\value{
The same type of object as entered in the \code{field} argument representing a subset of its input
}
\description{
Create a new field/multifield that is a subset of the input field
along the space and/or time dimensions
}
\details{
\strong{Time slicing}

Time slicing is performed on a yearly basis (i.e., sub-yearly specifications are not allowed). In case of
 year-crossing seasons (e.g. boreal winter (DJF), \code{season = c(12,1,2)}),
the season is assigned to the years of January and February
(i.e., winter of year 2000 corresponds to Dec 1999, Jan 2000 and Feb 2000). Thus,
the \code{years} argument must be introduced accordingly (See e.g. \code{\link{getYearsAsINDEX}}
function for details).

 \strong{Spatial slicing}

 Spatial subset definition is done via the \code{lonLim} and \code{latLim} arguments, in the same way as
  for instance the \code{\link{loadGridData}} function, with the exception that several chechks are undertaken
  to ensure that the subset is actually within the current extent of the input field. It is also possible to
  make single-point selections from a field, just by specifying a single coordinate instead of a range
   as the argument value. For instance \code{lonLim=c(-10,10)} and \code{latLim=c(35,45)} indicates a
 rectangular window centered in the Iberian Peninsula), and single grid-cell values
 (for instance \code{lonLim=-3.21} and \code{latLim=41.087} for retrieving the data in the closest grid
 point to the point coordinate -3.21E, 41.087N. In the last two cases, the function
 operates by finding the nearest (euclidean distance) grid-points to the coordinates introduced.
}
\author{
J. Bedia \email{joaquin.bedia@gmail.com}
}

