% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetGrid.R
\name{subsetGrid}
\alias{subsetGrid}
\title{Select an arbitrary subset from a grid or multigrid along one or more of its dimensions}
\usage{
subsetGrid(grid, var = NULL, members = NULL, years = NULL,
  season = NULL, latLim = NULL, lonLim = NULL)
}
\arguments{
\item{grid}{The input grid to be subset. This is either a grid, as returned e.g. by \code{loadeR::loadGridData}, a
multigrid, as returned by \code{makeMultiGrid}, or other types of multimember grids
(possibly multimember grids) as returned e.g. by \code{loadeR.ECOMS::loadECOMS}.}

\item{var}{Character vector indicating the variables(s) to be extracted. (Used for multigrid subsetting). See details.}

\item{members}{An integer vector indicating \strong{the position} of the members to be subset.}

\item{years}{The years to be selected. Note that this can be either a continuous or discontinuous
series of years, the latter option often used in a cross-validation framework.
 See details for year-crossing seasons. Default to \code{NULL} (no subsetting is performed on the time dimension).}

\item{season}{An integer vector indicating the months to be subset.}

\item{latLim}{Same as \code{lonLim} argument, but for latitude.}

\item{lonLim}{Vector of length = 2, with minimum and maximum longitude coordinates, in decimal degrees,
of the bounding box defining the subset. For single-point subsets, a numeric value with the
longitude coordinate. If \code{NULL} (default), no subsetting is performed on the longitude dimension}
}
\value{
A new grid object that is a logical subset of the input grid along the specified dimensions.
}
\description{
Create a new grid/multigrid that is a subset of the input grid along the selected dimensions
}
\details{
The attribute \code{subset} will be added to the different slots corresponding to the subset dimensions, taking
the value of the subroutine called in each case (e.g.: attribute subset will have the value \code{subsetSpatial}
in the xyCoords slot after spatial subsetting...).

\strong{Time slicing by years}

In case of year-crossing seasons (e.g. boreal winter (DJF), \code{season = c(12,1,2)}),
the season is assigned to the years of January and February 
(i.e., winter of year 2000 corresponds to Dec 1999, Jan 2000 and Feb 2000). Thus, 
the \code{years} argument must be introduced accordingly (See e.g. \code{\link{getYearsAsINDEX}}
function for details).

 \strong{Spatial slicing}
 
 Spatial subset definition is done via the \code{lonLim} and \code{latLim} arguments, in the same way as
  for instance the \code{loadGridData} function, from package \pkg{loadeR}, with the exception that several checks are undertaken
  to ensure that the subset is actually within the current extent of the input grid. It is also possible to
  make single-point selections from a grid, just by specifying a single coordinate instead of a range
   as the argument value. For instance \code{lonLim = c(-10,10)} and \code{latLim = c(35,45)} indicate a
 rectangular window centered in the Iberian Peninsula), and single grid-cell values
 (for instance \code{lonLim = -3.21} and \code{latLim = 41.087} for retrieving the data in the closest grid
 point to the point coordinate -3.21E, 41.087N. In the last two cases, the function
 operates by finding the nearest (euclidean distance) grid-points to the coordinates introduced.
 
 \strong{Extracting grids from multigrids}
 
 One or several variables from a multigrid object can be extracted. Note that argument \code{var} is 
 insensitive to the order of the variables, i.e.: variables will be always returned in the same order
  they are in the original multigrid.
}
\examples{
# Example 1 - Spatial / member subset
data(tasmax_forecast)
plotMeanGrid(tasmax_forecast, TRUE)
# Selection of a smaller domain over the Iberian Peninsula and members 3 and 7
sub <- subsetGrid(tasmax_forecast,
                   members = c(3,7),
                   lonLim = c(-10,5),
                   latLim = c(36,44))
plotMeanGrid(sub, multi.member = TRUE)
## Example 2 - Subsetting a multimember multigrid by variables
# Multimember multigrid creation
data(tasmax_forecast)
data(tasmin_forecast)
data(tp_forecast)
mm.mf <- makeMultiGrid(tasmax_forecast, tasmin_forecast, tp_forecast)
plotMeanGrid(mm.mf)
# Extracting just minimum temperature
sub1 <- subsetGrid(mm.mf, var = "tasmin", members = 1:4)
plotMeanGrid(sub1, multi.member = TRUE)
# Extracting precipitation and maximum temperature
# (Note that the grid variables are NOT re-ordered)
sub2 <- subsetGrid(mm.mf, var = c("tp", "tasmax"))
plotMeanGrid(sub2)
}
\author{
J. Bedia
}
\seealso{
Other subsetting: \code{\link{subsetDimension}},
  \code{\link{subsetMembers}}, \code{\link{subsetSeason}},
  \code{\link{subsetSpatial}}, \code{\link{subsetVar}},
  \code{\link{subsetYears}}
}

