% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetGrid.R
\name{subsetMembers}
\alias{subsetMembers}
\title{Member subsets from a multimember grid}
\usage{
subsetMembers(mmGrid, members = NULL)
}
\arguments{
\item{mmGrid}{Input multimember grid to be subset (possibly a multimember multigrid).}

\item{members}{An integer vector indicating \strong{the position} of the members to be subset.}
}
\value{
A grid (or multigrid) that is a logical subset of the input grid along its 'member' dimension.
}
\description{
Retrieves a grid that is a logical subset of a multimember grid along its 'member' dimension.
 Multimember multigrids are supported. Subroutine of \code{\link{subsetGrid}}.
}
\details{
An attribute 'subset' with value 'subsetMembers' is added to the Members slot of the output subset.
}
\author{
J. Bedia
}
\seealso{
Other subsetting: \code{\link{subsetDimension}},
  \code{\link{subsetGrid}}, \code{\link{subsetSeason}},
  \code{\link{subsetSpatial}}, \code{\link{subsetVar}},
  \code{\link{subsetYears}}
}
\keyword{internal}

