% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetGrid.R
\name{subsetRuntime}
\alias{subsetRuntime}
\title{Runtime subsets from a grid}
\usage{
subsetRuntime(mmGrid, runtime = NULL, drop = TRUE)
}
\arguments{
\item{mmGrid}{Input multiruntime grid to be subset.}

\item{runtime}{An integer vector indicating \strong{the position} of the runtimes to be subset.}

\item{drop}{Logical (default is TRUE). Drop or keep dimensions of length 1.}
}
\value{
A grid (or multigrid) that is a logical subset of the input grid along its 'runtime' dimension.
}
\description{
Retrieves a grid that is a logical subset of a multiruntime grid along its 'runtime' dimension.
 Multiruntime multigrids are supported. Subroutine of \code{\link{subsetGrid}}.
}
\details{
An attribute 'subset' with value 'subsetRuntime' is added to the Runtime slot of the output subset.
}
\author{
M. Iturbide
}
\seealso{
Other subsetting: \code{\link{subsetDimension}},
  \code{\link{subsetGrid}}, \code{\link{subsetMembers}},
  \code{\link{subsetSeason}}, \code{\link{subsetSpatial}},
  \code{\link{subsetVar}}, \code{\link{subsetYears}}
}
\keyword{internal}

