% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdas.R
\docType{data}
\name{tas_decadalForecast}
\alias{tas_decadalForecast}
\title{Decada forecast example grid}
\format{Grid data}
\source{
Subset of the SPECS_GFDL_decadal dataset. This data is accessible 
for loading in the \strong{The User Data Gateway (UDG)}, which is the one stop shop for 
climate data access maintained by the Santander MetGroup. 
The UDG builds on the THREDDS Access Portal (UDG-TAP) which is the entry point for 
authentication and data access (more info in \url{https://meteo.unican.es/trac/wiki/udg}).
}
\description{
Grid of decadal forecast data from the SPECS_GFDL_decadal dataset containing monthly temperature for the Iberian Peninsula.
}
\keyword{SPECS}
\keyword{decadal}
\keyword{forecast}
\keyword{temperature}

