% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdas.R
\name{tasmax_forecast}
\alias{tasmax_forecast}
\title{Multimember grid containing a seasonal forecast of maximum surface temperature for Europe}
\format{A multimember grid}
\source{
\url{http://www.meteo.unican.es/ecoms-udg}
}
\description{
CFSv2 forecast of maximum daily temperature for July 2001 over Europe. Lead-month 2, first 9 members.
}
\examples{
data(tasmax_forecast)
plotMeanGrid(tasmax_forecast, multi.member = TRUE)
}
\references{
Saha, S. \emph{et al.}, 2013. The NCEP Climate Forecast System Version 2. J Clim 130925135638001. doi:10.1175/JCLI-D-12-00823.1
}

