% Generated by roxygen2 (4.0.1): do not edit by hand
\name{tasmin_forecast}
\alias{tasmin_forecast}
\title{Multimember field containing a seasonal forecast of minimum surface temperature for Europe}
\format{A multimember field}
\source{
\url{http://www.meteo.unican.es/ecoms-udg}
}
\description{
CFSv2 forecast of minimum daily temperature for July 2001 over Europe. Lead-month 2, first 9 members.
}
\examples{
data(tasmin_forecast)
plotMeanField(tasmin_forecast, multi.member = TRUE)
}
\references{
Saha, S. \emph{et al.}, 2013. The NCEP Climate Forecast System Version 2. J Clim 130925135638001. doi:10.1175/JCLI-D-12-00823.1
}
\seealso{
\code{\link{makeMultiField}}, \code{\link[ecomsUDG.Raccess]{loadECOMS}}
}

