% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/conversion.R
\name{tdps2hurs}
\alias{tdps2hurs}
\title{Dew point to relative humidity}
\usage{
tdps2hurs(tas, tdps)
}
\arguments{
\item{tas}{vector of surface temperature (K)}

\item{tdps}{vector of surface dew point temperature (K)}
}
\value{
vector of relative humidity (\%)
}
\description{
Conversion of dew point to relative humidity
}
\author{
J Bedia \email{joaquin.bedia@gmail.com}, borrowing MatLab code from S. Herrera
}
\seealso{
Other conversion: \code{\link{hurs2huss}};
  \code{\link{mslp2ps}}
}
\keyword{internal}

