% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tercileBarplot.R
\name{tercileBarplot}
\alias{tercileBarplot}
\title{Tercile bar plot for visualization of the skill of an ensemble forecast prediction for a particular year.}
\usage{
tercileBarplot(mm.obj, obs, year.target, score.threshold = NULL,
  stationId = NULL)
}
\arguments{
\item{mm.obj}{A multi-member object with predictions, either a field or a multi-member station object as a result of
downscaling of a forecast using station data. See details.}

\item{obs}{The benchmarking observations for forecast verification}

\item{year.target}{Year selected to plot the probability of the tercile in bars}

\item{score.threshold}{Threshold to remark high positive score values}

\item{stationId}{In case of multi-member multi-station objects, one station can be selected to plot
 the diagram. Otherwise ignored.}
}
\description{
Tercile bar plot for visualization of the skill of an ensemble forecast prediction for a particular year.
}
\details{
For each member, the daily predictions are averaged to obtain a single seasonal forecast. For
rectangular spatial domains (i.e., for fields), the spatial average is first computed (with a warning) to obtain a
unique series for the whole domain. The corresponding terciles for each ensemble member are then computed
for the analysis period. Thus, each particular member and season, are categorized into three categories (above,
between or below), according to their respective climatological terciles. Then, a probabilistic forecast is computed
year by year by considering the number of members falling within each category. The probability for the selected year
is represented by the bars. The 1/3 probability is plotted by a grey line. For instance, probabilities below this
line are very low, indicating that a minority of the members falls in the tercile. Conversely, probabilities above 2/3
indicate a high level of member agreement (more than 66\% of members falling in the same tercile).

 Finally, the ROC Skill Score (ROCSS) is indicated at the bottom part of the bar plot for each tercile. It provides a
 quantitative measure of the forecast skill, and it is commonly used to evaluate the performance of probabilistic systems
 (Joliffe and Stephenson 2003). The value of this score ranges from 1 (perfect forecast system) to -1
 (perfectly bad forecast system). Zero indicates no skill compared with a random prediction. The negative values
 are written in red while high positive values are in blue. The threshold to highlight high positive values can be
 modified with the score.threshold argument.

 In case of multi-member fields, the field is spatially averaged to obtain one single time series
 for each member prior to data analysis, with a warning. In case of multimember stations, one single station
 can be selected through the \code{stationId} argument, otherwise all station series are also averaged.
}
\note{
The computation of climatological terciles requires a representative period to obtain meaningful results.
}
\author{
M.D. Frias, J. Fernandez and J. Bedia \email{joaquin.bedia@gmail.com}
}
\references{
Jolliffe, I. T. and Stephenson, D. B. 2003. Forecast Verification: A Practitioner's Guide in
 Atmospheri Science, Wiley, NY
}
\seealso{
Other visualization: \code{\link{bubblePlot}};
  \code{\link{spreadPlot}}; \code{\link{tercilePlot}}
}

