% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateGrid.R
\name{timeAggregation}
\alias{timeAggregation}
\title{Time aggregation}
\usage{
timeAggregation(grid, aggr.type = c("DD", "MM", "YY"), aggr.fun, parallel,
  max.ncores, ncores)
}
\arguments{
\item{grid}{A multimember grid to apply the aggregation}

\item{aggr.type}{Character string indicating the type of temporal aggregation: 
daily (\code{"DD"}), monthly (\code{"MM"}) or annual (\code{"YY"}).}

\item{aggr.fun}{One of \code{aggr.d}, \code{aggr.m} or \code{aggr.y} arguments, as passed by \code{aggregateGrid}}

\item{parallel.pars}{Arguments defining the parallelization options, as passed by \code{\link{parallelCheck}}}
}
\description{
Aggregate a grid along its time dimension
}
\author{
J. Bedia, M. Iturbide, M. de Felice
}
\keyword{internal}

