% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/timeBounds.R
\name{timeBounds}
\alias{timeBounds}
\title{Calculate time bounds of a verification time}
\usage{
timeBounds(dic, foreDates)
}
\description{
Internal function to compute the time bounds of each verification time
}
\author{
J. Bedia \email{joaquin.bedia@gmail.com}
}
\keyword{internal}

