% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rdas.R
\docType{data}
\name{VALUE_Iberia_tp}
\alias{VALUE_Iberia_tp}
\title{Station daily precipitation dataset in Iberia}
\format{Station data}
\source{
Subset of VALUE station data. Full dataset is accessible
for download in \url{http://meteo.unican.es/work/downscaler/data/VALUE_ECA_86_v2.tar.gz}.
}
\description{
Station data from the VALUE_ECA_86_v2 dataset containing daily precipitation for stations in the Iberian Peninsula.
}

