% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biasCorrection.R
\name{varCoeficient}
\alias{varCoeficient}
\title{VarCoeficient}
\usage{
varCoeficient(delta, data, cv)
}
\arguments{
\item{delta}{A vector of power parameter}

\item{data}{A vector containing the simulated climate by the model for training period}

\item{cv}{A vector containing coefficient of variation of observed climate data}
}
\description{
preprocess to power transformation of precipitation
}
\author{
S. Herrera and B. Szabo-Takacs
}

