% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{which.leap}
\alias{which.leap}
\title{Identification of leap years
Identification of leap years}
\usage{
which.leap(years)
}
\arguments{
\item{years}{a integer vector of (gregorian) years}
}
\value{
a vector of indices of the position of leap years
}
\description{
Identification of leap years
Identification of leap years
}
\examples{
leap.years <- which.leap(1885:1937)
(1885:1937)[leap.years]
}
\author{
J. Bedia
}
\references{
\url{https://en.wikipedia.org/wiki/Leap_year}
}
\keyword{internal}

