% Generated by roxygen2 (4.0.1): do not edit by hand
\name{analogs}
\alias{analogs}
\title{Analog downscaling}
\usage{
analogs(obs, pred, sim, n.neigh = 1, sel.fun = c("random", "mean"))
}
\arguments{
\item{obs}{A field or station data containing the observed climate data for the training period}

\item{pred}{A field containing the simulated climate by the model for the training period. This can be either
the same variable as \code{obs}, in the case of model calibration (bias correction and related techniques) or MOS (model
 output statistics) approaches, or a set of predictors in the case of \emph{perfect prog} downscaling approaches (possibly
 after principal component analysis via the \code{\link{prinComp}} output).}

\item{sim}{A field containing the simulated climate for the variables used in \code{pred}, but considering the test period.}

\item{n.neigh}{Integer indicating the number of closest neigbours to retain for analog construction. Default to 1.}

\item{sel.fun}{Criterion for the construction of analogs when several neigbours are chosen. Ignored when \code{n.neig = 1}.
Current values are \code{"random"} (the default) and \code{"mean"}. See details.}
}
\description{
Implementation of the downscaling analogs method
}
\details{
\strong{Spatial consistency}

Several checks of spatial consistency are performed. In particular, note that both 'pred' (reanalysis) and 'sim' (model
simulations) should be in the same grid. This consistency must be ensured by the user prior to entering these arguments,
for instance by means of the \code{\link{interpGridData}} function in conjunction with the \code{\link{getGrid}} method.

\strong{Scaling and centering}

When the climate variables are used as predictors instead of the PCs, these are previously centered and scaled
using the mean and sigma parameters globally computed for the whole spatial domain (This is equivalent to the \dQuote{field})
method in the \code{\link{prinComp}} function. The simulation data will use the parameters obtained when scaling and centering
the predictors dataset. In case that the predictors come from a PC analysis object (as returned by \code{\link{prinComp}}), the
parameters for rescaling the simulation data are passed by the predictors.

\strong{Construction of analogs using multiple neighbours}

The argument \code{sel.fun} controls how the analogs are constructed when considering more than the first neighbour (argument
\code{n.neigh} > 1). In this case the \code{"random"} choice randomly selects one of the \code{n.neigh} neighbours,
 while the \code{"mean"} choice will compute their average.
}
\author{
J. Bedia \email{joaquin.bedia@gmail.com}
}
\references{
Benestad, R.E., Hanssen-Bauer, I. and Chen, D., 2008. Empirical-Statistical Downscaling,
 1st ed. World Scientific Publishing, Singapore

Gutierrez, J.M. \emph{et al.}, 2013. Reassessing Statistical downscaling techniques for
 their robust application under climate change conditions. J. Clim. 26, 171-188

Bedia, J. \emph{et al.}, 2013. Robust projections of Fire Weather Index in the Mediterranean
 using statistical downscaling. Clim. Change 120, 229-247.
}
\seealso{
\code{\link{prinComp}} for details on principal component/EOF analysis
\code{\link{loadMultiField}}, \code{\link{makeMultiField}} for multifield creation
\code{\link{loadGridData}} and \code{\link{loadStationData}} for loading fields and station data respectively.

Other downscaling: \code{\link{biasCorrection}};
  \code{\link{calibrateProj}}; \code{\link{isimip}}
}

