% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsANALOGS.R
\name{analogs.train}
\alias{analogs.train}
\title{Analogs}
\usage{
analogs.train(
  x,
  y,
  dates,
  n.analogs = 4,
  sel.fun = "mean",
  window = 7,
  n.random = NULL,
  pool = 0
)
}
\arguments{
\item{x}{The grid data. Class: matrix.}

\item{y}{The observations data. Class: matrix.}

\item{dates}{Dates of the grid and observations data.}

\item{n.analogs}{An integer. Number of analogs. Default is 4.}

\item{sel.fun}{A string. Select a function to apply to the analogs selected for a given observation. Options are 
"mean", "wmean" (i.e., weighted mean), "max", "min", "median", "prcXX" 
(i.e., prc85 means the 85th percentile of the analogs values distribution). Default is "mean".
the function applied to the analogs values, (i.e., sel.fun = c("mean","max","min","median","prcXX"), with default "mean") 
and the temporal window, (i.e., window = 0).}

\item{window}{An integer. Window of days removed when selecting analogs. 
If window = 7, then 7 days after the observation date and the 7 days before the observation date are removed. Default is 0.}

\item{n.random}{An integer. Choose N random analogs among the closest n.analogs. Default is NULL.}

\item{pool}{An integer. Number of auxiliary analogs in case there are NaN or NA in the original analogs.}
}
\value{
A list containing the grid data, the observations and a list with information concerning the analogs.
}
\description{
Analog method implementation
}
\details{
The analogs actually is not a model in the sense that it is not really trained. 
The model would be the data where to search the analogs. For this reason this function only saves the information to search analogs in a list.
}
\author{
J. Bano-Medina
}
