% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeAggregation.R
\name{annualAggregationMain}
\alias{annualAggregationMain}
\title{Time aggregation of the data in a grid}
\usage{
annualAggregationMain(data, FUN = c("mean", "min", "max", "sum"), id, y.coord,
  x.coord, coords, obj, parallel.pars = parallel.pars)
}
\arguments{
\item{data}{array (can be multimember).}

\item{FUN}{a character indicating the aggregation function. Possibilities are: "mean", "min", "max", "sum".}

\item{id}{as argument INDEX in function tapply.}

\item{y.coord}{y.coord}

\item{x.coord}{x.coord}

\item{coords}{coords}

\item{obj}{obj}

\item{parallel.pars}{Arguments defining the parallelization options, as passed by \code{\link{parallelCheck}}}
}
\value{
array  with the time dimension aggregated.
}
\description{
Aggregates data daily, monthly or annually by specifiyng an aggregation function
}
\author{
M. Iturbide, M. de Felice, J. Bedia
}
\keyword{internal}

