% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeAggregation.R
\name{annualAggregationMain}
\alias{annualAggregationMain}
\title{Time aggregation of the data in a field}
\usage{
annualAggregationMain(data, FUN = c("mean", "min", "max", "sum"), id, y.coord,
  x.coord, coords, obj)
}
\arguments{
\item{data}{array (can be multimember).}

\item{FUN}{a character indicating the aggregation function. Possibilities are: "mean", "min", "max", "sum".}

\item{id}{as argument INDEX in function tapply.}

\item{y.coord}{y.coord}

\item{x.coord}{x.coord}

\item{coords}{coords}

\item{obj}{obj}
}
\value{
array  with the time dimension aggregated.
}
\description{
Aggregates data daily, monthly or annually by specifiyng an aggregation function
}
\author{
M. Iturbide \email{maibide@gmail.com}
}

