% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/biasCorrection.R
\name{biasCorrection}
\alias{biasCorrection}
\title{Bias correction methods}
\usage{
biasCorrection(obs, pred, sim, method = c("eqm", "delta", "scaling", "gqm",
  "gpqm"), pr.threshold = 1, multi.member = TRUE, window = NULL,
  scaling.type = c("additive", "multiplicative"), extrapolation = c("no",
  "constant"), theta = 0.95)
}
\arguments{
\item{obs}{A field or station data containing the observed climate data for the training period}

\item{pred}{A field containing the simulated climate by the model for the training period. This can be either
the same variable as \code{obs}, in the case of model calibration (bias correction and related techniques) or MOS (model
 output statistics) approaches, or a set of predictors in the case of \emph{perfect prog} downscaling approaches (possibly
 after principal component analysis via the \code{\link{prinComp}} output).}

\item{sim}{A field containing the simulated climate for the variables used in \code{pred}, but considering the test period.}

\item{method}{method applied. Current accepted values are \code{"eqm"}, \code{"delta"},
\code{"scaling"}, \code{"gqm"} and \code{"gpqm"}. See details.}

\item{pr.threshold}{The minimum value that is considered as a non-zero precipitation. Ignored for
\code{varcode} values different from \code{"pr"}. Default to 1 (assuming mm).}

\item{multi.member}{Should members be adjusted sepparately (TRUE, default), or jointly (FALSE)?.
Ignored if the dataset has no members.}

\item{window}{Numeric value specifying the time window width used to calibrate. The window is centered on the target day.
Default to \code{NULL}, which considers the whole period available.}

\item{scaling.type}{Character indicating the type of the scaling method. Options are \code{"additive"} (default)
or \code{"multiplicative"} (see details). This argument is ignored if \code{"scaling"} is not
selected as the bias correction method.}

\item{extrapolation}{Character indicating the extrapolation method to be applied to correct values in
\code{"sim"} that are out of the range of \code{"pred"}. Extrapolation is applied only to the \code{"eqm"} method,
thus, this argument is ignored if other bias correction method is selected. Default is \code{"no"} (do not extrapolate).}

\item{theta}{numeric indicating  upper threshold (and lower for the left tail of the distributions, if needed)
above which precipitation (temperature) values are fitted to a Generalized Pareto Distribution (GPD).
Values below this threshold are fitted to a gamma (normal) distribution. By default, 'theta' is the 95th
percentile (5th percentile for the left tail). Only for \code{"gpqm"} method.}
}
\value{
A calibrated object of the same spatio-temporal extent of the input field
}
\description{
Implementation of several standard bias correction methods
}
\details{
The methods available are \code{"eqm"}, \code{"delta"},
\code{"scaling"}, \code{"gqm"}, \code{"gpqm"} (the two latter used only for precipitation). In the case of precipitation a
frequency adaptation has been implemented in all versions of qqmap to alleviate the problem that arise when the dry day frequency
in the raw model output is larger than in the observations (Wilcke et al. 2013).

Next we make a brief description of each method:

\strong{Delta}

This method consists on adding to the observations the mean change signal (delta method).
This method is applicable to any kind of variable but it is preferable to avoid it for bounded variables
(e.g. precipitation, wind speed, etc.) because values out of the variable range could be obtained
(e.g. negative wind speeds...).

\strong{Scaling}

This method consists on scaling the simulation  with the difference (additive) or quotient (multiplicative)
between the observed and simulated means in the train period. The \code{additive} or \code{multiplicative}
correction is defined by parameter \code{scaling.type} (default is \code{additive}).
The additive version is preferably applicable to unbounded variables (e.g. temperature)
and the multiplicative to variables with a lower bound (e.g. precipitation, because it also preserves the frequency).

\strong{eqm}

Empirical Quantile Mapping. This is a very extended bias correction method which consists on calibrating the simulated Cumulative Distribution Function (CDF)
by adding to the observed quantiles both the mean delta change and the individual delta changes in the corresponding quantiles.
This method is applicable to any kind of variable.

\strong{gqm}

Gamma Quantile Mapping. This method is described in Piani et al. 2010 and is applicable only to precipitation. It is based on the initial assumption that both observed
and simulated intensity distributions are well approximated by the gamma distribution, therefore is a parametric q-q map
that uses the theorical instead of the empirical distribution.

\strong{gpqm}

Generalized Quantile Mapping. This method is described in Gutjahr and Heinemann 2013. It is applicable only to precipitation and is similar to the Piani method. It applies a
gamma distribution to values under the threshold given by the 95th percentile (following Yang et al. 2010) and a general Pareto
distribution (GPD) to values above the threshold.
}
\examples{
\dontrun{
# Download VALUE (station data) and NCEP (model data) datasets
dir.create("mydirectory")
download.file("http://meteo.unican.es/work/downscaler/data/VALUE_ECA_86_v2.tar.gz",
              destfile = "mydirectory/VALUE_ECA_86_v2.tar.gz")
download.file("http://meteo.unican.es/work/downscaler/data/Iberia_NCEP.tar.gz",
              destfile = "mydirectory/Iberia_NCEP.tar.gz")
# Extract files from the tar.gz file
untar("mydirectory/VALUE_ECA_86_v2.tar.gz", exdir = "mydirectory")
untar("mydirectory/NCEP_Iberia.tar.gz", exdir = "mydirectory")
# Path to the VALUE dataset and the NCEP ncml file.
value <- "mydirectory/VALUE_ECA_86_v2"
ncep <- "mydirectory/Iberia_NCEP/Iberia_NCEP.ncml"
# Data inventories provides a quick overview of the available data
value.inv <- dataInventory(value)
ncep.inv <- dataInventory(ncep)
str(value.inv)
str(ncep.inv)
# Load precipitation for boreal winter (DJF) in the train (1991-2000) and test (2001-2010) periods,
# for the observations (VALUE) and the Iberia_NCEP datasets
obs <- loadStationData(dataset = value,
                       var="precip",
                       lonLim = c(-12,10), latLim = c(33,47),
                       season = c(12,1,2),
                       years = 1991:2000)
prd <- loadGridData(dataset = ncep,
                    var = "tp",
                    lonLim = c(-12,10), latLim = c(33,47),
                    season = c(12,1,2),
                    years = 1991:2000)
sim <- loadGridData(dataset = ncep,
                    var = "tp",
                    lonLim = c(-12,10), latLim = c(33,47),
                    season = c(12,1,2),
                    years = 2001:2010)
# Interpolation of the observations onto the grid of model: we use the method "nearest" and the
# 'getGrid' function to ensure spatial consistency:
obs <- interpData(obs, new.Coordinates = getGrid(prd), method = "nearest")
# Apply the bias correction method:
simBC <- biasCorrection (obs, prd, sim, method = "eqm", pr.threshold = 1) # qq-mapping
par(mfrow = c(1,2))
plotMeanField(sim)
plotMeanField(simBC)
}
}
\author{
S. Herrera \email{sixto@predictia.es}
}
\references{
\itemize{
\item R.A.I. Wilcke, T. Mendlik and A. Gobiet (2013) Multi-variable error correction of regional climate models. Climatic Change, 120, 871-887

\item A. Amengual, V. Homar, R. Romero, S. Alonso, and C. Ramis (2012) A Statistical Adjustment of Regional Climate Model Outputs to Local Scales: Application to Platja de Palma, Spain. J. Clim., 25, 939-957

\item C. Piani, J. O. Haerter and E. Coppola (2009) Statistical bias correction for daily precipitation in regional climate models over Europe, Theoretical and Applied Climatology, 99, 187-192

\item O. Gutjahr and G. Heinemann (2013) Comparing precipitation bias correction methods for high-resolution regional climate simulations using COSMO-CLM, Theoretical and Applied Climatology, 114, 511-529
}
}
\seealso{
\code{\link{isimip}} for a trend-preserving method of model calibration

Other downscaling: \code{\link{analogs}};
  \code{\link{calibrateProj}}; \code{\link{glimpr}};
  \code{\link{isimip}}
}

