% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biasCorrection.R
\name{biasCorrection}
\alias{biasCorrection}
\title{Bias correction methods}
\usage{
biasCorrection(y, x, newdata, precipitation = FALSE, method = c("delta",
  "scaling", "eqm", "gqm", "gpqm"), cross.val = c("none", "loocv", "kfold"),
  folds = NULL, window = NULL, scaling.type = c("additive",
  "multiplicative"), wet.threshold = 1, n.quantiles = NULL,
  extrapolation = c("none", "constant"), theta = 0.95)
}
\arguments{
\item{y}{A grid or station data containing the observed climate data for the training period}

\item{x}{A grid containing the simulated climate by the model for the training period. This can be either
the same variable as \code{obs}, in the case of model calibration (bias correction and related techniques) or MOS (model
 output statistics) approaches, or a set of predictors in the case of \emph{perfect prog} downscaling approaches (possibly
 after principal component analysis via the \code{\link{prinComp}} output).}

\item{newdata}{A grid containing the simulated climate for the variables used in \code{x}, but considering the test period.}

\item{precipitation}{Logical indicating if the data to be corrected is precipitation data.}

\item{method}{method applied. Current accepted values are \code{"eqm"}, \code{"delta"},
\code{"scaling"}, \code{"gqm"} and \code{"gpqm"}. See details.}

\item{cross.val}{Should cross-validation be performed? methods available are leave-one-out ("loocv") and k-fold ("kfold"). The default 
option ("none") does not perform cross-validation.}

\item{folds}{Only requiered if cross.val = "kfold". A list of vectors, each containing the years to be grouped in 
the corresponding fold.}

\item{window}{vector of length = 2 specifying the time window width used to calibrate and the target days (days that are being corrected).
 The window is centered on the target day/s. 
Default to \code{NULL}, which considers the whole period available.}

\item{scaling.type}{Character indicating the type of the scaling method. Options are \code{"additive"} (default)
or \code{"multiplicative"} (see details). This argument is ignored if \code{"scaling"} is not 
selected as the bias correction method.}

\item{wet.threshold}{The minimum value that is considered as a non-zero precipitation. Ignored for
\code{varcode} values different from \code{"pr"}. Default to 1 (assuming mm). See details on bias correction for precipitation.}

\item{n.quantiles}{Integer indicating the number of quantiles to be considered when method = "eqm". Default is NULL, 
that considers all quantiles, i.e. \code{n.quantiles = length(x[i,j])} (being \code{i} and \code{j} the coordinates in a single location).}

\item{extrapolation}{Character indicating the extrapolation method to be applied to correct values in
\code{newdata} that are out of the range of \code{x}. Extrapolation is applied only to the \code{"eqm"} method, 
thus, this argument is ignored if other bias correction method is selected. Default is \code{"none"} (do not extrapolate).}

\item{theta}{numeric indicating  upper threshold (and lower for the left tail of the distributions, if needed) 
above which precipitation (temperature) values are fitted to a Generalized Pareto Distribution (GPD). 
Values below this threshold are fitted to a gamma (normal) distribution. By default, 'theta' is the 95th 
percentile (5th percentile for the left tail). Only for \code{"gpqm"} method.}
}
\value{
A calibrated grid of the same spatio-temporal extent than the input \code{"y"}
}
\description{
Implementation of several standard bias correction methods
}
\details{
The methods available are \code{"eqm"}, \code{"delta"}, 
\code{"scaling"}, \code{"gqm"}, \code{"gpqm"} (the two latter used only for precipitation).
 These are next briefly described: 
 
\strong{Delta}

This method consists on adding to the observations the mean change signal (delta method).
This method is applicable to any kind of variable but it is preferable to avoid it for bounded variables
(e.g. precipitation, wind speed, etc.) because values out of the variable range could be obtained
(e.g. negative wind speeds...).

\strong{Scaling}

This method consists on scaling the simulation  with the difference (additive) or quotient (multiplicative) 
between the observed and simulated means in the train period. The \code{additive} or \code{multiplicative}
correction is defined by parameter \code{scaling.type} (default is \code{additive}).
The additive version is preferably applicable to unbounded variables (e.g. temperature) 
and the multiplicative to variables with a lower bound (e.g. precipitation, because it also preserves the frequency). 

\strong{eqm}

Empirical Quantile Mapping. This is a very extended bias correction method which consists on calibrating the simulated Cumulative Distribution Function (CDF) 
by adding to the observed quantiles both the mean delta change and the individual delta changes in the corresponding quantiles. 
This method is applicable to any kind of variable.

\strong{gqm}

Gamma Quantile Mapping. This method is described in Piani et al. 2010 and is applicable only to precipitation. It is based on the initial assumption that both observed
and simulated intensity distributions are well approximated by the gamma distribution, therefore is a parametric q-q map 
that uses the theorical instead of the empirical distribution. 
 
\strong{gpqm}
 
Generalized Quantile Mapping. This method is described in Gutjahr and Heinemann 2013. It is applicable only to precipitation and is similar to the Piani method. It applies a 
gamma distribution to values under the threshold given by the 95th percentile (following Yang et al. 2010) and a general Pareto 
distribution (GPD) to values above the threshold.
}
\section{Note on the bias correction of precipitation}{


In the case of precipitation a frequency adaptation has been implemented in all versions of 
qqmap to alleviate the problems arising when the dry day frequency in the raw model output is larger
 than in the observations (Wilcke et al. 2013). 
 
 The precipitation subroutines are switched-on when the variable name of the grid 
 (i.e., the value returned by \code{gridData$Variable$varName}) is one of the following: 
 \code{"pr"}, \code{"tp"} (this is the standard name defined in the vocabulary (\code{\link[loadeR]{UDG.vocabulary}}), \code{"precipitation"} or \code{"precip"}.
 Thus, caution must be taken to ensure that the correct bias correction is being undertaken when dealing with
 non-standard variables.
}
\examples{
\dontrun{
data(VALUE_Igueldo_tp)
data(NCEP_Iberia_tp)
y <- VALUE_Igueldo_tp
x <- NCEP_Iberia_tp

eqm1 <- biasCorrection(y = y, x = x, newdata = x,
                       method = "eqm",
                       extrapolation = "none",
                       window = NULL)
eqm1win <- biasCorrection(y = y, x = x, newdata = x,
                          method = "eqm",
                          extrapolation = "none",
                          window = c(90, 10))
NCEP_Igueldo <- subsetGrid(x, latLim = y$xyCoords[,2], lonLim = y$xyCoords[,1])
par(mfrow = c(1,3))
qqplot(y$Data, NCEP_Igueldo$Data)
lines(c(0,100), c(0,100))
qqplot(y$Data, eqm1$Data)
lines(c(0,100), c(0,100))
qqplot(y$Data, eqm1win$Data)
lines(c(0,100), c(0,100))
par(mfrow = c(1,1))
}
}
\author{
S. Herrera and M. Iturbide
}
\references{
\itemize{
\item R.A.I. Wilcke, T. Mendlik and A. Gobiet (2013) Multi-variable error correction of regional climate models. Climatic Change, 120, 871-887

\item A. Amengual, V. Homar, R. Romero, S. Alonso, and C. Ramis (2012) A Statistical Adjustment of Regional Climate Model Outputs to Local Scales: Application to Platja de Palma, Spain. J. Clim., 25, 939-957

\item C. Piani, J. O. Haerter and E. Coppola (2009) Statistical bias correction for daily precipitation in regional climate models over Europe, Theoretical and Applied Climatology, 99, 187-192

\item O. Gutjahr and G. Heinemann (2013) Comparing precipitation bias correction methods for high-resolution regional climate simulations using COSMO-CLM, Theoretical and Applied Climatology, 114, 511-529
}
}
\seealso{
\code{\link{isimip}} for a trend-preserving method of model calibration

Other downscaling: \code{\link{analogs}},
  \code{\link{downscale}}, \code{\link{glimpr}},
  \code{\link{isimip}}
}

