% Generated by roxygen2 (4.0.1): do not edit by hand
\name{biasCorrection}
\alias{biasCorrection}
\title{Bias correction methods}
\usage{
biasCorrection(obs, pred, sim, method = c("qqmap", "delta", "scaling",
  "unbiasing", "piani"), pr.threshold = 1)
}
\arguments{
\item{obs}{A field or station data containing the observed climate data for the training period}

\item{pred}{A field containing the simulated climate by the model for the training period. This can be either
the same variable as \code{obs}, in the case of model calibration (bias correction and related techniques) or MOS (model
 output statistics) approaches, or a set of predictors in the case of \emph{perfect prog} downscaling approaches (possibly
 after principal component analysis via the \code{\link{prinComp}} output).}

\item{sim}{A field containing the simulated climate for the variables used in \code{pred}, but considering the test period.}

\item{method}{method applied. Current accepted values are \code{"qqmap"}, \code{"delta"},
\code{"scaling"}, \code{"unbiasing"} and \code{"piani"}. See details.}

\item{pr.threshold}{The minimum value that is considered as a non-zero precipitation. Ignored for
 \code{varcode} values different from \code{"pr"}. Default to 1 (assuming mm).}
}
\value{
A calibrated object of the same spatio-temporal extent of the input field
}
\description{
Implementation of several standard bias correction methods
}
\details{
The methods available are \code{"qqmap"}, \code{"delta"}, \code{"unbiasing"},
\code{"scaling"} and \code{"Piani"} (the latter used only for precipitation).
Next we make a brief description of each method:

\strong{Delta}

This method consists on adding to the observations the mean change signal (delta method).
This method is applicable to any kind of variable but it is preferable to avoid it for bounded variables
 (e.g. precipitation, wind speed, etc.) because values out of the variable range could be obtained (e.g. negative wind speeds...).

\strong{Unbiasing}

This correction consists on adding to the simulation the mean diference between the observations
and the simulation in the train period. This method is preferably applicable to unbounded variables (e.g. temperature).

\strong{Scaling}

This method consists on scaling the simulation by the quotient of the mean observed and simulated in the train period.
This method is preferably applicable to variables with a lower bound, such as precipitation, because it preserves the
frequency also.

\strong{Quantile-Quantile Mapping (qqmap)}

This is a very extended bias correction method which consists on calibrating the simulated Cumulative Distribution Function (CDF)
by adding to the observed quantiles both the mean delta change and the individual delta changes in the corresponding quantiles.
This method is applicable to any kind of variable.

\strong{Piani}

This method is described in Piani et al. 2010 and is applicable only to precipitation. It is based on the initial assumption that both observed
 and simulated intensity distributions are well approximated by the gamma distribution, therefore is a parametric q-q map
 that uses the theorical instead of the empirical distribution.
}
\examples{
\dontrun{
# These are the paths to the package built-in GSN and NCEP datasets
gsn.data.dir <- file.path(find.package("downscaleR"), "datasets/observations/GSN_Iberia")
ncep.data.dir <- file.path(find.package("downscaleR"), "datasets/reanalysis/Iberia_NCEP/Iberia_NCEP.ncml")
gsn.inv <- dataInventory(gsn.data.dir)
ncep.inv <- dataInventory(ncep.data.dir)
str(gsn.inv)
str(ncep.inv)
# Load precipitation for boreal winter (DJF) in the train (1991-2000) and test (2001-2010) periods, for the observations (GSN_Iberia) and the Iberia_NCEP datasets
obs <- loadStationData(dataset = gsn.data.dir, var="precip", lonLim = c(-12,10), latLim = c(33,47), season=c(12,1,2), years = 1991:2000)
prd <- loadGridData(ncep.data.dir, var = "tp", lonLim = c(-12,10), latLim = c(33,47), season = c(12,1,2), years = 1991:2000)
sim <- loadGridData(ncep.data.dir, var = "tp", lonLim = c(-12,10), latLim = c(33,47), season = c(12,1,2), years = 2001:2010)
# Interpolation of the observations onto the grid of model: we use the method "nearest" and the getGrid function to ensure spatial consistency:
obs <- interpGridData(obs, new.grid = getGrid(prd), method = "nearest")
# Apply the bias correction method:
simBC <- biasCorrection (obs, prd, sim, method = "qqmap", pr.threshold = 1) # qq-mapping
par(mfrow = c(1,2))
plotMeanField(sim)
plotMeanField(simBC)
}
}
\author{
S. Herrera \email{sixto@predictia.es}
}
\references{
\itemize{
\item R.A.I. Wilcke, T. Mendlik and A. Gobiet (2013) Multi-variable error correction of regional climate models. Climatic Change, 120, 871-887

\item A. Amengual, V. Homar, R. Romero, S. Alonso, and C. Ramis (2012) A Statistical Adjustment of Regional Climate Model Outputs to Local Scales: Application to Platja de Palma, Spain. J. Clim., 25, 939-957

\item C. Piani, J. O. Haerter and E. Coppola (2009) Statistical bias correction for daily precipitation in regional climate models over Europe, Theoretical and Applied Climatology, 99, 187-192
}
}
\seealso{
\code{\link{isimip}} for a trend-preserving method of model calibration

Other downscaling: \code{\link{analogs}};
  \code{\link{calibrateProj}}; \code{\link{isimip}}
}

