% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/biasCorrection.R
\name{calibrateProj}
\alias{calibrateProj}
\title{Bias correction methods}
\usage{
calibrateProj(obs, pred, sim, method = c("eqm", "delta", "scaling", "gqm",
  "gpqm"), varcode = c("tas", "hurs", "tp", "pr", "wss"), pr.threshold = 1,
  scaling.type = scaling.type, extrapolate = c("no", "constant"))
}
\arguments{
\item{obs}{A field or station data containing the observed climate data for the training period}

\item{pred}{A field containing the simulated climate by the model for the training period. This can be either
the same variable as \code{obs}, in the case of model calibration (bias correction and related techniques) or MOS (model
 output statistics) approaches, or a set of predictors in the case of \emph{perfect prog} downscaling approaches (possibly
 after principal component analysis via the \code{\link{prinComp}} output).}

\item{sim}{A field containing the simulated climate for the variables used in \code{pred}, but considering the test period.}

\item{method}{method applied. Current accepted values are \code{"eqm"}, \code{"delta"},
\code{"scaling"}, \code{"gqm"} and \code{"gpqm"}. See details.}

\item{varcode}{Variable code. This is not the variable itself to be corrected, but
rather it referes to its nature and distributional properties. For instance, \code{"tas"} applies for
temperature-like variables (i.e.: unbounded gaussian variables that can take both negative and positive values),
\code{"hurs"} for relative-humidity-like variables (i.e., bounded by both sides, with a maximum value of 1 -100-
and a minimum of zero, with no negatives), \code{"wss"} for wind-like variables (no possible values below zero,
but without an upper bound) and \code{"pr"}, specifically for precipitation.}

\item{pr.threshold}{The minimum value that is considered as a non-zero precipitation. Ignored for
\code{varcode} values different from \code{"pr"}. Default to 1 (assuming mm).}

\item{scaling.type}{Character indicating the type of the scaling method. Options are \code{"additive"} (default)
or \code{"multiplicative"}. This argument is ignored if \code{"scaling"} is not
selected as the bias correction method.}

\item{extrapolate}{Character indicating the extrapolation method to be applied to correct values in
\code{"sim"} that are out of the range of \code{"pred"}. Extrapolation is applied only to the \code{"qqmap"} method,
thus, this argument is ignored if other bias correction method is selected. Default is \code{"no"} (do not extrapolate).}
}
\description{
Implementation of several standard bias correction methods
}
\seealso{
Other downscaling: \code{\link{analogs}};
  \code{\link{biasCorrection}}; \code{\link{glimpr}};
  \code{\link{isimip}}
}
\keyword{internal}

