% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calibrateProj}
\alias{calibrateProj}
\title{Bias correction methods}
\usage{
calibrateProj(obs, pred, sim, method = c("qqmap", "delta", "scaling",
  "unbiasing", "piani"), varcode = c("tas", "hurs", "pr", "wss"),
  pr.threshold = 1)
}
\arguments{
\item{obs}{A field or station data containing the observed climate data for the training period}

\item{pred}{A field containing the simulated climate by the model for the training period. This can be either
the same variable as \code{obs}, in the case of model calibration (bias correction and related techniques) or MOS (model
 output statistics) approaches, or a set of predictors in the case of \emph{perfect prog} downscaling approaches (possibly
 after principal component analysis via the \code{\link{prinComp}} output).}

\item{sim}{A field containing the simulated climate for the variables used in \code{pred}, but considering the test period.}

\item{method}{method applied. Current accepted values are \code{"qqmap"}, \code{"delta"},
\code{"scaling"}, \code{"unbiasing"} and \code{"piani"}. See details.}

\item{varcode}{Variable code. This is not the variable itself to be corrected, but
rather it referes to its nature and distributional properties. For instance, \code{"tas"} applies for
temperature-like variables (i.e.: unbounded gaussian variables that can take both negative and positive values),
\code{"hurs"} for relative-humidity-like variables (i.e., bounded by both sides, with a maximum value of 1 -100-
and a minimum of zero, with no negatives), \code{"wss"} for wind-like variables (no possible values below zero,
but without an upper bound) and \code{"pr"}, specifically for precipitation.}

\item{pr.threshold}{The minimum value that is considered as a non-zero precipitation. Ignored for
\code{varcode} values different from \code{"pr"}. Default to 1 (assuming mm).}
}
\description{
Implementation of several standard bias correction methods
}
\author{
S. Herrera \email{sixto@predictia.es}
}
\seealso{
Other downscaling: \code{\link{analogs}};
  \code{\link{biasCorrection}}; \code{\link{isimip}}
}
\keyword{internal}

