% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dataInventory}
\alias{dataInventory}
\title{Dataset inventory}
\usage{
dataInventory(dataset, return.stats = FALSE)
}
\arguments{
\item{dataset}{A character string poiting to the target. Either a directory containing the data
in the case of station data in standard ASCII format (see \code{\link{loadStationData}}) for details),
or a target file (a NcML) in the case of other types of gridded data (reanalysis, gridded observations ...,
see \code{\link{loadGridData}} for details).}

\item{return.stats}{Optional logical flag indicating if summary statistics of the dataset
 should be returned with the inventory. Only used for station data.}
}
\value{
A list of components describing the variables and other characteristics of the target dataset.
}
\description{
Function to provide a quick overview of a climate dataset
 (either stations or gridded data)
}
\note{
The variable names returned correspond to the original names of the variables as stored in the dataset,
and not to the standard naming convention defined in the vocabulary.
}
\examples{
\donttest{
gsn <- file.path(find.package("downscaleR"), "datasets/observations/GSN_Iberia")
di <- dataInventory(gsn)
str(di)
# To obtain summary statistics of the variables stored:
di.stats <- dataInventory(gsn, return.stats = TRUE)
print(di.stats$Summary.stats)
}
}
\author{
J Bedia \email{joaquin.bedia@gmail.com}
}
\seealso{
\code{\link{stationInfo}} for a quick overview of available stations in station datasets.
}

