% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dictionaryTransform.R
\name{dictionaryTransform}
\alias{dictionaryTransform}
\title{Performs variable transformation}
\usage{
dictionaryTransform(dic, grid, timePars, mdArray)
}
\arguments{
\item{dic}{Dictionary line for the variable, as returned by \code{dictionaryLookup}}

\item{timePars}{A list of time selection parameters, as returned by \code{getTimeDomain}}

\item{mdArray}{A n-dimensional array, as returned by \code{makeSubset}}
}
\value{
a list with start/end dates and the transformed n-dimensional array of data.
}
\description{
Performs variable transformation according to dictionary specifications.
Sub-routine of \code{loadGridDataset}.
}
\author{
J. Bedia \email{joaquin.bedia@gmail.com}
}
\keyword{internal}

