% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscale.R
\name{downscale}
\alias{downscale}
\title{Downscale a Spatial Raster}
\usage{
downscale(from, to, FUN = "mean", progress = TRUE)
}
\arguments{
\item{from}{A RasterLayer, RasterStack or RasterBrick to downscale}

\item{to}{A RasterLayer with characteristics that from should be downscale to}

\item{FUN}{The function to aggregate values (DEFAULT: "mean")}

\item{progress}{If TRUE, a progress bar is displayed}
}
\value{
A RasterLayer, RasterStack or RasterBrick.
}
\description{
This function is an alternative to raster::resample() and downscales
a raster to coarser raster in a more efficient way (for large rasters).
}
\examples{
library(raster)

from  <- raster(system.file("external/test.grd", package = "raster"))
to    <- aggregate(from, fac = 2)

final <- downscale(from, to, FUN = "mean", progress = FALSE)

par(mfrow = c(1, 2))
plot(from)
plot(final)
}
\author{
Nicolas CASAJUS, \email{nicolas.casajus@fondationbiodiversite.fr}
}
