% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscale.R
\name{downscale}
\alias{downscale}
\title{...}
\usage{
downscale(from, to, FUN = "mean")
}
\arguments{
\item{from}{...}

\item{to}{...}

\item{FUN}{...}
}
\value{
...
}
\description{
...
}
\details{
...
}
\examples{
#...
}
\author{
Nicolas CASAJUS, \email{nicolas.casajus@fondationbiodiversite.fr}
}
