% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/downscale.R
\name{downscale}
\alias{downscale}
\title{Perfect-prog downscaling}
\usage{
downscale(obs, pred, sim = NULL, method = c("analogs", "glm"),
  n.neigh = 1, sel.fun = c("random", "mean"), analog.dates = FALSE,
  pr.threshold = 0.1, n.pcs = NULL, cross.val = c("none", "loocv"),
  parallel = FALSE, max.ncores = 16, ncores = NULL)
}
\arguments{
\item{obs}{A field or station data containing the observed climate data for the training period}

\item{pred}{A field containing the simulated climate by the model for the training period. This can be either
the same variable as \code{obs}, in the case of model calibration (bias correction and related techniques) or MOS (model
 output statistics) approaches, or a set of predictors in the case of \emph{perfect prog} downscaling approaches (possibly
 after principal component analysis via the \code{\link{prinComp}} output).}

\item{sim}{A field containing the simulated climate for the variables used in \code{pred}, but considering the test period.}

\item{method}{Downscaling method}

\item{n.neigh}{Applies only when \code{method="analogs"} (otherwise ignored). Integer indicating the number of closest neigbours to retain for analog construction. Default to 1.}

\item{sel.fun}{Applies only when \code{method="analogs"} (otherwise ignored). Criterion for the construction of analogs when several neigbours are chosen. Ignored when \code{n.neig = 1}.
Current values are \code{"random"} (the default) and \code{"mean"}. See details.}

\item{analog.dates}{Logical flag indicating whether the dates of the analogs should be returned. See the analogs section.}

\item{pr.threshold}{Value below which precipitation amount is considered zero}

\item{n.pcs}{Integer indicating the number of EOFs to be used as predictors}

\item{cross.val}{Should cross-validation be performed?}

\item{parallel}{Logical. Should parallel execution be used?}

\item{max.ncores}{Integer. Upper bound for user-defined number of cores.}

\item{ncores}{Integer number of cores used in parallel computation.
Self-selected number of cores is used when \code{ncpus = NULL} (the default),
 or when \code{maxcores} exceeds the default \code{ncores} value.}
}
\description{
Workhorse function to call the different perfect-prog downscaling methods
}
\details{
\strong{Spatial consistency}

Several checks of spatial consistency are performed. In particular, note that both \code{pred} (reanalysis) and \code{sim}
 (model simulations) should be in the same grid. This consistency must be ensured by the user prior to entering these arguments,
for instance by means of the \code{\link{interpData}} function in conjunction with the \code{\link{getGrid}} method.

\strong{Scaling and centering}

When the climate variables are used as predictors instead of the PCs, these are previously centered and scaled
using the mean and sigma parameters globally computed for the whole spatial domain (This is equivalent to the \dQuote{field})
method in the \code{\link{prinComp}} function. The simulation data will use the parameters obtained when scaling and centering
the predictors dataset. In case that the predictors come from a PC analysis object (as returned by \code{\link{prinComp}}), the
parameters for rescaling the simulation data are passed by the predictors.
}
\section{Analogs}{


\strong{Construction of analogs using multiple neighbours}

The argument \code{sel.fun} controls how the analogs are constructed when considering more than the first neighbour (argument
\code{n.neigh} > 1). In this case the \code{"random"} choice randomly selects one of the \code{n.neigh} neighbours,
 while the \code{"mean"} choice will compute their average.

\strong{Analog dates}

If the argument \code{analog.dates} is set to TRUE, these will be returned as a global attribute named \code{"analog.dates"}.
Note that the analog dates can be only returned for one single neighbour selections (i.e. \code{n.neigh = 1}),
 otherwise giving an error. The analog dates are different for each member in case of
 multimember downscaling, and are returned as a list, each element of the list corresponding to one member.
}

\section{Parallel Processing}{


Parallel processing is enabled using the \pkg{parallel} package.
Parallelization is undertaken by a FORK-type parallel socket cluster formed by \code{ncores}.
If \code{ncores} is not specified (default), \code{ncores} will be one less than the autodetected number of cores.
The maximum number of cores used for parallel processing can be set with the \code{max.ncores} argument,
although this will be reset to the auto-detected number of cores minus 1 if this number is exceeded. Note that not all
code, but just some critical loops within the function are parallelized.

In practice, parallelization does not always result in smaller execution times, due to the parallel overhead.
However, parallel computing may potentially provide a significant speedup for the
particular case of large multimember datasets or large grids.

Parallel computing is currently not available for Windows machines.
}
\author{
J Bedia and M Iturbide
}
\seealso{
\code{\link{prinComp}} for details on principal component/EOF analysis,
\code{rescaleMonthlyMeans} for data pre-processing,
\code{\link{makeMultiField}} for multifield creation
\code{\link{loadGridData}} and \code{\link{loadStationData}} for loading fields and station data respectively.

Other downscaling: \code{\link{analogs}};
  \code{\link{biasCorrection}};
  \code{\link{calibrateProj}}; \code{\link{glimpr}};
  \code{\link{isimip}}
}

