% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscaler-package.R
\docType{package}
\name{downscaler-package}
\alias{downscaler}
\alias{downscaler-package}
\title{downscaler: Downscale a Spatial Raster}
\description{
This package is an alternative to raster::resample() and downscales a raster to coarser resolution in a more efficient way (for large rasters).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/FRBCesab/downscaler}
  \item Report bugs at \url{https://github.com/FRBCesab/downscaler/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Casajus \email{nicolas.casajus@fondationbiodiversite.fr}

}
\keyword{internal}
