% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscaler-package.R
\docType{package}
\name{downscaler-package}
\alias{downscaler}
\alias{downscaler-package}
\title{downscaler: What the Package Does}
\description{
What the package does.
}
\author{
\strong{Maintainer}: Nicolas Casajus \email{nicolas.casajus@fondationbiodiversite.fr}

}
\keyword{internal}
