% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/biasCorrection.R
\name{eqm}
\alias{eqm}
\title{Empirical Quantile Mapping method for bias correction}
\usage{
eqm(o, p, s, precip, pr.threshold, n.quantiles, extrapolation)
}
\arguments{
\item{o}{A vector (e.g. station data) containing the observed climate data for the training period}

\item{p}{A vector containing the simulated climate by the model for the training period.}

\item{s}{A vector containing the simulated climate for the variable used in \code{p}, but considering the test period.}

\item{precip}{Logical indicating if o, p, s is precipitation data.}

\item{pr.threshold}{The minimum value that is considered as a non-zero precipitation. Ignored for
\code{varcode} values different from \code{"pr"}. Default to 1 (assuming mm).}

\item{extrapolation}{Character indicating the extrapolation method to be applied to correct values in
\code{"s"} that are out of the range of \code{"p"}. Extrapolation is applied only to the \code{"eqm"} method,
thus, this argument is ignored if other bias correction method is selected. Default is \code{"none"} (do not extrapolate).}
}
\description{
Implementation of Empirical Quantile Mapping method for bias correction
}
\author{
S. Herrera and M. Iturbide
}
\keyword{internal}

