% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getGrid}
\alias{getGrid}
\title{Get regular grid definition}
\usage{
getGrid(gridData)
}
\arguments{
\item{gridData}{A grid data object coming from \code{\link{loadGridData}} or \code{\link{interpGridData}}
or the function \code{\link[ecomsUDG.Raccess]{loadECOMS}} of package \pkg{ecomsUDG.Raccess}.}
}
\description{
Get the (regular) grid definition from an existing (gridded) dataset
}
\details{
The returned grid object inherits the attributes from the input \code{xyCoords} definition.
}
\author{
J Bedia \email{joaquin.bedia@gmail.com}
}
\seealso{
Other loading.grid: \code{\link{interpGridData}};
  \code{\link{loadGridData}}
}

