% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getGrid}
\alias{getGrid}
\title{Get regular grid definition}
\usage{
getGrid(gridData)
}
\arguments{
\item{gridData}{A grid data object coming from \code{\link{loadGridData}} or \code{\link{interpGridData}}
 or the function \code{\link[ecomsUDG.Raccess]{loadECOMS}} of package \pkg{ecomsUDG.Raccess}.}
}
\value{
A list of two named components, \code{x} and \code{y}, consisting of a vector of length two each one, defining
the x/y lower and upper bounds. The grid-cell resolution is given by the attributes \code{'resX'} and
 \code{'resY'} respectively.
}
\description{
Get the (regular) grid definition from an existing (gridded) dataset
}
\details{
In case of irregular grid definitions, the function forces the grid to regularity.
The returned grid object inherits the attributes from the input \code{xyCoords} definition.
}
\examples{
\donttest{
# Iberia domain
data(iberia_ncep_hus850)
getGrid(iberia_ncep_hus850)
# Europe domain
data(tasmin_forecast)
getGrid(tasmin_forecast)
plotMeanField(tasmin_forecast)
# Interpolate NCEP onto the System4 grid:
int <- interpGridData(iberia_ncep_hus850, getGrid(tasmin_forecast), "bilinear")
# Note the warnings because of the non-overlapping domain extents
plotMeanField(int)
# The other way round, using nearest neighbour interpolation:
int2 <- interpGridData(tasmin_forecast, getGrid(iberia_ncep_hus850))
plotMeanField(int2)
#' # In this case, the mismatch in domain extent occurs only in the longitudes (to the west)
}
}
\author{
S. Herrera and J. Bedia \email{joaquin.bedia@gmail.com}
}
\seealso{
Other loading.grid: \code{\link{interpGridData}};
  \code{\link{loadGridData}}
}

