% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getSeason}
\alias{getSeason}
\title{Get season from a station or field object}
\usage{
getSeason(obj)
}
\arguments{
\item{obj}{Any object extending the station or field classes}
}
\value{
An integer vector with the season
}
\description{
Retrieves the season encompassed by a station or field object
}
\examples{
data(iberia_ncep_ta850)
getSeason(iberia_ncep_ta850) # Boreal winter (DJF)
}
\author{
J. Bedia \email{joaquin.bedia@gmail.com}
}

