% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/glimpr.R
\name{glimpr}
\alias{glimpr}
\title{GLM downscaling}
\usage{
glimpr(obs, pred, sim, pr.threshold = 0.1)
}
\arguments{
\item{obs}{A field or station data containing the observed climate data for the training period}

\item{pred}{A field containing the simulated climate by the model for the training period. This can be either
the same variable as \code{obs}, in the case of model calibration (bias correction and related techniques) or MOS (model
 output statistics) approaches, or a set of predictors in the case of \emph{perfect prog} downscaling approaches (possibly
 after principal component analysis via the \code{\link{prinComp}} output).}

\item{sim}{A field containing the simulated climate for the variables used in \code{pred}, but considering the test period.}

\item{pr.threshold}{Value below which precipitation amount is considered zero}
}
\description{
Implementation of GLM's to downscale precipitation data
}
\author{
J Bedia \email{joaquin.bedia@gmail.com}
}
\seealso{
Other downscaling: \code{\link{analogs}};
  \code{\link{biasCorrection}};
  \code{\link{calibrateProj}}; \code{\link{isimip}}
}

