% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glimpr.R
\name{glimpr}
\alias{glimpr}
\title{GLM downscaling}
\usage{
glimpr(y = y, modelPars = modelPars, simulate = TRUE,
  return.models = FALSE, wet.threshold = wet.threshold, n.pcs = n.pcs)
}
\arguments{
\item{y}{A grid or station data containing the observed climate data for the training period}

\item{modelPars}{Output object from function ppModelSetup containing the predictors and test data}

\item{simulate}{Logical. Default is TRUE.}

\item{wet.threshold}{Value below which precipitation amount is considered zero}

\item{n.eofs}{Integer indicating the number of EOFs to be used as predictors}
}
\description{
Implementation of GLM's to downscale precipitation data
}
\author{
J Bedia, M Iturbide
}
\seealso{
Other downscaling: \code{\link{analogs}};
  \code{\link{biasCorrection}}; \code{\link{downscale}};
  \code{\link{isimip}}
}
\keyword{internal}

