% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biasCorrection.R
\name{gpqm}
\alias{gpqm}
\title{Generalized Quantile Mapping method for bias correction}
\usage{
gpqm(o, p, s, precip, pr.threshold, theta)
}
\arguments{
\item{o}{A vector (e.g. station data) containing the observed climate data for the training period}

\item{p}{A vector containing the simulated climate by the model for the training period.}

\item{s}{A vector containing the simulated climate for the variable used in \code{p}, but considering the test period.}

\item{precip}{Logical indicating if o, p, s is precipitation data.}

\item{pr.threshold}{The minimum value that is considered as a non-zero precipitation. Ignored for
\code{varcode} values different from \code{"pr"}. Default to 1 (assuming mm).}

\item{theta}{numeric indicating  upper threshold (and lower for the left tail of the distributions, if needed) 
above which precipitation (temperature) values are fitted to a Generalized Pareto Distribution (GPD). 
Values below this threshold are fitted to a gamma (normal) distribution. By default, 'theta' is the 95th 
percentile (5th percentile for the left tail). Only for \code{"gpqm"} method.}
}
\description{
Implementation of Generalized Quantile Mapping method for bias correction
}
\author{
S. Herrera and M. Iturbide
}
\keyword{internal}
