% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/conversion.R
\name{hurs2huss}
\alias{hurs2huss}
\title{Specific humidity from relative humidity}
\usage{
hurs2huss(tas, ps, hurs)
}
\arguments{
\item{tas}{surface air temperature (K)}

\item{ps}{surface pressure (Pa)}

\item{hurs}{surface relative humidity (\%)}
}
\value{
Specific humidity (kg.kg-1)
}
\description{
Derive specific humidity from relative humidity
}
\author{
S. Herrera \email{sixto@predictia.es}
}
\references{
Bohren & Albrecht (2000) Atmospheric thermodynamics. Oxford University Press. 402 pp
}
\seealso{
Other conversion: \code{\link{mslp2ps}};
  \code{\link{tdps2hurs}}
}
\keyword{internal}

