% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdas.R
\name{iberia_ncep_hus850}
\alias{iberia_ncep_hus850}
\title{Grid containing NCEP reanalysis data of specific humidity at 850mb for the Iberian Peninsula.}
\format{A grid}
\source{
\url{http://www.meteo.unican.es/ecoms-udg}
}
\description{
The data are daily means, wintertime (DJF) period 1991-2010.
}
\examples{
data(iberia_ncep_hus850)
plotMeanGrid(iberia_ncep_hus850)
}

