% Generated by roxygen2 (4.0.1): do not edit by hand
\name{iberia_ncep_hus850}
\alias{iberia_ncep_hus850}
\title{Field containing NCEP reanalysis data of specific humidity at 850mb for the Iberian Peninsula.}
\format{A field}
\source{
\url{http://www.meteo.unican.es/ecoms-udg}
}
\description{
The data correspond to the wintertime (DJF) period 1991-2010, and it consists of instantaneous data at 12:00 UTC
}
\examples{
data(iberia_ncep_hus850)
plotMeanField(iberia_ncep_hus850)
}
\seealso{
\code{\link{makeMultiField}}, \code{\link[ecomsUDG.Raccess]{loadECOMS}}
}

