% Generated by roxygen2 (4.0.1): do not edit by hand
\name{isimip}
\alias{isimip}
\title{Bias correction ISI-MIP method}
\usage{
isimip(obs, pred, sim, pr.threshold = 1)
}
\arguments{
\item{obs}{A field or station data containing the observed climate data for the training period}

\item{pred}{A field containing the simulated climate by the model for the training period. This can be either
the same variable as \code{obs}, in the case of model calibration (bias correction and related techniques) or MOS (model
 output statistics) approaches, or a set of predictors in the case of \emph{perfect prog} downscaling approaches (possibly
 after principal component analysis via the \code{\link{prinComp}} output).}

\item{sim}{A field containing the simulated climate for the variables used in \code{pred}, but considering the test period.}

\item{pr.threshold}{The minimum value that is considered as a non-zero precipitation. Ignored for
\code{varcode} values different from \code{"pr"}. Default to 1 (assuming mm).}
}
\value{
A calibrated object of the same spatio-temporal extent of the input field
}
\description{
Implementation of the ISI-MIP methodology
}
\details{
The methods available are qqmap, delta, unbiasing, scaling and Piani (only precipitation).

\strong{ISI-MIP}

Recently, Hempel et al.2013 proposed a new bias correction methodology within the ISI-MIP Project,
the first Inter-Sectoral Impact Model Intercomparison Project, funded by the German Federal Ministry
of Education and Research (BMBF). This method has been developed to preserve the change signal (trend,
climate change signal, etc.) and can be applied to several variables (precipitation, mean, maximum and
minimum temperature, windspeed and eastward/northward components, radiation, pressure and humidity).
The main difference with the rest of bias correction methods included in the \code{\link{biasCorrection}}
function is that the ISI-MIP method includes dependencies between some variables. That is, to correct some
of the variables (maximum/minimum temperatures and eastward/northward wind components) others are needed
(mean temperature and windspeed).

sim <- isimip(obs, pred, sim) # Temperature or other variable

sim <- isimip(obs, pred, sim, pr.threshold = threshold) # In the case of precipitation we should include the threshold considered of wet/dry days
}
\examples{
\dontrun{
# These are the paths to the package built-in GSN and NCEP datasets
gsn.data.dir <- file.path(find.package("downscaleR"), "datasets/observations/GSN_Iberia")
ncep.data.dir <- file.path(find.package("downscaleR"), "datasets/reanalysis/Iberia_NCEP/Iberia_NCEP.ncml")
# Data inventories provides a quick overview of the available data
gsn.inv <- dataInventory(gsn.data.dir)
ncep.inv <- dataInventory(ncep.data.dir)
str(gsn.inv)
str(ncep.inv)
# Load precipitation for boreal winter (DJF) in the train (1991-2000) and test (2001-2010) periods,
# for the observations (GSN_Iberia) and the Iberia_NCEP datasets
obs <- loadStationData(dataset = gsn.data.dir, var="precip", lonLim = c(-12,10), latLim = c(33,47), season=c(12,1,2), years = 1991:2000)
prd <- loadGridData(ncep.data.dir, var = "tp", lonLim = c(-12,10), latLim = c(33,47), season = c(12,1,2), years = 1991:2000)
sim <- loadGridData(ncep.data.dir, var = "tp", lonLim = c(-12,10), latLim = c(33,47), season = c(12,1,2), years = 2001:2010)
# Interpolate the observations onto the model's grid. We use the method "nearest" and the getGrid function to ensure spatial consistency:
obs <- interpGridData(obs, new.grid = getGrid(prd), method = "nearest")
# Apply the bias correction method:
simBC <- isimip (obs, prd, sim, pr.threshold = 1) # ISI-MIP Method
par(mfrow = c(1,2))
plotMeanField(sim)
plotMeanField(simBC)
par(mfrow = c(1,1))
}
}
\author{
S. Herrera \email{sixto@predictia.es}
}
\references{
\itemize{
\item Hempel, S., Frieler, K., Warszawski, L., Schewe, J., and Piontek, F. (2013) A trend-preserving bias correction: the ISI-MIP approach, Earth Syst. Dynam., 4, 219-236
}
}
\seealso{
\code{\link{biasCorrection}} for details on other standard methods for bias correction

Other downscaling: \code{\link{analogs}};
  \code{\link{biasCorrection}}; \code{\link{calibrateProj}}
}

