% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/loadGridDataset.R
\name{loadGridDataset}
\alias{loadGridDataset}
\title{Loads a user-defined subset of a gridded CDM dataset}
\usage{
loadGridDataset(var, grid, dic, level, season, years, time, latLon)
}
\description{
Loads a user-defined subset from a gridded dataset compliant with the Common
 Data Model interface
}
\author{
J. Bedia \email{joaquin.bedia@gmail.com}
}
\keyword{internal}

