% Generated by roxygen2 (4.0.1): do not edit by hand
\name{loadStationData}
\alias{loadStationData}
\title{Load station data}
\usage{
loadStationData(dataset, file.format = "ascii", var, stationID = NULL,
  lonLim = NULL, latLim = NULL, season = NULL, years = NULL, tz = "")
}
\arguments{
\item{dataset}{A character string indicating the database to be accessed. This is usually a path to a local file or a URL
pointing to a netCDF or NcML file in the case of netCDF and/or gridded datasets. For station data in standard ASCII format,
this is the path to the directory the dataset lives in.}

\item{var}{Variable code (character string). This is the name of the variable according to the R standard naming
(see the next argument). For variables with vertical levels, the vertical level is specified next to the variable name followed
 by the \dQuote{@} symbol (e.g. \code{var = "z@700"} for geopotential heigth at 700 mb isobaric surface pressure level).
 It is also possible to enter the variable name as originally coded in the dataset to skip data homogenization.}

\item{lonLim}{Vector of length = 2, with minimum and maximum longitude coordinates, in decimal degrees, of the bounding box selected.
For single-point queries, a numeric value with the longitude coordinate. If \code{NULL} (default), the whole longitudinal range
 is selected (Note that this may lead to a large output object size).}

\item{latLim}{Same as \code{lonLim}, but for the selection of the latitudinal range.}

\item{season}{An integer vector specifying the desired season (in months, January = 1 ..., December = 12).
Options include one to several (contiguous) months. Default to \code{NULL}, indicating a full year selection (same as \code{season = 1:12}).}

\item{years}{Optional vector of years to select. Default (\code{NULL}) to all available years. If the requested variable is static (e.g. orography)
it will be ignored.}

\item{file.format}{Wether the stations data are stored in a netCDF or ASCII (default) file.
Currently only the standard \dQuote{ASCII} format supported. See references for details.}

\item{stationID}{Optional. A character vector indicating the code names of the stations to be loaded.}

\item{tz}{A time zone specification to be used for the conversion of dates, if one is required
(i.e., if the time zone of the dataset does not correspond to the system-specific one; see
\code{\link[base]{timezones}} for details). Default to unspecified (i.e. \code{tz = ""}).}
}
\value{
a list with the following elements:
\itemize{
\item \code{Variable}. Name of the variable
\item \code{Data}. A 2-D matrix containing the data. Dates are ordered by rows and Stations by columns,
following the order indicated in the \code{Metadata}.
\item \code{xyCoords}. A 2-D matrix with longitude and latitudes of the stations
\item \code{Dates}. A list with the verification time interval of each record in the time series.
 This is represented by a list with two elements: \code{start} and \code{end}, representing the
 lower and upper bounds of the verification period
\item \code{Metadata}. A list of variable length depending on the available metadata associated
to each observation. If no metadata are provided, at least the station codes (compulsory) are displayed.
}
}
\description{
Load observations data from station datasets in standard ASCII format.
}
\note{
Unlike gridded datasets, station data do not use a dictionary for variable homogenization. Thus, users
must take care of variable units and eventual conversions when necessary.
}
\section{Geolocation parameters}{
 Regarding the selection of the spatial domain,
 it is possible to select the whole spatial domain of the dataset by defining the arguments \code{lonLim=NULL}
 and \code{latLim=NULL}. More often, rectangular domains are defined by the minimum and maximum coordinates
 in longitude and latitude (for instance \code{lonLim=c(-10,10)} and \code{latLim=c(35,45)} indicates a
 rectangular window centered in the Iberian Peninsula), and single grid-cell values
 (for instance \code{lonLim=-3.21} and \code{latLim=41.087} for retrieving the data in the closest grid
 point to the point coordinate -3.21E, 41.087N. In the last two cases, the function
 operates by finding the nearest (euclidean distance) grid-points to the coordinates introduced.

 In the case of station data (\code{\link{loadStationData}}), the logic is the same, taking into account that in the case
 of rectangular domains, all stations falling inside that window will be loaded. For single-point selections,
 the closest station will be chosen, and a note on-screen will inform about the distance from the selected point
 to the chosen station.
}
\author{
J. Bedia \email{joaquin.bedia@gmail.com}
}
\references{
\url{https://github.com/SantanderMetGroup/downscaleR/wiki/Observation-Data-format}
}
\seealso{
Other loading: \code{\link{loadGridData}};
  \code{\link{makeAggregatedDataset}}
}

