% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{mslp2ps}
\alias{mslp2ps}
\title{Sea-level pressure to surface pressure}
\usage{
mslp2ps(tas, zs, mslp)
}
\arguments{
\item{tas}{surface temperature}

\item{zs}{surface geopotential (m^2/s^2)}

\item{mslp}{sea-level pressure (Pa)}
}
\value{
surface pressure (Pa)
}
\description{
Conversion of sea-level pressure to surface pressure
}
\author{
S. Herrera \email{sixto@predictia.es}
}
\seealso{
Other conversion: \code{\link{hurs2huss}},
  \code{\link{tdps2hurs}}
}
\keyword{internal}

