% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdas.R
\docType{data}
\name{NCEP_Iberia_tp}
\alias{NCEP_Iberia_tp}
\title{Field containing NCEP reanalysis data of precipitation for the Iberian Peninsula.}
\format{A field}
\source{
subset of NCEP reanalysis data, which is accessible 
through the \strong{ECOMS User Data Gateway (ECOMS-UDG)}
}
\description{
NCEP_Iberia_tp is a field object returned by loadECOMS 
function (package ecomsUDG.Raccess):
season = c(12,1,2),  
years = 1991-2000, 
var = "tp", 
lonLim = c(-10,5),  
latLim= c(34,44), 
time = "DD", 
aggr.d = "sum".
}
\seealso{
\code{\link{makeMultiField}}, \code{\link[ecomsUDG.Raccess]{loadECOMS}}
}
\keyword{NCEP}
\keyword{reanalysis}

