% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClimatology.R
\name{plotClimatology}
\alias{plotClimatology}
\title{Lattice plot methods for climatological grids}
\usage{
plotClimatology(grid, backdrop.theme = "none", ...)
}
\arguments{
\item{grid}{Input grid}

\item{backdrop.theme}{Reference geographical lines to be added to the plot. See Details.}

\item{...}{Further arguments passed to \code{spplot}}
}
\value{
As spplot, \code{plotClimatology} returns a lattice plot of class \dQuote{trellis}. 
 If you fail to \dQuote{see} it, explicitly call \code{print(plotClimatology(...))}.
}
\description{
A wrapper for the lattice (trellis) plot methods for spatial data in \code{sp::spplot}
}
\details{
The function applies the \code{\link[sp]{spplot}} method after conversion of the climatological map(s) to a
 \code{SpatialGridDataFrame}.
 
 \strong{Multigrids}
 
 Multigrids of climatologies can be created using \code{makeMultiGrid} 
 for trellis visualization of different variables, or for instance, for the comparison of
 raw and corrected/downscaled scenarios side to side. In case of multimember multigrids, 
 the function will internally compute the ensemble mean of each variable in the multigrid
  for representation (with a message).
 
 \strong{Backdrop theme}
 
 Current implemented options are \code{"none"} and \code{"coastline"}, which contains
 a simplied vector theme delineating the world coastlines. Any other themes can be introduced
 by the user using the \code{sp.layout} options in \code{spplot}.
 
 \strong{Controlling graphical parameters}
 
 Many different aspects of the map can be controlled passing the relevant arguments to 
 spplot. Fine control of graphical parameters for the trellis display can
 be also controlled using \code{\link{lattice}{trellis.par.set}}.
}
\examples{
\donttest{
data(tasmax_forecast)
# Climatology is computed:
clim <- climatology(tasmax_forecast, by.member = TRUE)
plotClimatology(clim)
# Geographical lines can be added using the argument 'backdrop.theme':
plotClimatology(clim, backdrop.theme = "coastline")

# Further arguments can be passed to 'spplot'...

# ... a subset of members to be displayed, using 'zcol':
plotClimatology(clim,
                backdrop.theme = "coastline",
                zcol = 1:4)
                
# ... regional focuses (e.g. the Iberian Peninsula). 
plotClimatology(clim,
                backdrop.theme = "countries",
                xlim = c(-10,5), ylim = c(35,44),
                zcol = 1:4,
                scales = list(draw = TRUE))

# Changing the default color palette and ranges:
plotClimatology(clim,
                backdrop.theme = "coastline",
                zcol = 1:4,
                col.regions = cm.colors(27), at = seq(10,37,1))
                
# For ensemble means climatology should be called with 'by.member' set to FALSE:
clim <- climatology(tasmax_forecast, by.member = FALSE)

# Adding contours to the plot is direct with argument 'contour':
plotClimatology(clim,
                scales = list(draw = TRUE),
                contour = TRUE,
                main = "tasmax Predictions July Ensemble Mean")
}

}
\author{
J. Bedia
}
\seealso{
\code{\link{climatology}}. See \code{\link[sp]{spplot}} in package \pkg{sp} for further information on
plotting capabilities and options
}

