% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotMeanField}
\alias{plotMeanField}
\title{Plot a map of the mean value of a grid dataset}
\usage{
plotMeanField(gridData, multi.member = FALSE)
}
\arguments{
\item{gridData}{A grid dataset}

\item{multi.member}{Should members be plotted sepparately (TRUE), or just a plot
 of the multi-member mean (FALSE, default)?. Ignored if the dataset has no members.}
}
\value{
a plot of the mean/multi-member/multi-variable field with a world map superposed
}
\description{
Plot the spatial mean of a gridded variable, or variables in the case of multi-fields.
}
\details{
The function is a wrapper of the \code{\link[fields]{image.plot}} function
in package \pkg{fields}
}
\note{
The function plots a simple temporal mean of the loaded object in the form of
a map. It does not handle other temporal aggregations.
}
\examples{
# A field
data(iberia_ncep_ta850)
plotMeanField(iberia_ncep_ta850)
# A multifield
data(iberia_ncep_hus850)
mf <- makeMultiField(iberia_ncep_ta850, iberia_ncep_hus850)
plotMeanField(mf)
# A multimember field
data(tasmax_forecast)
plotMeanField(tasmax_forecast) # multimember mean
plotMeanField(tasmax_forecast, multi.member = TRUE) # by members
# A multimember multifield
data(tasmin_forecast)
mm.mf <- makeMultiField(tasmax_forecast, tasmin_forecast)
plotMeanField(mm.mf) # Note: multi-member not supported in this case
}
\author{
J Bedia \email{joaquin.bedia@gmail.com}
}

