% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareNewData.R
\name{prepareNewData}
\alias{prepareNewData}
\title{Prepare newdata for predictions}
\usage{
prepareNewData(newdata, data.structure)
}
\arguments{
\item{newdata}{A grid containing the prediction data.}

\item{data.structure}{A structure, as returned by \code{\link{prepareData}}}
}
\value{
A named list with the components required by the downscaling method in order to perform the predictions
}
\description{
Prepare the prediction data according to the definition of the experiment
}
\examples{
# See the dedicated vignette by typing:
# utils::vignette(topic = "configuring_newdata", package = "downscaleR")
}
\seealso{
Other downscaling.helpers: \code{\link{predictor.nn.indices}},
  \code{\link{predictor.nn.values}},
  \code{\link{prepareData}}
}
\author{
J Bedia
}
