% Generated by roxygen2 (4.0.1): do not edit by hand
\name{prinComp}
\alias{prinComp}
\title{Principal Component Analysis of field/multifield data}
\usage{
prinComp(gridData, n.eofs = NULL, v.exp = NULL, scaling = c("field",
  "gridbox"))
}
\arguments{
\item{gridData}{A multifield object or a single field as returned by loadGridData}

\item{n.eofs}{Number of EOFs to be retained. Default to \code{NULL}, indicating
that either all EOFs are kept, or that the next argument will be used as criterion
for its determination. See next argument and details.}

\item{v.exp}{Maximum fraction of explained variance, in the range (0,1]. Used to determine the number of EOFs
to be retained, as an alternative to \code{n.eofs}. Default to \code{NULL}. See details.}

\item{scaling}{Method for performing the scaling (and centering) of the input raw data matrix.
Currently accepted choices are \code{"field"} (the default) and \code{"gridbox"}. See details.}
}
\value{
A list of \emph{N + 1} elements for multifields, where \emph{N} is the number of input variables used
 and the last element contains the results of the combined PCA (See details). The list is named as the variables,
 including the last element named \code{"COMBINED"}. In case of single fields (1 variable only), a list of length 1
  (without the combined element). For each element of the list, the following objects are returned:

 \itemize{
 \item \code{PCs}: A matrix of principal components, arranged in columns by decreasing importance order
 \item \code{EOFs}: A matrix of EOFs, arranged in columns by decreasing importance order
 }
 The \dQuote{order of importance} is given by the explained variance of each PC, as indicated
 in the attribute \code{"explained_variance"} as a cumulative vector.
 Additional information is returned via the remaining attributes (see Details).
}
\description{
Performs a Principal Component Analysis of field or multifield data
}
\details{
\strong{Number of EOFs}

\code{n.eofs} and \code{v.exp} are alternative choices for the determination
 of the number of EOFs (hence also the corresponding PCs) to be retained. If both are \code{NULL} (the default)
 , all EOFs will be retained. If both are given a value different from \code{NULL},
 the \code{n.eofs} argument will prevail, and \code{v.exp} will be ignored, with a warning.
 Note that when dealing with multifields, the \code{n.eofs} argument will return the same number of EOFs
 for each variable by sepparate and also for the combined field, while \code{v.exp} will select
 a different number, depending on the explained variance in each case.

\strong{Scaling and centering}

In order to eliminate the effect of the varying scales of the different climate variables, the input
data matrix is always scaled and centered, and there is no choice to avoid this step. However, the mean
and standard deviation can be either computed for each grid box individually (\code{"gridbox"}) or for all
grid-boxes together (i.e., at the field level, \code{"field"}). The last case is the preferred choice and has
been set as default, returning one single mean and sigma parameter for each variable. If the \code{"gridbox"}
approach is selected, a vector of length \emph{n}, where \emph{n} is the number of grid-cells composing the
field, is returned for both the mean and sigma parameters (this is equivalent to using the \code{\link{scale}}
function with the input data matrix).

The method used is returned as a global attribute of the returned object (\code{"scaled:method"}), and the
\emph{mu} and \emph{sigma} parameters are returned as attributes of the corresponding variables
 (\code{"scaled:scale"} and \code{"scaled:center"} respectively).

\strong{Combined EOF analysis}

When dealing with multifield data, apart from the PCA analysis performed on each variable indivisually,
a combined analysis considering all variables together is done. This is always returned in the last element
of the output list.
}
\examples{
\dontrun{
# First a multifield containing a set of variables is loaded (e.g. data for spring spanning the 30-year period 1981--2010):
ncep <- file.path(find.package("downscaleR"), "datasets/reanalysis/Iberia_NCEP/Iberia_NCEP.ncml")
multifield <- loadMultiField(ncep, vars = c("hus@85000", "ta@85000", "psl"), season = c(3:5), years = 1981:2010)
# In this example, we retain the PCs explaining the 99\\\% of the variance
pca <- prinComp(multifield, v.exp = .99)
# Note that, apart from computing the principal components and EOFs for each field, it also returns, in the last element of the output list,
# the results of a PC analysis of all the variables combined:
str(pca)
# The different attributes of the pca object provide information regarding the variables involved and the geo-referencing information
attributes(pca)
# In addition, for each variable (and their combination), the scaling and centering parameters are also returned.
# These are either a single value in case of field scaling (the default), or a vector of values, one for each grid-cell, for the
# gridbox scaling. For instance, the parameters for the specific humidity field are:
attributes(pca$hus)$`scaled:center`
attributes(pca$hus)$`scaled:scale`
# In addition, the (cumulative) explained variance of each PC is also returned:
vexp <- attributes(pca$hus)$explained_variance
# The classical "scree plot":
barplot(1-vexp, names.arg = paste("PC",1:length(vexp)), las = 2, ylab = "Fraction of explained variance")
}
}
\author{
J. Bedia \email{joaquin.bedia@gmail.com}
}
\references{
Guti\'{e}rrez, J.M., R. Ancell, A. S. Cofi\~{n}o and C. Sordo (2004). Redes Probabil\'{i}sticas
 y Neuronales en las Ciencias Atmosf\'{e}ricas. MIMAM, Spain. 279 pp.
  \url{http://www.meteo.unican.es/en/books/dataMiningMeteo}
}
\seealso{
Other loading.grid: \code{\link{fieldFromPCs}};
  \code{\link{getGrid}}; \code{\link{interpGridData}};
  \code{\link{loadGridData}}; \code{\link{loadMultiField}}

Other multifield: \code{\link{fieldFromPCs}};
  \code{\link{loadMultiField}}
}

