% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickDiagnostics.R
\name{quickDiagnostics}
\alias{quickDiagnostics}
\title{Plots daily/annual series and the annual correlation map of different field objects}
\usage{
quickDiagnostics(obs, sim, downscaled = NULL, location = c(-42.5, -3),
  type = c("daily", "interannual"), ylim = NULL)
}
\arguments{
\item{obs}{Field of observations.}

\item{sim}{Field of model data.}

\item{downscaled}{Optional. Field of the downscaling output.}

\item{location}{Coordinates of a location in the geographic domain of the field.}

\item{type}{Character value, either \code{"daily"} or \code{"interannual"}, indicating is the assessment is to
be performed on a daily or interannual basis.}

\item{ylim}{'ylim' argument passed to the time series plot.}
}
\value{
Two diagnostic plots with observed, simulated and (possibly) downscaled time series, and a QQ-plot by percentlies.
}
\description{
Plots daily/annual series and the annual correlation map of different field objects
}
\author{
M. Iturbide \email{maibide@gmail.com}
}
\seealso{
Other visualization: \code{\link{bubbleValidation}},
  \code{\link{plotMeanField}},
  \code{\link{tercileValidation}}
}

