% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redim.R
\name{redim}
\alias{redim}
\title{Complete missing dimensions of Grid objects}
\usage{
redim(obj, drop = FALSE)
}
\arguments{
\item{grid}{A grid}
}
\value{
The same object with all the dimensions (i.e. member, time, station)
}
\description{
Inverse of drop to complete all dimensions of the Data array
}
\author{
M. Iturbide
}
\keyword{internal}

