% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetField.R
\name{subsetDimension}
\alias{subsetDimension}
\title{Select an arbitrary subset from a field or multifield along one of its dimensions}
\usage{
subsetDimension(field, dimension = NULL, indices = NULL)
}
\arguments{
\item{field}{The input field to be subset. This is either a field, as returned e.g. by \code{loadGridData}, a
multifield, as returned by \code{loadMultiField} or \code{makeMultiField}, or other types of multimember fields
(possibly multimember multifields) as returned e.g. by \code{\link[ecomsUDG.Raccess]{loadECOMS}}.}

\item{dimension}{Character vector indicating the dimension along which the positions indicated by the \code{indices} paraneter.}

\item{indices}{An integer vector indicating \strong{the positions} of the dimension to be extracted.}
}
\value{
A new field object that is a logical subset of the input field along the specified dimension.
}
\description{
Create a new field/multifield that is a subset of the input field 
along the selected dimension
}
\details{
The attribute \code{subset} will be added taking the value of the \code{dimension} parameter.
}
\examples{
# Example - Member subset
data(tasmax_forecast)
plotMeanField(tasmax_forecast, TRUE)
# Selection of a smaller domain over the Iberian Peninsula and members 3 and 7
sub <- subsetDimension(tasmax_forecast,
                   dimension = "member",
                   indices = c(1,3))
plotMeanField(sub, multi.member = TRUE)
}
\author{
J. Bedia \email{joaquin.bedia@gmail.com} and S. Herrera
}
\seealso{
Other subsetting: \code{\link{subsetField}},
  \code{\link{subsetMembers}}, \code{\link{subsetSeason}},
  \code{\link{subsetSpatial}}, \code{\link{subsetVar}},
  \code{\link{subsetYears}}
}

