% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetField.R
\name{subsetMembers}
\alias{subsetMembers}
\title{Member subsets from a multimember field}
\usage{
subsetMembers(mmField, members = NULL)
}
\arguments{
\item{mmField}{Input multimember field to be subset (possibly a multimember multifield).}

\item{members}{An integer vector indicating \strong{the position} of the members to be subset.}
}
\value{
A field (or multifield) that is a logical subset of the input field along its 'member' dimension.
}
\description{
Retrieves a field that is a logical subset of a multimember field along its 'member' dimension.
 Multimember multifields are supported. Subroutine of \code{\link{subsetField}}.
}
\details{
An attribute 'subset' with value 'subsetMembers' is added to the Members slot of the output subset.
}
\author{
J. Bedia (joaquin.bedia@gmail.com)
}
\seealso{
Other subsetting: \code{\link{subsetDimension}},
  \code{\link{subsetField}}, \code{\link{subsetSeason}},
  \code{\link{subsetSpatial}}, \code{\link{subsetVar}},
  \code{\link{subsetYears}}
}
\keyword{internal}

