% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetField.R
\name{subsetSeason}
\alias{subsetSeason}
\title{Monthly subset of a field}
\usage{
subsetSeason(field, season = NULL)
}
\arguments{
\item{field}{Input field to be subset (possibly a multimember/multifield).}

\item{season}{An integer vector indicating the months to be subset.}
}
\value{
A field (or multifield) that is a logical subset of the input field along its 'time' dimension.
}
\description{
Retrieves a field that is a logical subset of the input field along its 'time' dimension,
 on a monthly basis. Multimember multifields are supported. Subroutine of \code{\link{subsetField}}.
}
\details{
An attribute 'subset' with value 'subsetSeason' is added to the Dates slot of the output subset.
}
\author{
J. Bedia (joaquin.bedia@gmail.com)
}
\seealso{
Other subsetting: \code{\link{subsetDimension}},
  \code{\link{subsetField}}, \code{\link{subsetMembers}},
  \code{\link{subsetSpatial}}, \code{\link{subsetVar}},
  \code{\link{subsetYears}}
}
\keyword{internal}

