% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetGrid.R
\name{subsetSpatial}
\alias{subsetSpatial}
\title{Spatial subset from a grid}
\usage{
subsetSpatial(grid, lonLim, latLim, outside, drop)
}
\arguments{
\item{grid}{Input grid to be subset (possibly a multimember multigrid).}

\item{lonLim}{Vector of length = 2, with minimum and maximum longitude coordinates, in decimal degrees,
of the bounding box defining the subset. For single-point subsets, a numeric value with the
longitude coordinate. If \code{NULL} (default), no subsetting is performed on the longitude dimension}

\item{latLim}{Same as \code{lonLim} argument, but for latitude.}

\item{outside}{Logical. Default to \code{FALSE}. If \code{TRUE}, subset coordinates outside the grid extent 
are allowed.}

\item{drop}{Logical (default to \code{TRUE}). Drop or keep dimensions of length 1.}
}
\value{
A grid (or multigrid) that is a logical spatial subset of the input grid.
}
\description{
Retrieves a grid that is a logical subset of the input grid along its 'lat' and 'lon' dimensions.
 Multimember multigrids are supported. Subroutine of \code{\link{subsetGrid}}.
}
\details{
An attribute \code{subset} with value \code{subsetSpatial} is added to the \code{xyCoords}
component of the output grid.
}
\author{
J. Bedia
}
\seealso{
Other subsetting: \code{\link{subsetDimension}},
  \code{\link{subsetGrid}}, \code{\link{subsetMembers}},
  \code{\link{subsetRuntime}}, \code{\link{subsetSeason}},
  \code{\link{subsetVar}}, \code{\link{subsetYears}}
}
\keyword{internal}

