% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetField.R
\name{subsetSpatial}
\alias{subsetSpatial}
\title{Spatial subset from a field}
\usage{
subsetSpatial(field, lonLim = NULL, latLim = NULL)
}
\arguments{
\item{field}{Input field to be subset (possibly a multimember multifield).}

\item{lonLim}{Vector of length = 2, with minimum and maximum longitude coordinates, in decimal degrees,
of the bounding box defining the subset. For single-point subsets, a numeric value with the
longitude coordinate. If \code{NULL} (default), no subsetting is performed on the longitude dimension}

\item{latLim}{Same as \code{lonLim} argument, but for latitude.}
}
\value{
A field (or multifield) that is a logical spatial subset of the input field.
}
\description{
Retrieves a field that is a logical subset of the input field along its 'lat' and 'lon' dimensions.
 Multimember multifields are supported. Subroutine of \code{\link{subsetField}}.
}
\details{
An attribute 'subset' with value 'subsetSpatial' is added to the xyCoords slot of the output subset.
}
\author{
J. Bedia (joaquin.bedia@gmail.com)
}
\seealso{
Other subsetting: \code{\link{subsetDimension}},
  \code{\link{subsetField}}, \code{\link{subsetMembers}},
  \code{\link{subsetSeason}}, \code{\link{subsetVar}},
  \code{\link{subsetYears}}
}
\keyword{internal}

