% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetField.R
\name{subsetVar}
\alias{subsetVar}
\title{Extract a field from a multifield object}
\usage{
subsetVar(multiField, var)
}
\arguments{
\item{multiField}{Input multifield to be subset}

\item{var}{Character vector indicating the variables(s) to be extracted}
}
\value{
Either a (multimember)field or (multimember)multifield if one ore more variables
are selected respectively.
}
\description{
Extracts a field from a multifield object. Multimember multifields are supported. Subroutine of subsetField.
}
\details{
Argument \code{var} is insensitive to the order of the variables, i.e.: variables
will be always returned in the same order they are in the original multifield.

An attribute 'subset' with value 'subsetVar' is added to the Variable slot of the output subset.
}
\author{
J. Bedia (joaquin.bedia@gmail.com)
}
\seealso{
Other subsetting: \code{\link{subsetDimension}},
  \code{\link{subsetField}}, \code{\link{subsetMembers}},
  \code{\link{subsetSeason}}, \code{\link{subsetSpatial}},
  \code{\link{subsetYears}}
}
\keyword{internal}

