% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetField.R
\name{subsetYears}
\alias{subsetYears}
\title{Year subsets from a multimember field}
\usage{
subsetYears(field, years = NULL)
}
\arguments{
\item{field}{Input field to be subset (possibly a multimember/multifield).}

\item{years}{An integer vector indicating the years to be subset.}
}
\value{
A field (or multifield) that is a logical subset of the input field along its 'time' dimension.
}
\description{
Retrieves a field that is a logical subset of a multimember field along its 'time' dimension,
 on a yearly basis. Multimember multifields are supported. Subroutine of \code{\link{subsetField}}.
}
\details{
An attribute 'subset' with value 'subsetYears' is added to the Dates slot of the output subset.
}
\author{
J. Bedia (joaquin.bedia@gmail.com)
}
\seealso{
Other subsetting: \code{\link{subsetDimension}},
  \code{\link{subsetField}}, \code{\link{subsetMembers}},
  \code{\link{subsetSeason}}, \code{\link{subsetSpatial}},
  \code{\link{subsetVar}}
}
\keyword{internal}

