% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/timeAggregation.R
\name{timeAggregation}
\alias{timeAggregation}
\title{Annual aggregation of the data in a field/multifield}
\usage{
timeAggregation(obj, aggr.d = NULL, aggr.y = "mean")
}
\arguments{
\item{obj}{a field or multifield (can be multimember).}

\item{aggr.d}{a single character or a character vector indicating the daily aggregation function.
Possibilities are:  "mean", "min", "max", "sum" (default is NULL).}

\item{aggr.y}{a single character or a character vector indicating the annual aggregation function.
Possibilities are: "none", "mean", "min", "max", "sum" (default is "mean").}
}
\value{
a field or multifield with the time dimension daily or annually aggregated (see details).
}
\description{
Aggregates data sub-annual data annually by specifiyng an aggregation function
}
\details{
If aggr.d = NULL and the data in the field is sub-daily, annualAggregation will return an error.
If the data is sub-daily and FUN = "none", the function will return the daily aggregated field.
}
\author{
M. Iturbide \email{maibide@gmail.com}
}

