% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rdas.R
\docType{data}
\name{VALUE_Igueldo_tas}
\alias{VALUE_Igueldo_tas}
\title{Station data from the VALUE_ECA_86_v2 dataset containing daily mean temperature for the Igueldo-SanSebastian station.}
\format{Station data}
\source{
Subset of VALUE station data. Full dataset is accessible
for download in \strong{"http://meteo.unican.es/work/downscaler/data/VALUE_ECA_86_v2.tar.gz"}.
}
\description{
season = c(12,1,2),
years = 1991-2000,
stationID = "000234"
var= "tmean".
}
\seealso{
\code{\link{NCEP_Iberia_tas}}, \code{\link{VALUE_tas}}
}
\keyword{Igueldo}
\keyword{VALUE}
\keyword{station}
\keyword{temperature}

