% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rdas.R
\docType{data}
\name{VALUE_Igueldo_tp}
\alias{VALUE_Igueldo_tp}
\title{Station data from the VALUE_ECA_86_v2 dataset containing daily precipitation for the Igueldo-SanSebastian station.}
\format{Station data}
\source{
Subset of VALUE station data. Full dataset is accessible
for download in \strong{"http://meteo.unican.es/work/downscaler/data/VALUE_ECA_86_v2.tar.gz"}.
}
\description{
season = c(12,1,2),
years = 1991-2000,
stationID = "000234"
var= "precip".
}
\seealso{
\code{\link{NCEP_Iberia_tp}}, \code{\link{VALUE_tp}}
}
\keyword{Igueldo}
\keyword{VALUE}
\keyword{precipitation}
\keyword{station}

