% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdas.R
\docType{data}
\name{VALUE_tas}
\alias{VALUE_tas}
\title{Station data from the VALUE_ECA_86_v2 dataset containing daily mean temperature for 82 stations in Europe.}
\format{Station data}
\source{
Subset of VALUE station data. Full dataset is accessible 
for download in \strong{"http://meteo.unican.es/work/downscaler/data/VALUE_ECA_86_v2.tar.gz"}.
}
\description{
season = c(12,1,2),  
years = 1991-2000, 
var= "tmean".
}
\seealso{
\code{\link{NCEP_Iberia_tas}}, \code{\link{VALUE_Igueldo_tas}}
}
\keyword{VALUE}
\keyword{station}
\keyword{temperature}

